% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ABF.R
\name{ABF}
\alias{ABF}
\title{Fine-map with ABF}
\source{
\itemize{
\item JB Maller et al., Bayesian refinement of association signals for 14 loci in 3 common diseases. Nature Genetics. 44, 1294–1301 (2012).
\item J Wakefield, A bayesian measure of the probability of false discovery in genetic epidemiology studies. American Journal of Human Genetics. 81, 208–227 (2007).
}
}
\usage{
ABF(subset_DT, PP_threshold = 0.95)
}
\description{
Conduct statistical (non-functional) fine-mapping with approximate Bayes factor (ABF).
}
\keyword{internal}
