% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate.R
\name{ANNOTATE.annotate_missense}
\alias{ANNOTATE.annotate_missense}
\title{Annotate any missense variants}
\usage{
ANNOTATE.annotate_missense(merged_DT, snp_filter = "Support>0")
}
\description{
Annotate any missense variants
}
\examples{
data("merged_DT");
annotated_DT <- ANNOTATE.annotate_missense(merged_DT=merged_DT, snp_filter="Support>0")
}
\seealso{
Other annotate: 
\code{\link{ANNOTATE.plot_missense}()},
\code{\link{SNPs_by_mutation_type}()},
\code{\link{biomart_geneInfo}()},
\code{\link{biomart_snp_info}()},
\code{\link{biomart_snps_to_geneInfo}()},
\code{\link{epigenetics_enrichment}()},
\code{\link{epigenetics_summary}()},
\code{\link{haploR.HaploReg}()},
\code{\link{haploR.regulomeDB}()}
}
\concept{annotate}
