% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arg_list_handler.R
\name{arg_list_handler}
\alias{arg_list_handler}
\title{Argument list handler}
\usage{
arg_list_handler(arg, i, loci, use_names = FALSE, error = TRUE)
}
\arguments{
\item{arg}{Name of the argument to evaluate.}

\item{i}{Iterator index.}

\item{loci}{A vector of loci that are being iterated over.}

\item{use_names}{Whether to identify locus-specific value
by name instead of index.}

\item{error}{Throw an error (\code{TRUE})
instead of a warning (\code{FALSE}).}
}
\value{
Selected argument value.
}
\description{
Infer whether the argument should be applied to all loci or matched by index.
}
\concept{general}
\keyword{internal}
