% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arg_types.R
\name{arg_types}
\alias{arg_types}
\title{Argument types}
\usage{
arg_types(
  fun,
  keep_types = c("value", "function", "NULL", "no_default"),
  verbose = FALSE
)
}
\arguments{
\item{fun}{Function to evaluate.}

\item{keep_types}{Argument types to keep in output.}

\item{verbose}{Print messages.}
}
\value{
Named list of argument types.
}
\description{
Categorize arguments of a given function by type.
}
\keyword{internal}
