% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_lead_SNP.R
\name{assign_lead_SNP}
\alias{assign_lead_SNP}
\title{Assign a lead GWAS SNP to a locus}
\usage{
assign_lead_SNP(new_DT, verbose = T)
}
\arguments{
\item{data.frame}{Fine-mapping results data.frame.}
}
\value{
Fine-mapping results data.frame with new boolean \strong{leadSNP} column,
 indicating whether each SNPs is the lead GWAS SNP in that locus or not.
}
\description{
If none of the SNPs in the data.frame have \code{leadSNP==T},
then sort by lowest p-value (and then highest Effect size) and assign the top SNP as the lead SNP.
}
\keyword{internal}
