% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_topSNPs_sub.R
\name{auto_topSNPs_sub}
\alias{auto_topSNPs_sub}
\title{Automatically identify top SNP per locus}
\usage{
auto_topSNPs_sub(top_SNPs, query, locus)
}
\description{
If no \code{top_SNPs} dataframe is supplied,
 this function will sort by p-value and then effect size,
 and use the SNP in the first row.
}
\seealso{
Other standardization functions: 
\code{\link{calculate_tstat}()}
}
\concept{standardization functions}
