% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batapply.R
\name{batapply}
\alias{batapply}
\title{\pkg{echolocatoR}-themed progress bar}
\usage{
batapply(
  X,
  FUN = function(l) Sys.sleep(5/100),
  ...,
  apply_func = lapply,
  total = length(X),
  name = NULL,
  .envir = parent.frame(),
  cli.progress_show_after = 0,
  style = c("bar", "squares", "dot", "fillsquares", "classic"),
  type = c("iterator", "tasks", "download", "custom"),
  clear = FALSE
)
}
\arguments{
\item{X}{a vector (atomic or list) or an \code{\link[base]{expression}}
    object.  Other objects (including classed objects) will be coerced
    by \code{base::\link[base]{as.list}}.}

\item{FUN}{the function to be applied to each element of \code{X}:
    see \sQuote{Details}.  In the case of functions like
    \code{+}, \code{\%*\%}, the function name must be backquoted or quoted.}

\item{...}{optional arguments to \code{FUN}.}

\item{apply_func}{Iterator function to use (default: \link[base]{lapply}).}

\item{total}{Passed to \code{\link[cli:cli_progress_bar]{cli_progress_bar()}}.}

\item{name}{Name of the progress bar, a label, passed to
\code{\link[cli:cli_progress_bar]{cli_progress_bar()}}.}

\item{.envir}{Passed to \code{\link[cli:cli_progress_bar]{cli_progress_bar()}}.}
}
\value{
A (named) list.
}
\description{
Iterator function with \pkg{echolocatoR}-themed progress bar.
}
\examples{
out <- batapply(X = seq_len(30))
}
