% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate.R
\name{biomart_geneInfo}
\alias{biomart_geneInfo}
\title{Get gene info using Biomart}
\source{
\href{https://bioconductor.org/packages/release/bioc/html/biomaRt.html}{biomaRt}
}
\usage{
biomart_geneInfo(geneList, reference_genome = "grch37")
}
\description{
Get gene info using Biomart
}
\examples{
gene_info <- biomart_geneInfo(c("PTK2B","CLU","APOE"))
}
\seealso{
Other annotate: 
\code{\link{SNPs_by_mutation_type}()},
\code{\link{biomart_snp_info}()},
\code{\link{biomart_snps_to_geneInfo}()},
\code{\link{epigenetics_enrichment}()},
\code{\link{epigenetics_summary}()},
\code{\link{haploR.HaploReg}()},
\code{\link{haploR.regulomeDB}()}
}
\concept{annotate}
\keyword{internal}
