% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate.R
\name{biomart_snps_to_geneInfo}
\alias{biomart_snps_to_geneInfo}
\title{Identify which genes SNPs belong to using Biomart}
\source{
\href{https://bioconductor.org/packages/release/bioc/html/biomaRt.html}{biomaRt}
}
\usage{
biomart_snps_to_geneInfo(snp_list, reference_genome = "grch37")
}
\description{
Identify which genes SNPs belong to using Biomart
}
\seealso{
Other annotate: 
\code{\link{SNPs_by_mutation_type}()},
\code{\link{biomart_geneInfo}()},
\code{\link{biomart_snp_info}()},
\code{\link{epigenetics_enrichment}()},
\code{\link{epigenetics_summary}()},
\code{\link{haploR.HaploReg}()},
\code{\link{haploR.regulomeDB}()}
}
\concept{annotate}
\keyword{internal}
