% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize.R
\name{calculate_tstat}
\alias{calculate_tstat}
\title{Compute t-stat}
\usage{
calculate_tstat(finemap_dat, tstat_col = "t_stat")
}
\description{
If \strong{tstat} column is missing,
compute t-statistic from: \code{Effect / StdErr}.
}
\seealso{
Other standardization functions: 
\code{\link{auto_topSNPs_sub}()}
}
\concept{standardization functions}
