% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_deprecated.R
\name{check_deprecated}
\alias{check_deprecated}
\title{Check deprecated arguments}
\usage{
check_deprecated(
  fun = "finemap_loci",
  pkg = "echolocatoR",
  when = "2.0.0",
  args = match.call(),
  lifecycle_fun = lifecycle::deprecate_warn,
  map = list(A1_col = "colmap", A2_col = "colmap", chrom_col = "colmap", position_col =
    "colmap", effect_col = "colmap", freq_col = "colmap", gene_col = "colmap", locus_col
    = "colmap", MAF_col = "colmap", N_cases = "colmap", N_controls = "colmap",
    N_cases_col = "colmap", N_controls_col = "colmap", MAF_col = "colmap", pval_col =
    "colmap", stderr_col = "colmap", tstat_col = "colmap", snp_col = "colmap", file_sep =
    NULL, probe_path = NULL, chrom_type = NULL, PAINTOR_QTL_datasets = NULL, QTL_prefixes
    = NULL, 
     proportion_cases = NULL, server = NULL, vcf_folder = NULL, top_SNPs =
    "topSNPs", PP_threshold = "credset_thresh", consensus_threshold = "consensus_thresh",
    sample_size = "compute_n", plot.types = "plot_types", plot.Roadmap = "roadmap",
    plot.Roadmap_query = "roadmap_query", plot.XGR_libnames = "xgr_libnames", plot.zoom =
    "zoom", plot.zoom = "zoom", plot.Nott_epigenome = "nott_epigenome",
    plot.Nott_show_placseq = "nott_show_placseq")
)
}
\arguments{
\item{fun}{Function to check.}

\item{pkg}{Package that the function is from.}

\item{when}{A string giving the version when the behaviour was deprecated.}

\item{args}{Argument calls to assess.}

\item{lifecycle_fun}{Which \pkg{lifecycle} function to use by default.}

\item{map}{Mapping between old:new argument names. Use \code{NULL}
if the argument is no longer used at all.}
}
\description{
Semi-automatically check all deprecated args in a given function.
}
\examples{
topSNPs <- echodata::topSNPs_Nalls2019
fullSS_path <- echodata::example_fullSS()
testthat::expect_error(
  echolocatoR::finemap_loci(
    fullSS_path = fullSS_path,
    topSNPs = topSNPs,
    loci = c("BST1","MEX3C"),
    chrom_col = "CHR",
    position_col = "BP")
)
}
