% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_genome.R
\name{check_genome}
\alias{check_genome}
\title{Check genome build}
\usage{
check_genome(
  fullSS_genome_build = NULL,
  munged = FALSE,
  fullSS_path = NULL,
  sampled_snps = 10000,
  names_from_paths = TRUE,
  dbSNP = 144,
  nThread = 1,
  verbose = TRUE
)
}
\arguments{
\item{fullSS_genome_build}{Genome build of the full summary statistics
 (\code{fullSS_path}). Can be "GRCH37" or "GRCH38" or one of their synonyms..
If \code{fullSS_genome_build==NULL} and \code{munged=TRUE},
infers genome build (hg19 vs. hg38)
from summary statistics using \link[MungeSumstats]{get_genome_builds}.}

\item{munged}{Whether \code{fullSS_path} have already been
standardised/filtered full summary stats
with \link[MungeSumstats]{format_sumstats}.
If \code{munged=FALSE} you'll need to provide the necessary
 column names to the \code{colmap} argument.}

\item{fullSS_path}{Path to the full summary statistics file (GWAS or QTL)
that you want to fine-map.
It is usually best to provide the absolute path rather
than the relative path.}

\item{sampled_snps}{Downsample the number of SNPs used when inferring genome
build to save time.}

\item{names_from_paths}{Infer the name of each item in \code{sumstats_list}
from its respective file path.
Only works if \code{sumstats_list} is a list of paths.}

\item{dbSNP}{version of dbSNP to be used (144 or 155). Default is 155.}

\item{nThread}{Number of threads to parallelise saving across.}

\item{verbose}{Print messages.}
}
\value{
Character string indicating genome build.
}
\description{
If \code{fullSS_genome_build==NULL} and \code{munged=TRUE},
infers genome build (hg19 vs. hg38)
from summary statistics using \link[MungeSumstats]{get_genome_builds}.
This can only be done with summary statistics that have already been
munged by \link[MungeSumstats]{format_sumstats}.
When \code{fullSS_genome_build} is a synonym of hg19 or hg38, this function
simply returns a standardized version of the user-provided
genome build.
}
\examples{
fullSS_path <- echodata::example_fullSS()
build <- check_genome(fullSS_genome_build="hg19",
                      fullSS_path=fullSS_path)
}
