% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_finemap.R
\name{check_necessary_cols}
\alias{check_necessary_cols}
\title{Check for necessary columns}
\usage{
check_necessary_cols(
  subset_DT,
  finemap_methods,
  sample_size = NULL,
  dataset_type = "GWAS",
  verbose = TRUE
)
}
\description{
Check for necessary columns
}
\examples{
BST1 <- echodata::BST1;
finemap_methods <- c("ABF","FINEMAP","SUSIE","POLYFUN_SUSIE")
finemap_methods <- check_necessary_cols(subset_DT=BST1, finemap_methods=finemap_methods)
}
