% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COJO.R
\name{COJO.conditional}
\alias{COJO.conditional}
\title{Run \emph{GCTA-COJO} conditional analysis}
\source{
\url{https://www.nature.com/articles/ng.2213}
\url{https://www.cell.com/ajhg/fulltext/S0002-9297(10)00598-7}
\url{https://cnsgenomics.com/software/gcta/#Overview}
}
\usage{
COJO.conditional(
  GCTA_path = system.file("tools/gcta_1.92.1beta5_mac/bin", "gcta64", package =
    "echolocatoR"),
  locus_dir,
  conditioned_snps,
  min_MAF = 0,
  diff_freq = 0.1,
  excluded_path
)
}
\description{
Condition on a given SNP (or list of SNPs)
and calculate residual effects of all other SNPs.
}
\seealso{
Other COJO: 
\code{\link{COJO.make_locus_subdir}()}
}
\concept{COJO}
