% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COJO.R
\name{COJO.make_locus_subdir}
\alias{COJO.make_locus_subdir}
\title{Make \emph{GCTA-COJO} path}
\source{
\url{https://www.nature.com/articles/ng.2213}
\url{https://www.cell.com/ajhg/fulltext/S0002-9297(10)00598-7}
\url{https://cnsgenomics.com/software/gcta/#Overview}
}
\usage{
COJO.make_locus_subdir(locus_dir)
}
\description{
Make \emph{GCTA-COJO} path
}
\examples{
\dontrun{
locus_dir <- echodata::locus_dir
cojo_dir <- COJO.make_locus_subdir(locus_dir=locus_dir)
}
}
\seealso{
Other COJO: 
\code{\link{COJO.conditional}()}
}
\concept{COJO}
\keyword{internal}
