% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COJO.R
\name{COJO}
\alias{COJO}
\title{Run \emph{GCTA-COJO}}
\usage{
COJO(
  subset_DT,
  fullSS_path,
  locus_dir,
  conditioned_snps,
  excluded_snps = "",
  min_MAF = 0,
  GCTA_path = system.file("tools/gcta_1.92.1beta5_mac/bin", "gcta64", package =
    "echolocatoR"),
  bfiles = "plink_tmp/plink",
  stepwise_procedure = TRUE,
  conditional_analysis = TRUE,
  snp_col = "SNP",
  freq_col = "Freq",
  effect_col = "Effect",
  stderr_col = "StdErr",
  pval_col = "P",
  A1_col = "A1",
  A2_col = "A2",
  full_genome = F
)
}
\description{
Main function to run either the conditional stepwise procedure (genome-wide)
or the conditional analysis (locus-specific) from \emph{GCTA-COJO}.
}
\seealso{
Other COJO
\url{https://www.nature.com/articles/ng.2213}
\url{https://www.cell.com/ajhg/fulltext/S0002-9297(10)00598-7}
\url{https://cnsgenomics.com/software/gcta/#Overview}: 
\code{\link{COJO.stepwise}()},
\code{\link{get_stepwise_results}()},
\code{\link{process_COJO_results}()}
}
\concept{COJO
\url{https://www.nature.com/articles/ng.2213}
\url{https://www.cell.com/ajhg/fulltext/S0002-9297(10)00598-7}
\url{https://cnsgenomics.com/software/gcta/#Overview}}
