% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COJO.R
\name{COJO.stepwise}
\alias{COJO.stepwise}
\title{Run \emph{GCTA-COJO} conditional stepwise procedure}
\usage{
COJO.stepwise(
  GCTA_path = system.file("tools/gcta_1.92.1beta5_mac/bin", "gcta64", package =
    "echolocatoR"),
  locus_dir,
  min_MAF,
  excluded_path
)
}
\description{
Runs the \emph{GCTA-COJO} conditional stepwise procedure to identify independent signals.
Should only be run on full, genome-wide summary stats at once.
}
\seealso{
Other COJO
\url{https://www.nature.com/articles/ng.2213}
\url{https://www.cell.com/ajhg/fulltext/S0002-9297(10)00598-7}
\url{https://cnsgenomics.com/software/gcta/#Overview}: 
\code{\link{COJO}()},
\code{\link{get_stepwise_results}()},
\code{\link{process_COJO_results}()}
}
\concept{COJO
\url{https://www.nature.com/articles/ng.2213}
\url{https://www.cell.com/ajhg/fulltext/S0002-9297(10)00598-7}
\url{https://cnsgenomics.com/software/gcta/#Overview}}
