% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MungeSumstats.R
\name{column_map}
\alias{column_map}
\title{Column name mappings for a given package}
\source{
https://github.com/neurogenomics/MungeSumstats
}
\usage{
column_map(package = "MungeSumstats")
}
\description{
Supports \pkg{MungeSumstats} and \pkg{echolocatoR}
}
