% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conda.R
\name{CONDA.activate_env}
\alias{CONDA.activate_env}
\title{Activate conda env}
\usage{
CONDA.activate_env(conda_env = "echoR", verbose = T)
}
\description{
Activate conda env
}
\examples{
CONDA.activate_env(conda_env="echoR")
}
\seealso{
Other conda: 
\code{\link{CONDA.create_echoR_env}()},
\code{\link{CONDA.env_from_yaml}()},
\code{\link{CONDA.find_env_Rlib}()},
\code{\link{CONDA.find_python_path}()},
\code{\link{CONDA.install}()}
}
\concept{conda}
