% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conda.R
\name{CONDA.create_env}
\alias{CONDA.create_env}
\title{Install necessary command line and python tools}
\usage{
CONDA.create_env(envname = "echoR", packages = c("plink", "tabix", "pandas"))
}
\arguments{
\item{envname}{The conda environment where you want to install \emph{plink}.
By default uses \emph{echoR}, the conda environment distributed with \emph{echolocatoR}.}
}
\description{
\describe{
\item{plink}{https://anaconda.org/bioconda/plink}
\item{tabix}{https://anaconda.org/bioconda/tabix}
}
}
\seealso{
Other conda: 
\code{\link{CONDA.activate_env}()},
\code{\link{CONDA.env_from_list}()},
\code{\link{CONDA.env_from_yaml}()},
\code{\link{CONDA.find_env_path}()},
\code{\link{CONDA.install}()}
}
\concept{conda}
