% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conda.R
\name{CONDA.env_from_list}
\alias{CONDA.env_from_list}
\title{Create new conda env from list of dependencies}
\usage{
CONDA.env_from_list(
  libraries = c("numpy", "scipy", "scikit-learn", "pandas", "tqdm", "pyarrow",
    "bitarray", "networkx", "rpy2", "r-ckmeans.1d.dp")
)
}
\description{
Create new conda env from list of dependencies
}
\seealso{
Other conda: 
\code{\link{CONDA.activate_env}()},
\code{\link{CONDA.create_env}()},
\code{\link{CONDA.env_from_yaml}()},
\code{\link{CONDA.find_env_path}()},
\code{\link{CONDA.install}()}
}
\concept{conda}
\keyword{internal}
