% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conda.R
\name{CONDA.env_from_yaml}
\alias{CONDA.env_from_yaml}
\title{Create conda env from yaml file}
\usage{
CONDA.env_from_yaml(
  yaml_path = system.file("conda", "echoR.yml", package = "echolocatoR")
)
}
\description{
Create conda env from yaml file
}
\seealso{
Other conda: 
\code{\link{CONDA.activate_env}()},
\code{\link{CONDA.create_echoR_env}()},
\code{\link{CONDA.find_env_Rlib}()},
\code{\link{CONDA.find_python_path}()},
\code{\link{CONDA.install}()}
}
\concept{conda}
\keyword{internal}
