% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conda.R
\name{CONDA.find_env_path}
\alias{CONDA.find_env_path}
\title{Identify the right conda env to use}
\usage{
CONDA.find_env_path(conda_env = "echoR")
}
\description{
Identify the right conda env to use
}
\seealso{
Other conda: 
\code{\link{CONDA.activate_env}()},
\code{\link{CONDA.create_env}()},
\code{\link{CONDA.env_from_list}()},
\code{\link{CONDA.env_from_yaml}()},
\code{\link{CONDA.install}()}
}
\concept{conda}
