% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conda.R
\name{CONDA.find_package}
\alias{CONDA.find_package}
\title{Find package executable}
\usage{
CONDA.find_package(package, conda_env = "echoR", verbose = T)
}
\description{
Find package executable
}
\examples{
# Tabix
tabix <- CONDA.find_package(package="tabix", conda_env="echoR")
tabix <- CONDA.find_package(package="tabix", conda_env=NULL)
# bgzip
bgzip <- CONDA.find_package(package="bgzip", conda_env="echoR")
bgzip <- CONDA.find_package(package="bgzip", conda_env=NULL)
}
\concept{CONDA}
\keyword{internal}
