% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conda.R
\name{CONDA.find_python_path}
\alias{CONDA.find_python_path}
\title{Find the python file for a specific env}
\usage{
CONDA.find_python_path(conda_env = "echoR")
}
\description{
Find the python file for a specific env
}
\seealso{
Other conda: 
\code{\link{CONDA.activate_env}()},
\code{\link{CONDA.create_echoR_env}()},
\code{\link{CONDA.env_from_list}()},
\code{\link{CONDA.env_from_yaml}()},
\code{\link{CONDA.find_env_Rlib}()},
\code{\link{CONDA.install}()}
}
\concept{conda}
