% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conda.R
\name{CONDA.install}
\alias{CONDA.install}
\title{Install conda if it's missing}
\usage{
CONDA.install()
}
\description{
Install conda if it's missing
}
\seealso{
Other conda functions: 
\code{\link{CONDA.activate_env}()},
\code{\link{CONDA.create_env}()}
}
\concept{conda functions}
