% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conda.R
\name{CONDA.install_reticulate}
\alias{CONDA.install_reticulate}
\title{Install reticulate}
\usage{
CONDA.install_reticulate(
  dependencies = c("devtools", "reticulate", "lattice", "jsonlite", "Matrix",
    "rappdirs", "Rcpp")
)
}
\description{
\emph{reticulate} often doesn't install very well via CRAN.
This function helps do it correctly.
}
