% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_locus_dir.R
\name{construct_locus_dir}
\alias{construct_locus_dir}
\title{Make locus-specific results folder}
\usage{
construct_locus_dir(results_dir = tempdir(), dataset_type, dataset_name, locus)
}
\description{
Make locus-specific results folder
}
\examples{
locus_dir <- echolocatoR:::construct_locus_dir(dataset_type="GWAS",
                                               dataset_name="Nalls2019",
                                               locus="BST1")
}
\seealso{
Other directory functions: 
\code{\link{construct_subset_path}()},
\code{\link{get_locus_dir}()},
\code{\link{get_study_dir}()}
}
\concept{directory functions}
\keyword{internal}
