% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_subset_path.R
\name{construct_subset_path}
\alias{construct_subset_path}
\title{Construct the path of the locus subset}
\usage{
construct_subset_path(
  subset_path = "auto",
  results_dir = tempdir(),
  dataset_type,
  dataset_name,
  locus = NULL,
  suffix = ".tsv.gz"
)
}
\description{
Construct the path of the locus subset
}
\examples{
subset_path <- echolocatoR:::construct_subset_path(dataset_type="GWAS",
                                                   dataset_name="Nalls2019",
                                                   locus="BST1")
}
\seealso{
Other directory functions: 
\code{\link{construct_locus_dir}()},
\code{\link{get_locus_dir}()},
\code{\link{get_study_dir}()}
}
\concept{directory functions}
\keyword{internal}
