% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CORCES_2020.bulkATACseq_peaks}
\alias{CORCES_2020.bulkATACseq_peaks}
\title{bulkATACseq peaks from Alzheimer's disease brain tissue}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 186559 rows and 10 columns.
}
\source{
\url{https://www.biorxiv.org/content/10.1101/2020.01.06.896159v1}
}
\usage{
CORCES_2020.bulkATACseq_peaks
}
\description{
Each row represents an individual peak identified in the bulk ATAC-seq data.
}
\details{
Data originally from \href{https://www.biorxiv.org/content/10.1101/2020.01.06.896159v1}{Corces et al. (bioRxiv)}, as of May 2020.
Specifically: \emph{STable2_Features_bulkATAC-seq_Peaks}
}
\seealso{
Other CORCES_2020: 
\code{\link{CORCES_2020.scATACseq_celltype_peaks}},
\code{\link{CORCES_2020.scATACseq_peaks}}
}
\concept{CORCES_2020}
\keyword{datasets}
