% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CORCES_2020.cicero_coaccessibility}
\alias{CORCES_2020.cicero_coaccessibility}
\title{Cicero_coaccessibility from human brain tissue}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 9795 rows and 14 columns.
}
\source{
\url{https://www.biorxiv.org/content/10.1101/2020.01.06.896159v1}
}
\usage{
CORCES_2020.cicero_coaccessibility
}
\description{
Cicero coaccessibility analysis for peaks that overlap SNPs derived from analysis of scATAC-seq data.
Each row represents an individual peak identified from the feature binarization analysis (see methods).
}
\details{
Data originally from \href{https://www.biorxiv.org/content/10.1101/2020.01.06.896159v1}{Corces et al. (bioRxiv)}, as of May 2020.
Specifically: \emph{STable10_Coacessibility_Peak_loop_connection}, \emph{Cicero Coaccessibility} sheet.
Peak_ID_Peak1 - A unique number that identifies the peak across supplementary tables.

\strong{Column dictionary}:
\describe{
\item{hg38_Chromosome_Peak1}{The hg38 chromosome of the first loop Peak.}
\item{hg38_Start_Peak1}{The hg38 start position of the first loop Peak.}
\item{hg38_Stop_Peak1}{The hg38 stop position of the first loop Peak.}
\item{Width_Peak1}{The width of the first loop Peak.}
\item{Peak_ID_Peak2}{A unique number that identifies the peak across supplementary tables.}
\item{hg38_Chromosome_Peak2}{The hg38 chromosome of the second loop Peak.}
\item{hg38_Start_Peak2}{The hg38 start position of the second loop Peak.}
\item{hg38_Stop_Peak2}{The hg38 stop position of the second loop Peak.}
\item{Width_Peak2}{The width of the second loop Peak.}
\item{Coaccessibility}{The coaccessibility correlation for the given peak pair.}
\item{Peak1_hasSNP}{A boolean variable determining whether the first peak overlaps a SNP from our AD/PD GWAS analyses.}
\item{Peak2_hasSNP}{A boolean variable determining whether the second peak overlaps a SNP from our AD/PD GWAS analyses.}
}
}
\examples{
\dontrun{
dat <- readxl::read_excel("~/Desktop/Fine_Mapping/echolocatoR/annotations/Coceres_2020/STable10_Coacessibility_Peak_loop_connection.xlsx", skip = 21, sheet=2)
CORCES_2020.cicero_coaccessibility <- data.table::data.table(dat)
usethis::use_data(CORCES_2020.cicero_coaccessibility)
}
}
\seealso{
Other CORCES_2020: 
\code{\link{CORCES_2020.HiChIP_FitHiChIP_loop_calls}},
\code{\link{CORCES_2020.bulkATACseq_peaks}},
\code{\link{CORCES_2020.get_ATAC_peak_overlap}()},
\code{\link{CORCES_2020.get_HiChIP_FitHiChIP_overlap}()},
\code{\link{CORCES_2020.prepare_bulkATAC_peak_overlap}()},
\code{\link{CORCES_2020.prepare_scATAC_peak_overlap}()},
\code{\link{CORCES_2020.scATACseq_celltype_peaks}},
\code{\link{CORCES_2020.scATACseq_peaks}}
}
\concept{CORCES_2020}
\keyword{datasets}
