% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CORCES_2020.R
\name{CORCES_2020.get_HiChIP_FitHiChIP_overlap}
\alias{CORCES_2020.get_HiChIP_FitHiChIP_overlap}
\title{Get overlap between data table of SNPs and HiChIP_FitHiChIP coaccessibility anchors}
\source{
\url{https://www.biorxiv.org/content/10.1101/2020.01.06.896159v1}
}
\usage{
CORCES_2020.get_HiChIP_FitHiChIP_overlap(finemap_dat, verbose = T)
}
\description{
Anchors are the genomic regions that have evidence of being
functionally connected to one another (coaccessible),
 e.g. enhancer-promoter interactions.
}
\seealso{
Other CORCES_2020: 
\code{\link{CORCES_2020.HiChIP_FitHiChIP_loop_calls}},
\code{\link{CORCES_2020.bulkATACseq_peaks}},
\code{\link{CORCES_2020.cicero_coaccessibility}},
\code{\link{CORCES_2020.get_scATAC_peak_overlap}()},
\code{\link{CORCES_2020.prepare_peak_overlap}()},
\code{\link{CORCES_2020.scATACseq_celltype_peaks}},
\code{\link{CORCES_2020.scATACseq_peaks}}
}
\concept{CORCES_2020}
