% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CORCES_2020.HiChIP_FitHiChIP_loop_calls}
\alias{CORCES_2020.HiChIP_FitHiChIP_loop_calls}
\title{FitHiChIP loop calls from human brain tissue}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 11542 rows and 11 columns.
}
\source{
\url{https://www.biorxiv.org/content/10.1101/2020.01.06.896159v1}
}
\usage{
CORCES_2020.HiChIP_FitHiChIP_loop_calls
}
\description{
FitHiChIP loop calls that overlap SNPs derived from analysis of H3K27ac HiChIP data.
Each row represents an individual peak identified from the feature binarization analysis (see methods).
}
\details{
Data originally from \href{https://www.biorxiv.org/content/10.1101/2020.01.06.896159v1}{Corces et al. (bioRxiv)}, as of May 2020.
Specifically: \emph{STable10_Coacessibility_Peak_loop_connection}, \emph{HiChIP FitHiChIP Loop Calls} sheet.

\strong{Column dictionary}
\describe{
\item{hg38_Chromosome_Anchor1}{The hg38 chromosome of the first loop Anchor.}
\item{hg38_Start_Anchor1}{The hg38 start position of the first loop Anchor.}
\item{hg38_Stop_Anchor1}{The hg38 stop position of the first loop Anchor.}
\item{Width_Anchor1}{The width of the first loop Anchor.}
\item{hg38_Chromosome_Anchor2}{The hg38 chromosome of the second loop Anchor.}
\item{hg38_Start_Anchor2}{The hg38 start position of the second loop Anchor.}
\item{hg38_Stop_Anchor2}{The hg38 stop position of the second loop Anchor.}
\item{Width_Anchor2}{The width of the second loop Anchor.}
\item{Score}{The -log10(q-value) of the loop call from FitHiChIP.}
\item{Anchor1_hasSNP}{A boolean variable determining whether the first anchor overlaps a SNP from our AD/PD GWAS analyses.}
\item{Anchor2_hasSNP}{A boolean variable determining whether the second anchor overlaps a SNP from our AD/PD GWAS analyses.}
}
}
\examples{
\dontrun{
dat <- readxl::read_excel("~/Desktop/Fine_Mapping/echolocatoR/annotations/Coceres_2020/STable10_Coacessibility_Peak_loop_connection.xlsx", skip = 19, sheet=1)
CORCES_2020.HiChIP_FitHiChIP_loop_calls <- data.table::data.table(dat)
usethis::use_data(CORCES_2020.HiChIP_FitHiChIP_loop_calls)
}
}
\seealso{
Other CORCES_2020: 
\code{\link{CORCES_2020.bulkATACseq_peaks}},
\code{\link{CORCES_2020.cicero_coaccessibility}},
\code{\link{CORCES_2020.get_ATAC_peak_overlap}()},
\code{\link{CORCES_2020.get_HiChIP_FitHiChIP_overlap}()},
\code{\link{CORCES_2020.prepare_bulkATAC_peak_overlap}()},
\code{\link{CORCES_2020.prepare_scATAC_peak_overlap}()},
\code{\link{CORCES_2020.scATACseq_celltype_peaks}},
\code{\link{CORCES_2020.scATACseq_peaks}}
}
\concept{CORCES_2020}
\keyword{datasets}
