% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CORCES_2020.R
\name{CORCES_2020.prepare_bulkATAC_peak_overlap}
\alias{CORCES_2020.prepare_bulkATAC_peak_overlap}
\title{Prepare data to plot overlap between datatable of SNPs and
cell-type-specific epigenomic peaks and coaccessibility data.}
\source{
\url{https://www.biorxiv.org/content/10.1101/2020.01.06.896159v1}
}
\usage{
CORCES_2020.prepare_bulkATAC_peak_overlap(
  merged_DT,
  FDR_filter = NULL,
  snp_filter = "Consensus_SNP==T",
  add_HiChIP_FitHiChIP = T,
  annotate_genes = F,
  verbose = T
)
}
\description{
Prepare data to plot overlap between datatable of SNPs and
cell-type-specific epigenomic peaks and coaccessibility data.
}
\examples{
data("merged_DT");
finemap_dat <- subset(merged_DT, Consensus_SNP)
dat_melt <- CORCES_2020.prepare_bulkATAC_peak_overlap(merged_DT=merged_DT)
}
\seealso{
Other CORCES_2020: 
\code{\link{CORCES_2020.HiChIP_FitHiChIP_loop_calls}},
\code{\link{CORCES_2020.bulkATACseq_peaks}},
\code{\link{CORCES_2020.cicero_coaccessibility}},
\code{\link{CORCES_2020.get_ATAC_peak_overlap}()},
\code{\link{CORCES_2020.get_HiChIP_FitHiChIP_overlap}()},
\code{\link{CORCES_2020.prepare_scATAC_peak_overlap}()},
\code{\link{CORCES_2020.scATACseq_celltype_peaks}},
\code{\link{CORCES_2020.scATACseq_peaks}}
}
\concept{CORCES_2020}
