% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CORCES_2020.scATACseq_celltype_peaks}
\alias{CORCES_2020.scATACseq_celltype_peaks}
\title{scATACseq cell type-specific peaks from Alzheimer's disease brain tissue.}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 221062 rows and 13 columns.

An object of class \code{data.table} (inherits from \code{data.frame}) with 221062 rows and 13 columns.

An object of class \code{data.table} (inherits from \code{data.frame}) with 221062 rows and 13 columns.
}
\source{
\url{https://www.biorxiv.org/content/10.1101/2020.01.06.896159v1}

\url{https://www.biorxiv.org/content/10.1101/2020.01.06.896159v1}

\url{https://www.biorxiv.org/content/10.1101/2020.01.06.896159v1}
}
\usage{
CORCES_2020.scATACseq_celltype_peaks

CORCES_2020.scATACseq_celltype_peaks

CORCES_2020.scATACseq_celltype_peaks
}
\description{
Each row represents an individual peak identified from the feature binarization analysis (see methods).

FitHiChIP loop calls that overlap SNPs derived from analysis of H3K27ac HiChIP data.
Each row represents an individual peak identified from the feature binarization analysis (see methods).

Cicero coaccessibility analysis for peaks that overlap SNPs derived from analysis of scATAC-seq data.
Each row represents an individual peak identified from the feature binarization analysis (see methods).
}
\details{
Data originally from \href{https://www.biorxiv.org/content/10.1101/2020.01.06.896159v1}{Corces et al. (bioRxiv)}, as of May 2020.
Specifically: \emph{STable6_Features_scATAC-seq_celltype_Peaks}

Data originally from \href{https://www.biorxiv.org/content/10.1101/2020.01.06.896159v1}{Corces et al. (bioRxiv)}, as of May 2020.
Specifically: \emph{STable10_Coacessibility_Peak_loop_connection}, \emph{HiChIP FitHiChIP Loop Calls} sheet.
\strong{Column dictionary:}
hg38_Chromosome_Anchor1 - The hg38 chromosome of the first loop Anchor.
hg38_Start_Anchor1 - The hg38 start position of the first loop Anchor.
hg38_Stop_Anchor1 - The hg38 stop position of the first loop Anchor.
Width_Anchor1 - The width of the first loop Anchor.
hg38_Chromosome_Anchor2 - The hg38 chromosome of the second loop Anchor.
hg38_Start_Anchor2 - The hg38 start position of the second loop Anchor.
hg38_Stop_Anchor2 - The hg38 stop position of the second loop Anchor.
Width_Anchor2 - The width of the second loop Anchor.
Score - The -log10(q-value) of the loop call from FitHiChIP.
Anchor1_hasSNP - A boolean variable determining whether the first anchor overlaps a SNP from our AD/PD GWAS analyses.
Anchor2_hasSNP - A boolean variable determining whether the second anchor overlaps a SNP from our AD/PD GWAS analyses.

Data originally from \href{https://www.biorxiv.org/content/10.1101/2020.01.06.896159v1}{Corces et al. (bioRxiv)}, as of May 2020.
Specifically: \emph{STable10_Coacessibility_Peak_loop_connection}, \emph{Cicero Coaccessibility} sheet.
Peak_ID_Peak1 - A unique number that identifies the peak across supplementary tables.
\strong{Columns dictionary}:
hg38_Chromosome_Peak1 - The hg38 chromosome of the first loop Peak.
hg38_Start_Peak1 - The hg38 start position of the first loop Peak.
hg38_Stop_Peak1 - The hg38 stop position of the first loop Peak.
Width_Peak1 - The width of the first loop Peak.
Peak_ID_Peak2 - A unique number that identifies the peak across supplementary tables.
hg38_Chromosome_Peak2 - The hg38 chromosome of the second loop Peak.
hg38_Start_Peak2 - The hg38 start position of the second loop Peak.
hg38_Stop_Peak2 - The hg38 stop position of the second loop Peak.
Width_Peak2 - The width of the second loop Peak.
Coaccessibility - The coaccessibility correlation for the given peak pair.
Peak1_hasSNP - A boolean variable determining whether the first peak overlaps a SNP from our AD/PD GWAS analyses.
Peak2_hasSNP - A boolean variable determining whether the second peak overlaps a SNP from our AD/PD GWAS analyses.
}
\seealso{
Other CORCES_2020: 
\code{\link{CORCES_2020.bulkATACseq_peaks}},
\code{\link{CORCES_2020.scATACseq_peaks}}

Other CORCES_2020: 
\code{\link{CORCES_2020.bulkATACseq_peaks}},
\code{\link{CORCES_2020.scATACseq_peaks}}

Other CORCES_2020: 
\code{\link{CORCES_2020.bulkATACseq_peaks}},
\code{\link{CORCES_2020.scATACseq_peaks}}
}
\concept{CORCES_2020}
\keyword{datasets}
