% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CORCES_2020.scATACseq_peaks}
\alias{CORCES_2020.scATACseq_peaks}
\title{scATACseq peaks from human brain tissue}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 359022 rows and 10 columns.
}
\source{
\url{https://www.biorxiv.org/content/10.1101/2020.01.06.896159v1}
}
\usage{
CORCES_2020.scATACseq_peaks
}
\description{
Each row represents an individual peak identified in the single-cell ATAC-seq data.
}
\details{
Data originally from \href{https://www.biorxiv.org/content/10.1101/2020.01.06.896159v1}{Corces et al. (bioRxiv)}, as of May 2020.
Specifically: \emph{STable5_Features_scATAC-seq_Peaks_all}
}
\examples{
\dontrun{
dat <- readxl::read_excel("~/Desktop/Fine_Mapping/echolocatoR/annotations/Coceres_2020/STable5_Features_scATAC-seq_Peaks_all.xlsx", skip = 18)
CORCES_2020.scATACseq_peaks <- data.table::data.table(dat)
usethis::use_data(CORCES_2020.scATACseq_peaks, overwrite = T)
}
}
\seealso{
Other CORCES_2020: 
\code{\link{CORCES_2020.HiChIP_FitHiChIP_loop_calls}},
\code{\link{CORCES_2020.bulkATACseq_peaks}},
\code{\link{CORCES_2020.cicero_coaccessibility}},
\code{\link{CORCES_2020.get_ATAC_peak_overlap}()},
\code{\link{CORCES_2020.get_HiChIP_FitHiChIP_overlap}()},
\code{\link{CORCES_2020.prepare_bulkATAC_peak_overlap}()},
\code{\link{CORCES_2020.prepare_scATAC_peak_overlap}()},
\code{\link{CORCES_2020.scATACseq_celltype_peaks}}
}
\concept{CORCES_2020}
\keyword{datasets}
