% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{createDT}
\alias{createDT}
\title{Interactive DT}
\usage{
createDT(DF, caption = "", scrollY = 400)
}
\description{
Generate an interactive data table with download buttons.
}
\seealso{
Other general: 
\code{\link{.rbind.file.list}()},
\code{\link{createDT_html}()},
\code{\link{dt.replace}()},
\code{\link{get_os}()},
\code{\link{startup_image}()},
\code{\link{tryFunc}()}
}
\concept{general}
\keyword{internal}
