% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{createDT_html}
\alias{createDT_html}
\title{Interactive DT (html)}
\usage{
createDT_html(DF, caption = "", scrollY = 400)
}
\description{
Generate an interactive data table with download buttons.
Use this function when manually constructing rmarkdown chunks using cat() in a for loop.
}
