% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe_2_vcf.R
\name{dataframe_2_vcf}
\alias{dataframe_2_vcf}
\title{Convert a dataframe to a vcf}
\usage{
dataframe_2_vcf(
  dat,
  samplename = "GWAS",
  reference_fasta = "/pd-omics/tools/polyfun/reference_fasta/hg19.fa.gz",
  output_vcf = "./GWAS_converted.vcf",
  conda_env = "echoR"
)
}
\description{
Convert a dataframe to a vcf
}
\examples{
data("merge_DT")

}
\concept{utils}
