% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dey_DeepLearning.R
\name{DEEPLEARNING.melt}
\alias{DEEPLEARNING.melt}
\title{Melt deep learning annotations into long-format}
\usage{
DEEPLEARNING.melt(
  ANNOT,
  model = c("Basenji", "BiClassCNN", "DeepSEA", "ChromHMM", "Roadmap", "Others"),
  metric_str = "max",
  replace_NA = NA,
  replace_negInf = NA
)
}
\description{
Melt deep learning annotations into long-format
}
\examples{
merged_dat <- merge_finemapping_results(dataset = "Data/GWAS/Nalls23andMe_2019", minimum_support = 0, LD_reference = "UKB")
ANNOT <- DEEPLEARNING.query(merged_dat=merged_dat, level="Allelic_Effect", type="annot")

annot.melt <- DEEPLEARNING.melt(ANNOT=ANNOT, metric_str="mean")
base_url = "/sc/arion/projects/pd-omics/brian/Fine_Mapping"
data.table::fwrite(annot.melt, file.path(base_url,"Data/GWAS/Nalls23andMe_2019/_genome_wide/Dey_DeepLearning/Nalls23andMe_2019.Dey_DeepLearning.snp_groups.mean.csv.gz"))
}
