% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dey_DeepLearning.R
\name{DEEPLEARNING.plot}
\alias{DEEPLEARNING.plot}
\title{Plot deep learning predictions}
\usage{
DEEPLEARNING.plot(
  annot.melt,
  snp_groups = c("GWAS lead", "UCS", "Consensus"),
  comparisons_filter = function(x) {     if ("Consensus" \%in\% x)          return(x) },
  model.metric = c("MEAN"),
  facet_formula = "Assay ~ Model + Tissue",
  shift_points = T,
  show_plot = T,
  save_path = F,
  height = 9,
  width = 8
)
}
\description{
Plot deep learning predictions
}
\examples{
base_url = "/sc/arion/projects/pd-omics/brian/Fine_Mapping"
annot.melt <- data.table::fread(file.path(base_url,"Data/GWAS/Nalls23andMe_2019/_genome_wide/Dey_DeepLearning/Nalls23andMe_2019.Dey_DeepLearning.snp_groups.mean.csv.gz"))
}
