% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dey_DeepLearning.R
\name{DEEPLEARNING.query}
\alias{DEEPLEARNING.query}
\title{Iteratively collect deep learning annotations}
\usage{
DEEPLEARNING.query(
  merged_dat,
  base_url = "/sc/arion/projects/pd-omics/data/Dey_DeepLearning",
  level = c("Variant_Level", "Allelic_Effect"),
  tissue = c("NTS", "Blood", "Brain"),
  model = c("Basenji", "BiClassCNN", "DeepSEA", "ChromHMM", "Roadmap", "Others"),
  mean_max = c("MEAN", "MAX"),
  type = c("annot", "ldscore"),
  nThread = 4,
  verbose = T
)
}
\description{
Iteratively collect deep learning annotations
}
\seealso{
Other DEEPLEARNING: 
\code{\link{DEEPLEARNING.query_multi_chr}()},
\code{\link{DEEPLEARNING.query_one_chr}()}
}
\concept{DEEPLEARNING}
