% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dep_graph.R
\name{dep_graph}
\alias{dep_graph}
\title{Dependency graph plot}
\usage{
dep_graph(
  pkg_name = "echolocatoR",
  deps = "echoverse",
  shape = c("image", "hexagon"),
  image = file.path("https://github.com/RajLabMSSM",
    "Fine_Mapping/blob/master/echolocatoR", "images/bat_silhouette.png?raw=true"),
  show_plot = TRUE,
  save_path = file.path(tempdir(), paste0(pkg_name, ".dep_graph.html")),
  background = "#25355c",
  verbose = TRUE
)
}
\arguments{
\item{pkg_name}{Package to search dependencies for.}

\item{deps}{A subset of of the main package's (\code{pkg_name} )
dependencies to include in the plot visualization.}

\item{shape}{: String. Default to 'ellipse'. The shape defines what the node looks like. There are two types of nodes. One type has the label inside of it and the other type has the label underneath it. The types with the label inside of it are: ellipse, circle, database, box, text. The ones with the label outside of it are: image, circularImage, diamond, dot, star, triangle, triangleDown, hexagon, square and icon.}

\item{image}{: List or String. Default to undefined. When the shape is set to image or circularImage, this option should be the URL to an image. If the image cannot be found, the brokenImage option can be used.
\itemize{
  \item{"unselected"}{ : String. Unselected (default) image URL.}
  \item{"selected"}{ : String. Selected image URL.}
}}

\item{show_plot}{Whether to print the plot.}

\item{save_path}{Path to save the plot to, as an interactive,
self-container HTML file.}

\item{background}{: Text string giving the html background color of the widget. Defaults to white.}

\item{verbose}{Print messages.}
}
\description{
Create a dependency graph between a set of R packages and plot them
as an interactive network. By default,
plots only packages within the
 \href{https://github.com/topics/echoverse}{\code{echoverse}}.
}
\examples{
vis <- echolocatoR::dep_graph()
}
