% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directory.R
\name{Directory_info}
\alias{Directory_info}
\title{Retrieve the location of summary stats files}
\usage{
Directory_info(dataset_name, variable = "fullSS.local")
}
\description{
Retrieve the location of summary stats files
}
\concept{directory}
