% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloaders.R
\name{downloader}
\alias{downloader}
\title{Downloaders}
\usage{
downloader(
  input_url,
  output_path,
  download_method = "axel",
  background = F,
  force_overwrite = F,
  quiet = F,
  show_progress = T,
  continue = T,
  nThread = 4,
  alternate = T,
  check_certificates = F,
  verbose = T
)
}
\description{
R wrapper for wget and axel
}
\seealso{
Other downloaders: 
\code{\link{axel}()},
\code{\link{wget}()}
}
\concept{downloaders}
\keyword{internal}
