% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{dt.replace}
\alias{dt.replace}
\title{Replace items within DT object}
\usage{
dt.replace(DT, target, replacement)
}
\value{
data.frame
}
\description{
Annoyingly, there is no native function to do simple find-and-replace in the `DT` library.
}
