% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt_replace.R
\name{dt_replace}
\alias{dt_replace}
\title{Replace items within DT object}
\usage{
dt_replace(DT, target, replacement)
}
\value{
data.frame
}
\description{
Annoyingly, there is no native function to do simple find-and-replace in the `DT` library.
}
\concept{general}
\keyword{internal}
