% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XGR.R
\name{DT_to_GRanges}
\alias{DT_to_GRanges}
\title{Convert data.table to GRanges object}
\usage{
DT_to_GRanges(subset_DT)
}
\description{
Convert data.table to GRanges object
}
\seealso{
Other XGR: 
\code{\link{GRanges_to_BED}()},
\code{\link{XGR.download_and_standardize}()},
\code{\link{XGR.import_annotations}()},
\code{\link{XGR.iterate_enrichment}()},
\code{\link{XGR.iterate_overlap}()},
\code{\link{XGR.merge_and_process}()},
\code{\link{XGR.plot_enrichment}()},
\code{\link{XGR.plot_peaks}()},
\code{\link{XGR.prepare_foreground_background}()}
}
\concept{XGR}
\keyword{internal}
