% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{echolocatoR-package}
\alias{echolocatoR}
\alias{echolocatoR-package}
\title{echolocatoR: Automated genomic fine-mapping}
\description{
Automated statistical and functional fine-mapping pipeline with extensive API access to datasets.
}
\details{
Fine-mapping methods are a powerful means of identifying causal variants underlying a given phenotype,
but are  underutilized due to the technical challenges of implementation.
\emph{echolocatoR} is an R package that automates end-to-end genomics fine-mapping, annotation,
and plotting in order to identify the most probable causal variants associated with a given phenotype.

It requires minimal input from users (a GWAS or QTL summary statistics file),
and includes a suite of statistical and functional fine-mapping tools.
It also includes extensive access to datasets
(linkage disequilibrium panels, epigenomic and genome-wide annotations, QTL).

The elimination of data gathering and preprocessing steps enables rapid fine-mapping of many loci in any phenotype,
 complete with locus-specific publication-ready figure generation.
 All results are merged into a single per-SNP summary file for additional downstream analysis
  and results sharing. Therefore \emph{echolocatoR} drastically reduces the barriers to identifying
  causal variants by making the entire fine-mapping pipeline rapid, robust and scalable.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/RajLabMSSM/echolocatoR}
  \item Report bugs at \url{https://github.com/RajLabMSSM/echolocatoR/issues}
}

}
\author{
\strong{Maintainer}: Brian Schilder \email{brian_schilder@alumni.brown.edu}

Other contributors:
\itemize{
  \item Jack Humphrey [contributor]
}

}
\keyword{internal}
