% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAIN.R
\name{echoR.filter_snps}
\alias{echoR.filter_snps}
\title{Filter SNPs}
\usage{
echoR.filter_snps(
  subset_DT,
  bp_distance,
  remove_variants,
  locus,
  verbose = T,
  min_POS = NULL,
  max_POS = NULL,
  max_snps = NULL,
  min_MAF = NULL,
  trim_gene_limits = F
)
}
\description{
Filter SNps by MAF, window size, min/max position, maxmimum number of SNPs, or gene coordinates.
You can also explicitly remove certain variants.
}
