% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{effective_sample_size}
\alias{effective_sample_size}
\title{Calculate effective sample size}
\usage{
effective_sample_size(finemap_DT)
}
\arguments{
\item{finemap_DT}{Preprocessed \emph{echolocatoR} locus subset file.
Requires the columns \strong{N_cases} and \strong{N_controls}.}
}
\description{
This is an often-cited approximation for the effective sample size in a case-control study.
i.e., this is the sample size required to identify an association with a
 quantitative trait with the same power as in your present study.
 (from email correpsondence with Omer Weissbrod)
}
\examples{
data("finemap_DT")
finemap_DT$effective_sample_size <- effective_sample_size(finemap_DT)
}
\keyword{internal}
