% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate.R
\name{epigenetics_enrichment}
\alias{epigenetics_enrichment}
\title{Test for enrichment of \code{HaploR} annotations}
\source{
\href{https://cran.r-project.org/web/packages/haploR/vignettes/haplor-vignette.html}{HaploR}
}
\usage{
epigenetics_enrichment(
  snp_list1,
  snp_list2,
  chisq = T,
  fisher = T,
  epigenetic_variables = c("Promoter_histone_marks", "Enhancer_histone_marks"),
  tissue_list = c("BRN", "BLD")
)
}
\description{
Test for enrichment of \code{HaploR} annotations
}
\seealso{
Other annotate: 
\code{\link{ANNOTATE.annotate_missense}()},
\code{\link{ANNOTATE.plot_missense}()},
\code{\link{SNPs_by_mutation_type}()},
\code{\link{biomart_geneInfo}()},
\code{\link{biomart_snp_info}()},
\code{\link{biomart_snps_to_geneInfo}()},
\code{\link{epigenetics_summary}()},
\code{\link{haploR.HaploReg}()},
\code{\link{haploR.regulomeDB}()}
}
\concept{annotate}
\keyword{internal}
