% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate.R
\name{epigenetics_summary}
\alias{epigenetics_summary}
\title{Summarise \code{HaploR} annotations}
\source{
\href{https://cran.r-project.org/web/packages/haploR/vignettes/haplor-vignette.html}{HaploR}
}
\usage{
epigenetics_summary(
  merged_results,
  tissue_list = c("BRN", "BLD"),
  epigenetic_variables = c("Promoter_histone_marks", "Enhancer_histone_marks")
)
}
\description{
Summarise \code{HaploR} annotations
}
\seealso{
Other annotate: 
\code{\link{SNPs_by_mutation_type}()},
\code{\link{biomart_geneInfo}()},
\code{\link{biomart_snp_info}()},
\code{\link{biomart_snps_to_geneInfo}()},
\code{\link{epigenetics_enrichment}()},
\code{\link{haploR.HaploReg}()},
\code{\link{haploR.regulomeDB}()}
}
\concept{annotate}
\keyword{internal}
