% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_snp_subset.R
\name{extract_snp_subset}
\alias{extract_snp_subset}
\title{Extract a subset of the summary stats}
\usage{
extract_snp_subset(
  locus = NULL,
  locus_dir,
  results_dir = NULL,
  fullSS_path,
  fullSS_genome_build = "hg19",
  subset_path,
  LD_reference,
  force_new_subset = FALSE,
  top_SNPs = "auto",
  bp_distance = 5e+05,
  chrom_col = "CHR",
  chrom_type = NULL,
  position_col = "POS",
  snp_col = "SNP",
  locus_col = "Locus",
  pval_col = "P",
  effect_col = "Effect",
  stderr_col = "StdErr",
  MAF_col = "MAF",
  freq_col = "Freq",
  tstat_col = "t-stat",
  A1_col = "A1",
  A2_col = "A2",
  gene_col = "gene",
  N_cases_col = "N_cases",
  N_controls_col = "N_controls",
  N_cases = NULL,
  N_controls = NULL,
  proportion_cases = "calculate",
  sample_size = NULL,
  superpopulation = "",
  min_POS = NA,
  max_POS = NA,
  genes_detected = FALSE,
  file_sep = "\\t",
  query_by = "coordinates",
  qtl_prefixes = NULL,
  remove_tmps = TRUE,
  conda_env = "echoR",
  verbose = TRUE
)
}
\description{
Use either \emph{tabix} or \emph{awk} to extract a locus subset
 from the full summary statistics file.
}
\seealso{
Other query functions: 
\code{\link{query_handler}()}
}
\concept{query functions}
\keyword{internal}
