% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fillNA_CS_PP.R
\name{fillNA_CS_PP}
\alias{fillNA_CS_PP}
\title{Fill NA in PP and CS columns}
\usage{
fillNA_CS_PP(finemap_dat, fillNA_CS = 0, fillNA_PP = 0)
}
\description{
Fill NA in PP and CS columns
}
\examples{
data("BST1");
finemap_dat <- BST1
# finemap_dat <- data.table::fread("~/Desktop/results/GWAS/Kunkle_2019.microgliaQTL/ABCA7/Multi-finemap/ABCA7_Kunkle_2019.microgliaQTL_Multi-finemap.tsv.gz")
finemap_dat <- fillNA_CS_PP(finemap_dat=finemap_dat)
}
\seealso{
Other general: 
\code{\link{createDT_html}()},
\code{\link{createDT}()},
\code{\link{dt.replace}()},
\code{\link{example_fullSS}()},
\code{\link{get_os}()},
\code{\link{get_sample_size}()},
\code{\link{startup_image}()},
\code{\link{tryFunc}()}
}
\concept{general}
\keyword{internal}
