% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{filter_snps}
\alias{filter_snps}
\title{Filter SNPs}
\usage{
filter_snps(
  subset_DT,
  bp_distance = 5e+05,
  remove_variants = F,
  min_POS = NA,
  max_POS = NA,
  max_snps = NULL,
  min_MAF = NULL,
  trim_gene_limits = F,
  verbose = T
)
}
\description{
Filter SNps by MAF, window size, min/max position, maxmimum number of SNPs, or gene coordinates.
You can also explicitly remove certain variants.
}
\examples{
data("BST1");
subset_DT <- filter_snps(subset_DT=BST1)
}
\seealso{
Other SNP filters: 
\code{\link{gene_trimmer}()},
\code{\link{limit_SNPs}()},
\code{\link{snps_to_condition}()},
\code{\link{subset_common_snps}()}
}
\concept{SNP filters}
\keyword{internal}
