% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FINEMAP.R
\name{FINEMAP.construct_data}
\alias{FINEMAP.construct_data}
\title{Prepare input files for \code{FINEMAP}}
\source{
\url{http://www.christianbenner.com}
}
\usage{
FINEMAP.construct_data(
  locus_dir,
  subset_DT,
  LD_matrix,
  nThread = 1,
  verbose = TRUE
)
}
\description{
Creates and saves 1) the summary stats file, and 2) the LD matrix.
"Columns beta and se are required for fine-mapping.
Column maf is needed to output posterior effect size estimates on the
allelic scale. All other columns are not required for computations and
can be specified arbitrarily."
}
\examples{
locus_dir <- echodata::locus_dir; BST1 <- echodata::BST1; data("BST1_LD_matrix");
finemap_DT <- BST1
dir.create(file.path(locus_dir,"FINEMAP"), showWarnings = FALSE, recursive = TRUE)
out <- echoLD::subset_common_snps(LD_matrix=LD_matrix, dat=finemap_DT)
LD_matrix <- out$LD
finemap_DT <- out$DT
dat_paths <- FINEMAP.construct_data(locus_dir=locus_dir, subset_DT=finemap_DT, LD_matrix=LD_matrix)
}
\seealso{
Other FINEMAP: 
\code{\link{FINEMAP.construct_master}()},
\code{\link{FINEMAP.find_executable}()},
\code{\link{FINEMAP.process_results}()},
\code{\link{FINEMAP}()}
}
\concept{FINEMAP}
\keyword{internal}
