% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FINEMAP.R
\name{FINEMAP.construct_master}
\alias{FINEMAP.construct_master}
\title{Construct the \code{FINAMAP} master file}
\usage{
FINEMAP.construct_master(locus_dir, n_samples, dataset_number = 1, file.k = NA)
}
\description{
Creates and saves the master file
which tells \code{FINEMAP} where to find each input file.
}
\examples{
data("locus_dir"); data("BST1");
finemap_DT <- BST1
n_samples <- get_sample_size(subset_DT = finemap_DT)
master_path <- FINEMAP.construct_master(locus_dir=locus_dir, n_samples=n_samples)
}
\seealso{
Other FINEMAP: 
\code{\link{FINEMAP.construct_data}()},
\code{\link{FINEMAP.find_executable}()},
\code{\link{FINEMAP.process_results}()},
\code{\link{FINEMAP}()}
}
\concept{FINEMAP}
\keyword{internal}
