% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FINEMAP.R
\name{FINEMAP.find_executable}
\alias{FINEMAP.find_executable}
\title{Retrieve location of \code{FINEMAP} executable}
\source{
\url{http://www.christianbenner.com}
}
\usage{
FINEMAP.find_executable(
  FINEMAP_path = NULL,
  OS = NULL,
  version = "1.4",
  verbose = TRUE
)
}
\description{
Retrieve location of \code{FINEMAP} executable
}
\examples{
FINEMAP_path <- FINEMAP.find_executable()
}
\seealso{
Other FINEMAP: 
\code{\link{FINEMAP.construct_data}()},
\code{\link{FINEMAP.construct_master}()},
\code{\link{FINEMAP.process_results}()},
\code{\link{FINEMAP}()}
}
\concept{FINEMAP}
\keyword{internal}
