% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{finemap_DT}
\alias{finemap_DT}
\title{echolocatoR output example}
\format{
data.table
\describe{
  \item{SNP}{SNP RSID}
  \item{CHR}{Chromosome}
  \item{POS}{Genomic positiion (in basepairs)}
  ...
}
}
\source{
\url{https://www.biorxiv.org/content/10.1101/388165v3}
}
\usage{
finemap_DT
}
\description{
An example results file after running \code{\link{finemap_loci}} on the \emph{BST1} locus.
Data originally comes from the Parkinson's disease GWAS
by \href{https://www.biorxiv.org/content/10.1101/388165v3}{Nalls et al. (bioRxiv)}.
}
\keyword{datasets}
