% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fullSS_dat}
\alias{fullSS_dat}
\title{"Example subset of full summary stats}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 38066 rows and 11 columns.
}
\source{
\url{https://www.biorxiv.org/content/10.1101/388165v3}
}
\usage{
fullSS_dat
}
\description{
Data originally comes from the Parkinson's disease GWAS
by \href{https://www.biorxiv.org/content/10.1101/388165v3}{Nalls et al. (bioRxiv)}.
}
\examples{
A subset of the full GWAS summary stats from Nalls et al. (2019)
\dontrun{
data("top_SNPs")
loci <- c("BST1","LRRK2","MEX3C")
fullSS_dat <- lapply(loci, function(locus){
  top_sub <- subset(top_SNPs, Locus==locus)
  TABIX(fullSS_path = "~/Desktop/nallsEtAl2019_allSamples_allVariants.mod.txt.gz",
        save_subset = F,
        subset_path = NULL,
        chrom_col = "CHR", position_col = "POS",
        chrom = top_sub$CHR[1],
        min_POS = top_sub$POS - 500000*4,
        max_POS = top_sub$POS + 500000*4)
}) \%>\% data.table::rbindlist()
usethis::use_data(fullSS_dat, overwrite=T)
}
}
\keyword{datasets}
