% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fullSS_munged}
\alias{fullSS_munged}
\title{Example subset of full summary stats (munged)}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 37033 rows and 11 columns.
}
\source{
\url{https://www.biorxiv.org/content/10.1101/388165v3}
}
\usage{
fullSS_munged
}
\description{
Data originally comes from the Parkinson's disease GWAS
by \href{https://www.biorxiv.org/content/10.1101/388165v3}{Nalls et al. (bioRxiv)}.
}
\details{
Munged using \pkg{MungeSumstats}.
}
\examples{
A subset of the full GWAS summary stats from Nalls et al. (2019)
\dontrun{
data("fullSS_dat")
tmp_file <- tempfile(fileext = ".tsv.gz")
data.table::fwrite(fullSS_dat, tmp_file)
fullSS_munged <- MungeSumstats::format_sumstats(tmp_file, ref_genome = "GRCh37", return_data = TRUE)
usethis::use_data(fullSS_munged, overwrite=TRUE)
}
}
\keyword{datasets}
