% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PAINTOR.R
\name{gather.transethnic.LD}
\alias{gather.transethnic.LD}
\title{Prepare transethnic PAINTOR results}
\usage{
gather.transethnic.LD(
  merged_DT,
  locus,
  conditions = c("Nalls23andMe_2019", "MESA_AFA", "MESA_CAU", "MESA_HIS")
)
}
\description{
Prepare transethnic PAINTOR results
}
\keyword{internal}
