% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAIN.R
\name{gene_trimmer}
\alias{gene_trimmer}
\title{Remove all SNPs outside of of a given gene.}
\usage{
gene_trimmer(subset_DT, gene, min_POS = NULL, max_POS = NULL)
}
\description{
Get the min/max coordinates of a given gene (including known regulatory regions, introns, and exons).
Remove any SNPs from the data.frame that fall outside these coordinates.
}
\examples{
subset_DT <- gene_trimmer(subset_DT, gene, min_POS=NULL, max_POS=NULL)
}
