% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{genome_wide_dir}
\alias{genome_wide_dir}
\title{Example results path for genome-wide results}
\format{
An object of class \code{character} of length 1.
}
\usage{
genome_wide_dir
}
\description{
Example results path for genome-wide results
}
\examples{
\dontrun{
genome_wide_dir <- "results/GWAS/Nalls23andMe_2019/_genome_wide"
usethis::use_data(genome_wide_dir, overwrite=T)
}
}
\keyword{datasets}
