% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_locus_dir.R
\name{get_locus_dir}
\alias{get_locus_dir}
\title{Extract the locus dir}
\usage{
get_locus_dir(subset_path)
}
\description{
Extract the locus dir
}
\seealso{
Other directory functions: 
\code{\link{construct_locus_dir}()},
\code{\link{construct_subset_path}()},
\code{\link{get_study_dir}()}
}
\concept{directory functions}
\keyword{internal}
