% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_multifinemap_path}
\alias{get_multifinemap_path}
\title{Create multi-finemap path}
\usage{
get_multifinemap_path()
}
\description{
Create multi-finemap path
}
\seealso{
Other directory functions: 
\code{\link{get_locus_dir}()},
\code{\link{get_study_dir}()},
\code{\link{get_subset_path}()},
\code{\link{make_locus_dir}()}
}
\concept{directory functions}
\keyword{internal}
