% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_os.R
\name{get_os}
\alias{get_os}
\title{Identify current operating system (OS).}
\usage{
get_os()
}
\description{
Identify current operating system (OS).
}
\seealso{
Other general: 
\code{\link{dt_replace}()},
\code{\link{get_sample_size}()}
}
\concept{general}
\keyword{internal}
