% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sample_size.R
\name{get_sample_size}
\alias{get_sample_size}
\title{Infer sample size from summary stats}
\usage{
get_sample_size(
  subset_DT,
  sample_size = NULL,
  effective_ss = TRUE,
  verbose = TRUE
)
}
\description{
Infer sample size from summary stats
}
\examples{
BST1 <- echodata::BST1
BST1 <- finemap_DT
subset_DT <- get_sample_size(subset_DT = finemap_DT)
}
\seealso{
Other general: 
\code{\link{dt_replace}()},
\code{\link{get_os}()}
}
\concept{general}
\keyword{internal}
