% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COJO.R
\name{get_stepwise_results}
\alias{get_stepwise_results}
\title{Gather stepwise conditional results}
\usage{
get_stepwise_results(cojo_dir)
}
\description{
Gather and preprocess the results of the \emph{GCTA-COJO} conditional stepwise procedure.
}
\seealso{
Other COJO
\url{https://www.nature.com/articles/ng.2213}
\url{https://www.cell.com/ajhg/fulltext/S0002-9297(10)00598-7}
\url{https://cnsgenomics.com/software/gcta/#Overview}: 
\code{\link{COJO.stepwise}()},
\code{\link{COJO}()},
\code{\link{process_COJO_results}()}
}
\concept{COJO
\url{https://www.nature.com/articles/ng.2213}
\url{https://www.cell.com/ajhg/fulltext/S0002-9297(10)00598-7}
\url{https://cnsgenomics.com/software/gcta/#Overview}}
