% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_study_dir}
\alias{get_study_dir}
\title{Extract the study dir}
\usage{
get_study_dir(locus_dir)
}
\description{
Extract the study dir
}
\seealso{
Other directory functions: 
\code{\link{get_locus_dir}()},
\code{\link{get_multifinemap_path}()},
\code{\link{get_subset_path}()},
\code{\link{make_locus_dir}()}
}
\concept{directory functions}
\keyword{internal}
