% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_subset_path.R
\name{get_subset_path}
\alias{get_subset_path}
\title{Construct the path of the locus subset}
\usage{
get_subset_path(
  subset_path = "auto",
  results_dir,
  dataset_type = "dataset_type",
  dataset_name = "dataset_name",
  locus = NULL,
  suffix = ".tsv.gz"
)
}
\description{
Construct the path of the locus subset
}
\examples{
subset_path <- echolocatoR:::get_subset_path(
  results_dir=file.path(tempdir(),"Data/GWAS/Nalls23andMe_2019/BST1"),
  locus="BST1")
}
\seealso{
Other directory functions: 
\code{\link{get_locus_dir}()},
\code{\link{get_study_dir}()},
\code{\link{make_locus_dir}()}
}
\concept{directory functions}
\keyword{internal}
