% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize.R
\name{get_UKB_MAF}
\alias{get_UKB_MAF}
\title{Get MAF from UK Biobank.}
\usage{
get_UKB_MAF(
  subset_DT,
  output.path = "./Data/Reference/UKB_MAF",
  force_new_maf = F
)
}
\description{
If \strong{MAF} column is missing,
download MAF from UK Biobank and use that instead.
}
\seealso{
Other standardizing functions: 
\code{\link{standardize_subset}()}
}
\concept{standardizing functions}
