% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggbio.R
\name{get_window_limits}
\alias{get_window_limits}
\title{Get window size limits for plot}
\usage{
get_window_limits(finemap_dat, plot.window = NULL, verbose = T)
}
\description{
Get window size limits for plot
}
\examples{
data("BST1");
xlims <- get_window_limits(finemap_dat=BST1, plot.window=50000)
}
\seealso{
Other plot: 
\code{\link{GGBIO.QTL_track}()},
\code{\link{GGBIO.SNP_track}()},
\code{\link{GGBIO.add_lines}()},
\code{\link{GGBIO.dot_summary}()},
\code{\link{GGBIO.invisible_legend}()},
\code{\link{GGBIO.plot}()},
\code{\link{GGBIO.track_heights_dict}()},
\code{\link{GGBIO.transcript_model_track}()}
}
\concept{plot}
\keyword{internal}
