% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggbio.R
\name{get_window_suffix}
\alias{get_window_suffix}
\title{Determine the plot suffix indicating its window size}
\usage{
get_window_suffix(finemap_dat, plot.zoom)
}
\description{
Determine the plot suffix indicating its window size
}
\examples{
data("BST1")
window_suffix <- get_window_suffix(finemap_dat=BST1, plot.zoom=1000)
window_suffix <- get_window_suffix(finemap_dat=BST1, plot.zoom=NULL)
window_suffix <- get_window_suffix(finemap_dat=BST1, plot.zoom="all")
window_suffix <- get_window_suffix(finemap_dat=BST1, plot.zoom="2x")
}
\seealso{
Other plot: 
\code{\link{GGBIO.QTL_track}()},
\code{\link{GGBIO.SNP_track}()},
\code{\link{GGBIO.add_lines}()},
\code{\link{GGBIO.dot_summary}()},
\code{\link{GGBIO.invisible_legend}()},
\code{\link{GGBIO.plot}()},
\code{\link{GGBIO.track_heights_dict}()},
\code{\link{GGBIO.transcript_model_track}()},
\code{\link{get_window_limits}()}
}
\concept{plot}
\keyword{internal}
