% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggbio.R
\name{GGBIO.add_lines}
\alias{GGBIO.add_lines}
\title{Add vertical lines}
\usage{
GGBIO.add_lines(trks, finemap_dat, alpha = 0.7, size = 0.3)
}
\description{
Adds vertical lines indicates key SNPs
}
\seealso{
Other plot: 
\code{\link{GGBIO.QTL_track}()},
\code{\link{GGBIO.SNP_track}()},
\code{\link{GGBIO.invisible_legend}()},
\code{\link{GGBIO.plot}()},
\code{\link{GGBIO.track_heights_dict}()},
\code{\link{GGBIO.transcript_model_track}()}
}
\concept{plot}
\keyword{internal}
