% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggbio.R
\name{GGBIO.dot_summary}
\alias{GGBIO.dot_summary}
\title{Multi-fine-map summary dot plot}
\usage{
GGBIO.dot_summary(finemap_dat, PP_threshold = 0.95, show_plot = T)
}
\description{
Multi-fine-map summary dot plot
}
\examples{
data("BST1")
gp <- GGBIO.dot_summary(finemap_dat=BST1)
}
\seealso{
Other plot: 
\code{\link{GGBIO.QTL_track}()},
\code{\link{GGBIO.SNP_track}()},
\code{\link{GGBIO.add_lines}()},
\code{\link{GGBIO.invisible_legend}()},
\code{\link{GGBIO.plot}()},
\code{\link{GGBIO.track_heights_dict}()},
\code{\link{GGBIO.transcript_model_track}()},
\code{\link{get_window_limits}()},
\code{\link{get_window_suffix}()}
}
\concept{plot}
