% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggbio.R
\name{GGBIO.invisible_legend}
\alias{GGBIO.invisible_legend}
\title{Make legend invisible}
\usage{
GGBIO.invisible_legend(gg)
}
\description{
Make legend invisible
}
\seealso{
Other plot: 
\code{\link{GGBIO.QTL_track}()},
\code{\link{GGBIO.SNP_track}()},
\code{\link{GGBIO.add_lines}()},
\code{\link{GGBIO.dot_summary}()},
\code{\link{GGBIO.plot}()},
\code{\link{GGBIO.track_heights_dict}()},
\code{\link{GGBIO.transcript_model_track}()},
\code{\link{get_window_limits}()},
\code{\link{get_window_suffix}()}
}
\concept{plot}
\keyword{internal}
