% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggbio.R
\name{GGBIO.plot}
\alias{GGBIO.plot}
\title{Multi-track plot of fine-mapping results}
\usage{
GGBIO.plot(
  finemap_DT,
  LD_matrix,
  locus = NULL,
  results_path = NULL,
  method_list = c("SUSIE", "FINEMAP", "PAINTOR", "PAINTOR_Fairfax"),
  mean.PP = T,
  PP_threshold = 0.95,
  consensus_thresh = 2,
  XGR_libnames = c("ENCODE_TFBS_ClusteredV3_CellTypes",
    "ENCODE_DNaseI_ClusteredV3_CellTypes", "Broad_Histone"),
  n_top_xgr = 5,
  ROADMAP = F,
  ROADMAP_query = NULL,
  n_top_roadmap = 7,
  annot_overlap_threshold = 5,
  Nott_sn_epigenome = F,
  show_regulatory_rects = T,
  show_placseq = T,
  plot_Nott_binwidth = 2500,
  Nott_bigwig_dir = NULL,
  save_plot = T,
  show_plot = T,
  max_transcripts = 1,
  plot_window = NULL,
  dpi = 200,
  height = 12,
  width = 10
)
}
\arguments{
\item{PP_threshold}{The minimum fine-mapped posterior probability for a SNP to be considered part of a Credible Set.
For example, \code{PP_threshold=.95} means that all Credible Set SNPs will be 95\% Credible Set SNPs.}

\item{plot_Nott_binwidth}{When including Nott et al. (2019) epigenomic data in the track plots,
adjust the bin width of the histograms.}

\item{Nott_bigwig_dir}{Instead of pulling Nott et al. (2019) epigenomic data from the UCSC Genome Browsers, use a set of local bigwig files.}

\item{max_transcripts}{Max number of transcripts per gene to show.}

\item{plot_window}{Zoom into the center of the locus when plotting (without editing the fine-mapping results file).}
}
\description{
Plots fine-mapping results from \emph{echolocatoR}.
Create separate tracks for the original results (-log10(P))
and each fine-mapping method (Posterior Probability).
Uses \code{\link{ggbio}}.
}
\examples{
data("finemap_DT")
data("LD_matrix")

# Using only XGR annotations
trk_plot <- GGBIO.plot(finemap_DT=finemap_DT, LD_matrix=LD_matrix, XGR_libnames=c("ENCODE_TFBS_ClusteredV3_CellTypes"))

# Using only Roadmap annotations
trk_plot <- GGBIO.plot(finemap_DT=finemap_DT, LD_matrix=LD_matrix, XGR_libnames=NULL, ROADMAP=T, ROADMAP_query="monocyte")

# Using only Nott_2019 annotations
trk_plot <- GGBIO.plot(finemap_DT=finemap_DT, LD_matrix=LD_matrix, XGR_libnames=NULL, Nott_sn_epigenome=T)
}
\seealso{
Other plot: 
\code{\link{SNP_track}()},
\code{\link{invisible_legend}()},
\code{\link{transcript_model_track}()}
}
\concept{plot}
\keyword{internal}
