% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggbio.R
\name{GGBIO.QTL_track}
\alias{GGBIO.QTL_track}
\title{Plot QTL data}
\usage{
GGBIO.QTL_track(
  gr.snp,
  QTL_prefix = "QTL",
  labels_subset = c("Lead SNP", "Credible Set", "Consensus SNP"),
  color_r2 = T,
  show.legend = T,
  PP_threshold = 0.95,
  sig_cutoff = 5e-08
)
}
\arguments{
\item{gr.snp}{\emph{echolocator} fine-mapping results
in \code{\link{GRanges}} format.}

\item{labels_subset}{Which SNP groups to label.}

\item{show.legend}{Whether to show legend.}

\item{PP_threshold}{The minimum fine-mapped posterior probability for a SNP to be considered part of a Credible Set.
For example, \code{PP_threshold=.95} means that all Credible Set SNPs will be 95\\% Credible Set SNPs.}
}
\description{
Plot QTL data
}
\examples{
data("BST1")
gr.snp <- DT_to_GRanges(dat)
## Create fake QTL P-values
gr.snp$fake_eQTL.P <- gr.snp$P  * c(1,.9,.7)
qtl.track <- GGBIO.QTL_track(gr.snp=gr.snp, QTL_prefix="fake_eQTL.", labels_subset=c("Lead SNP"), color_r2=F)
}
\seealso{
Other plot: 
\code{\link{GGBIO.SNP_track}()},
\code{\link{GGBIO.add_lines}()},
\code{\link{GGBIO.invisible_legend}()},
\code{\link{GGBIO.plot}()},
\code{\link{GGBIO.track_heights_dict}()},
\code{\link{GGBIO.transcript_model_track}()}
}
\concept{plot}
\keyword{internal}
