% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggbio.R
\name{GGBIO.SNP_track}
\alias{GGBIO.SNP_track}
\title{Track plot for SNPs}
\usage{
GGBIO.SNP_track(
  gr.snp,
  method = "original",
  labels_subset = c("Lead SNP", "Credible Set", "Consensus SNP"),
  color_r2 = F,
  show.legend = T,
  PP_threshold = 0.95,
  sig_cutoff = 5e-08
)
}
\arguments{
\item{gr.snp}{\emph{echolocator} fine-mapping results
in \code{\link{GRanges}} format.}

\item{method}{Which method to plot.}

\item{labels_subset}{Which SNP groups to label.}

\item{show.legend}{Whether to show legend.}

\item{PP_threshold}{The minimum fine-mapped posterior probability for a SNP to be considered part of a Credible Set.
For example, \code{PP_threshold=.95} means that all Credible Set SNPs will be 95\% Credible Set SNPs.}

\item{r2}{Whether color the points by LD r2.}
}
\description{
Uses \code{\link{ggbio}}.
}
\examples{
data("BST1");
gr.snp <- DT_to_GRanges(dat)
snp.track <- GGBIO.SNP_track(gr.snp=gr.snp, method = "original", color_r2=F)
}
\seealso{
Other plot: 
\code{\link{GGBIO.QTL_track}()},
\code{\link{GGBIO.add_lines}()},
\code{\link{GGBIO.invisible_legend}()},
\code{\link{GGBIO.plot}()},
\code{\link{GGBIO.track_heights_dict}()},
\code{\link{GGBIO.transcript_model_track}()}
}
\concept{plot}
\keyword{internal}
