% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggbio.R
\name{GGBIO.track_heights_dict}
\alias{GGBIO.track_heights_dict}
\title{Determine track heights}
\usage{
GGBIO.track_heights_dict(
  TRACKS_list,
  Manhattan = 0.2,
  Gene = 0.2,
  XGR = 0.5,
  Roadmap = 1,
  Nott_epigenome = 1,
  Nott_placseq = 0.33
)
}
\description{
Determine track heights
}
\seealso{
Other plot: 
\code{\link{GGBIO.QTL_track}()},
\code{\link{GGBIO.SNP_track}()},
\code{\link{GGBIO.add_lines}()},
\code{\link{GGBIO.invisible_legend}()},
\code{\link{GGBIO.plot}()},
\code{\link{GGBIO.transcript_model_track}()}
}
\concept{plot}
\keyword{internal}
