% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggbio.R
\name{GGBIO.transcript_model_track}
\alias{GGBIO.transcript_model_track}
\title{Track plot for transcript models}
\usage{
GGBIO.transcript_model_track(
  gr.snp_CHR,
  max_transcripts = 1,
  show.legend = T,
  show_plot = F
)
}
\arguments{
\item{gr.snp_CHR}{\emph{echolocator} fine-mapping results
in \code{\link{GRanges}} format.}

\item{max_transcripts}{Max number of transcripts per gene to show.}

\item{show.legend}{Whether to show legend.}
}
\description{
Uses \code{\link{ggbio}}.
}
\examples{
data("BST1"); data("LD_matrix");
finemap_DT <- BST1
#finemap_DT <- data.table::fread("~/Desktop/Fine_Mapping/Data/GWAS/Nalls23andMe_2019/LRRK2/Multi-finemap/Multi-finemap_results.txt")
#LD_matrix <- readRDS("~/Desktop/Fine_Mapping/Data/GWAS/Nalls23andMe_2019/LRRK2/plink/UKB_LD.RDS")
LD_SNP <- subset(finemap_DT, leadSNP==T)$SNP
LD_sub <- LD_with_leadSNP(LD_matrix = LD_matrix, LD_SNP = LD_SNP)
dat <- data.table:::merge.data.table(finemap_DT, LD_sub, by = "SNP", all.x = T)
gr.snp <- DT_to_GRanges(dat)
gr.snp_CHR <- biovizBase::transformDfToGr(dat, seqnames = "CHR", start = "POS", end = "POS")
track.genes <- GGBIO.transcript_model_track(gr.snp_CHR, max_transcripts=1)
}
\seealso{
Other plot: 
\code{\link{GGBIO.QTL_track}()},
\code{\link{GGBIO.SNP_track}()},
\code{\link{GGBIO.add_lines}()},
\code{\link{GGBIO.invisible_legend}()},
\code{\link{GGBIO.plot}()},
\code{\link{GGBIO.track_heights_dict}()}
}
\concept{plot}
\keyword{internal}
