% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finemapping_portal_API.R
\name{GITHUB.make_data_dict}
\alias{GITHUB.make_data_dict}
\title{echolocatoR Fine-mapping portal: data dictionary}
\usage{
GITHUB.make_data_dict(named_lists)
}
\description{
echolocatoR Fine-mapping portal: data dictionary
}
