% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finemapping_portal_API.R
\name{GITHUB.portal_metadata}
\alias{GITHUB.portal_metadata}
\title{echolocatoR Fine-mapping portal: metadata}
\usage{
GITHUB.portal_metadata(verbose = T)
}
\description{
echolocatoR Fine-mapping portal: metadata
}
