% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finemapping_portal_API.R
\name{GITHUB.portal_query}
\alias{GITHUB.portal_query}
\title{Search and download fine-mapping files}
\usage{
GITHUB.portal_query(
  dataset_types = NULL,
  datasets = NULL,
  phenotypes = NULL,
  loci = NULL,
  LD_panels = c("UKB", "1KGphase1", "1KGphase3"),
  file_types = c("multi_finemap", "LD", "plot"),
  results_dir = "./",
  overwrite = F,
  nThread = parallel::detectCores() - 2,
  verbose = T
)
}
\description{
Search the \href{https://github.com/RajLabMSSM/Fine_Mapping_Shiny}{echolocatoR Fine-mapping Portal}
for fine-mapping results, LD, and locus plots.
}
\examples{
local_finemap <- GITHUB.portal_query(dataset_types="GWAS", phenotypes = c("schizophrenia","parkinson"),file_types = "multi_finemap", LD_panels=c("UKB","1KGphase3"))
}
