% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GoShifter.R
\name{GOSHIFTER.create_snpmap}
\alias{GOSHIFTER.create_snpmap}
\title{Prepare SNP input for \emph{GoShifter}}
\usage{
GOSHIFTER.create_snpmap(finemap_dat, GS_results, nThread = 4, verbose = T)
}
\description{
Prepare SNP input for \emph{GoShifter}
}
\examples{
data("BST1"); data("locus_dir")
snpmap <- GOSHIFTER.create_snpmap(finemap_dat=BST1, GS_results=file.path(locus_dir,"GoShifter"))
}
\seealso{
Other GOSHIFTER: 
\code{\link{GOSHIFTER.create_LD}()}
}
\concept{GOSHIFTER}
