% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GoShifter.R
\name{GOSHIFTER}
\alias{GOSHIFTER}
\title{Run \code{GoShifter} enrichment pipeline}
\source{
\href{https://github.com/immunogenomics/goshifter}{GitHub}
\href{https://pubmed.ncbi.nlm.nih.gov/26140449/}{Publication}
}
\usage{
GOSHIFTER(
  locus_dir,
  snp_df,
  SNP.Group = "",
  goshifter_path = NULL,
  permutations = 1000,
  ROADMAP_search = "",
  ROADMAP_type = NA,
  chromatin_states = c("TssA"),
  R2_filter = 0.8,
  overlap_threshold = 1,
  force_new_goshifter = F,
  remove_tmps = T,
  verbose = T,
  save_results = T
)
}
\description{
\code{GoShifter}: Locus-specific enrichment tool.
}
