% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GoShifter.R
\name{GOSHIFTER.run}
\alias{GOSHIFTER.run}
\title{Run \code{GoShifter} enrichment test}
\source{
\href{https://github.com/immunogenomics/goshifter}{GitHub}
\href{https://pubmed.ncbi.nlm.nih.gov/26140449/}{Publication}
}
\usage{
GOSHIFTER.run(
  dat,
  locus_dir,
  GRlist,
  permutations = 1000,
  goshifter_path = NULL,
  chromatin_state = "TssA",
  R2_filter = 0.8,
  overlap_threshold = 1,
  remove_tmps = TRUE,
  verbose = TRUE
)
}
\description{
\code{GoShifter}: Locus-specific enrichment tool.
}
\examples{
BST1 <- echodata::BST1; locus_dir <- echodata::locus_dir
peaks <- echoannot::NOTT2019_get_epigenomic_peaks()
grl.peaks <- GenomicRanges::makeGRangesListFromDataFrame(peaks, split.field ="Cell_type")
GS_results <- GOSHIFTER.run(dat=subset(BST1, P<5e-8), locus_dir=locus_dir, GRlist=grl.peaks)
}
