% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XGR.R
\name{GRanges_to_BED}
\alias{GRanges_to_BED}
\title{Convert GRanges object to BED format and save}
\usage{
GRanges_to_BED(GR.annotations, output_path, sep = "\\t", nThread = 4, gzip = F)
}
\description{
Convert GRanges object to BED format and save
}
\seealso{
Other XGR: 
\code{\link{DT_to_GRanges}()},
\code{\link{XGR.download_and_standardize}()},
\code{\link{XGR.filter_assays}()},
\code{\link{XGR.filter_sources}()},
\code{\link{XGR.import_annotations}()},
\code{\link{XGR.iterate_enrichment}()},
\code{\link{XGR.iterate_overlap}()},
\code{\link{XGR.merge_and_process}()},
\code{\link{XGR.plot_enrichment}()},
\code{\link{XGR.plot_peaks}()},
\code{\link{XGR.prepare_foreground_background}()}
}
\concept{XGR}
\keyword{internal}
