% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate.R
\name{haploR.regulomeDB}
\alias{haploR.regulomeDB}
\title{Download SNP-wise annotations from RegulomeDB}
\source{
\href{https://cran.r-project.org/web/packages/haploR/vignettes/haplor-vignette.html}{HaploR}
}
\usage{
haploR.regulomeDB(snp_list, verbose = T, chunk_size = NA)
}
\description{
Download SNP-wise annotations from RegulomeDB
}
\seealso{
Other annotate: 
\code{\link{ANNOTATE.annotate_missense}()},
\code{\link{ANNOTATE.plot_missense}()},
\code{\link{SNPs_by_mutation_type}()},
\code{\link{biomart_geneInfo}()},
\code{\link{biomart_snp_info}()},
\code{\link{biomart_snps_to_geneInfo}()},
\code{\link{epigenetics_enrichment}()},
\code{\link{epigenetics_summary}()},
\code{\link{haploR.HaploReg}()},
\code{\link{merge_finemapping_results_each}()}
}
\concept{annotate}
\keyword{internal}
