% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IMPACT.R
\name{hierarchical_colors}
\alias{hierarchical_colors}
\title{Color tissues/cell types}
\usage{
hierarchical_colors(mat_meta)
}
\description{
Use the same color for groups of tissues/cell types but vary the shades within subgroups.
}
\keyword{internal}
