% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IMPACT.R
\name{IMPACT.compute_enrichment}
\alias{IMPACT.compute_enrichment}
\title{Compute enrichment of IMPACT scores}
\usage{
IMPACT.compute_enrichment(annot_melt, locus = NULL)
}
\description{
Conduct IMPACT enrichment between SNP groups a fine-mapping methods.
}
\keyword{internal}
