% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IMPACT.R
\name{IMPACT.get_ldscores}
\alias{IMPACT.get_ldscores}
\title{Download LD scores from IMPACT publication}
\source{
T Amariuta et al., IMPACT: Genomic Annotation of Cell-State-Specific Regulatory Elements Inferred from the Epigenome of Bound Transcription Factors. The American Journal of Human Genetics, 1–17 (2019).
}
\usage{
IMPACT.get_ldscores(chrom = NULL, subset_DT = NULL, nThread = 1)
}
\description{
Downloads per-SNP LD scores from LD SCore regression,
first conducted in the original IMPACT publication.
}
\keyword{internal}
