% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IMPACT.R
\name{IMPACT.get_top_annotations}
\alias{IMPACT.get_top_annotations}
\title{Get the top annotation(s)}
\usage{
IMPACT.get_top_annotations(
  ANNOT_MELT,
  snp.filter = "!is.na(IMPACT_score)",
  top_annotations = 1,
  force_one_annot_per_locus = F
)
}
\description{
Get the annotation(s) with the top mean IMPACT for a given set of SNPs.
}
\keyword{internal}
