% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IMPACT.R
\name{IMPACT.iterate_enrichment}
\alias{IMPACT.iterate_enrichment}
\title{Iterate IMPACT enrichment tests}
\usage{
IMPACT.iterate_enrichment(
  gwas_paths,
  annot_baseURL = "../../data/IMPACT/IMPACT707/Annotations"
)
}
\description{
Iterate IMPACT enrichment tests
}
\keyword{internal}
