% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IMPACT.R
\name{IMPACT.iterate_get_annotations}
\alias{IMPACT.iterate_get_annotations}
\title{Download IMPACT annotations}
\usage{
IMPACT.iterate_get_annotations(
  merged_DT,
  IMPACT_score_thresh = 0.1,
  baseURL = "/Volumes/Steelix/IMPACT/IMPACT707/Annotations",
  all_snps_in_range = T,
  top_annotations = F,
  force_one_annot_per_locus = F,
  snp.filter = "!is.na(SNP)"
)
}
\description{
Includes the raw annotation itself, as well as per-SNP IMPACT scores for each annotation.
}
\keyword{internal}
