% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IMPACT.R
\name{IMPACT.plot_impact_score}
\alias{IMPACT.plot_impact_score}
\title{Locus plot of IMPACT scores}
\usage{
IMPACT.plot_impact_score(annot_melt, save_path = FALSE, show_plot = TRUE)
}
\description{
Locus plot of IMPACT scores
}
\keyword{internal}
