% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IMPACT.R
\name{IMPACT.plot_top_annotations}
\alias{IMPACT.plot_top_annotations}
\title{IMPACT locus plots of top annotations}
\usage{
IMPACT.plot_top_annotations()
}
\description{
IMPACT locus plots of top annotations
}
\keyword{internal}
