% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IMPACT.R
\name{IMPACT.postprocess_annotations}
\alias{IMPACT.postprocess_annotations}
\title{Prepare IMPACT annotations}
\usage{
IMPACT.postprocess_annotations(ANNOT_MELT, order_loci = T, no_no_loci = NULL)
}
\description{
Transform IMPACT annotations from wide format (one row per SNP) to
long format (multiple rows per SNP).
}
\keyword{internal}
