% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IMPACT.R
\name{IMPACT.snp_group_boxplot}
\alias{IMPACT.snp_group_boxplot}
\title{IMPACT box plot
Box plot of IMPACT scores from each SNP group.}
\usage{
IMPACT.snp_group_boxplot(TOP_IMPACT_all)
}
\description{
IMPACT box plot
Box plot of IMPACT scores from each SNP group.
}
\keyword{internal}
