% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_topSNPs.R
\name{import_topSNPs}
\alias{import_topSNPs}
\title{Import top GWAS/QTL summary statistics}
\usage{
import_topSNPs(
  topSS,
  show_table = T,
  sheet = 1,
  munge = T,
  ref_genome = NULL,
  chrom_col = NULL,
  position_col = NULL,
  min_POS_col = NULL,
  max_POS_col = NULL,
  snp_col = NULL,
  pval_col = NULL,
  effect_col = NULL,
  locus_col = "Locus",
  grouping_vars = c("Locus"),
  gene_col = "Gene",
  remove_variants = NULL,
  nThread = 4,
  verbose = T
)
}
\arguments{
\item{show_table}{Create an interative data table.}

\item{sheet}{If the \emph{topSS} file is an excel sheet, you can specify which tab to use.
You can provide either a number to identify the tab by order,
or a string to identify the tab by name.}

\item{chrom_col}{Name of the chromosome column in the full summary stats file.
Can be "chr1" or "1" format.
(\emph{default: ="CHR"})}

\item{position_col}{Name of the genomic position column in the full summary stats file.
Must be in units of basepairs.
(\emph{default: ="POS"})}

\item{snp_col}{Name of the SNP RSID column in the full summary stats file.
(\emph{default: ="SNP"})}

\item{pval_col}{Name of the p-value column in the full summary stats file.
Raw p-values are preferred, but if not available corrected p-values (e.g. FDR) can be used instead.
(\emph{default: ="P"})}

\item{effect_col}{Name of the effect size column in the full summary stats file.
Effect size is preferred, but if not available other metrics like Beta for Odds Ratio can be used instead.
(\emph{default: ="Effect"})}

\item{locus_col}{Name of the locus column in the full summary stats file.
(\emph{default: ="Locus"})}

\item{grouping_vars}{The variables that you want to group by
such that each grouping_var combination has its own index SNP.
For example, if you want one index SNP per QTL eGene - GWAS locus pair, you could supply:
\code{grouping_vars=c("Locus","Gene")}.}

\item{gene_col}{An optional column to keep track of the causal gene(s) in each locus (if known).}

\item{remove_variants}{A list of variants to remove from the locus subset file.}

\item{verbose}{Whether \pkg{echolocatoR} should be verbose or silent.}

\item{top_SNPs}{Can be a data.frame with the top sumary stats per locus.
Alternatively, you can provide a path to the stored top summary stats file.
Can be in any tabular format (e.g. excel, .tsv, .csv, etc.).
This file should have one lead GWAS/QTL hits per locus.
If there is more than one SNP per locus, the one with the smallest p-value (then the largest effect size) is selected as the lead SNP.
The lead SNP will be used as the center of the locus when constructing the locus subset files.}
}
\description{
The resulting top_SNPs data.frame can be used to guide
the \code{\link{finemap_loci}} in querying and fine-mapping loci.
}
\seealso{
Other query functions: 
\code{\link{TABIX.convert_file}()},
\code{\link{TABIX.query}()},
\code{\link{TABIX}()},
\code{\link{extract_SNP_subset}()},
\code{\link{query_by_coordinates_merged}()},
\code{\link{query_by_coordinates}()},
\code{\link{query_by_gene}()},
\code{\link{query_by_probe}()},
\code{\link{query_fullSS}()},
\code{\link{query_handler}()}
}
\concept{query functions}
