% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_tricky_packages.R
\name{install_tricky_packages}
\alias{install_tricky_packages}
\title{Install tricky packages}
\usage{
install_tricky_packages()
}
\description{
Some packages don't install very well via the DESCRIPTION file
(e.g. wrong versions, wrong sources).
This function ensures they're actually installed properly.
}
