% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggbio.R
\name{invisible_legend}
\alias{invisible_legend}
\title{Make legend invisible}
\usage{
invisible_legend(gg)
}
\description{
Make legend invisible
}
\seealso{
Other plot: 
\code{\link{GGBIO.plot}()},
\code{\link{SNP_track}()},
\code{\link{transcript_model_track}()}
}
\concept{plot}
\keyword{internal}
