% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LD.R
\name{LD.calculate_LD}
\alias{LD.calculate_LD}
\title{Calculate LD}
\usage{
LD.calculate_LD(results_path, ld_window = 1000, ld_format = "r")
}
\arguments{
\item{results_path}{Locus-specific results directory.}

\item{ld_window}{Set --r/--r2 max variant ct pairwise distance (usu. 10).}

\item{ld_format}{Whether to produce an LD matrix with
r (\code{ld_format="r"}) or D' (\code{ld_format="D"}) as the pairwise SNP correlation metric.}
}
\description{
Calculate a pairwise LD matrix from a vcf file using \emph{plink}.
}
