% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LD.R
\name{LD.dprime_table}
\alias{LD.dprime_table}
\title{Calculate LD (D')}
\usage{
LD.dprime_table(SNP_list, plink_folder)
}
\description{
This appriach computes an LD matrix of D' (instead of r or r2) from a vcf.
See \code{\link{LD.run_plink_LD}} for a faster (but less flexible) alternative to computing LD.
}
