% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LD.R
\name{LD.fill_NA}
\alias{LD.fill_NA}
\title{Fill NAs in an LD matrix}
\usage{
LD.fill_NA(LD_matrix, fillNA = 0, verbose = F)
}
\description{
Trickier than it looks.
}
\examples{
\dontrun{
data("LD_matrix");
LD_matrix <- LD.fill_NA(LD_matrix)
}
}
