% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LD.R
\name{LD.filter_vcf}
\alias{LD.filter_vcf}
\title{Filter a vcf by min/max coordinates}
\usage{
LD.filter_vcf(subset_vcf, popDat, superpopulation, remove_tmp = F)
}
\arguments{
\item{subset_vcf}{Path to the locus subset vcf.}

\item{popDat}{The metadata file listing the superpopulation to which each sample belongs.}
}
\description{
Uses \emph{bcftools} to filter a vcf by min/max genomic coordinates (in basepairs).
}
