% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LD.R
\name{LD.filter_vcf_population}
\alias{LD.filter_vcf_population}
\title{Subset a vcf by superpopulation}
\usage{
LD.filter_vcf_population(
  subset_vcf,
  subset_DT,
  locus_dir,
  superpopulation,
  popDat
)
}
\arguments{
\item{subset_vcf}{Path to the locus subset vcf.}

\item{popDat}{The metadata file listing the superpopulation to which each sample belongs.}
}
\description{
Subset a vcf by superpopulation
}
\seealso{
Other LD: 
\code{\link{LD.1KG_download_vcf}()},
\code{\link{LD.1KG}()},
\code{\link{LD.LD_blocks}()},
\code{\link{LD.UKBiobank}()},
\code{\link{LD.calculate_LD}()},
\code{\link{LD.dprime_table}()},
\code{\link{LD.filter_vcf_rsamtools}()},
\code{\link{LD.filter_vcf}()},
\code{\link{LD.leadSNP_block}()},
\code{\link{LD.load_or_create}()},
\code{\link{LD.plink_LD}()},
\code{\link{LD.plot}()},
\code{\link{LD.read_bin}()},
\code{\link{LD.read_ld_table}()},
\code{\link{LD.run_plink_LD}()},
\code{\link{LD.translate_population}()},
\code{\link{LD.vcf_to_bed}()},
\code{\link{get_vcf_folder}()}
}
\concept{LD}
\keyword{internal}
