% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LD.R
\name{LD.get_lead_r2}
\alias{LD.get_lead_r2}
\title{Find correlates of the lead GWAS/QTL SNP}
\usage{
LD.get_lead_r2(
  finemap_dat,
  LD_matrix = NULL,
  fillNA = 0,
  LD_format = "matrix",
  verbose = T
)
}
\description{
Find correlates of the lead GWAS/QTL SNP
}
