% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LD.R
\name{LD.leadSNP_block}
\alias{LD.leadSNP_block}
\title{Identify the LD block in which the lead SNP resides}
\usage{
LD.leadSNP_block(leadSNP, LD_folder, LD_block_size = 0.7)
}
\description{
Identify the LD block in which the lead SNP resides
}
\seealso{
Other LD: 
\code{\link{LD.1KG_download_vcf}()},
\code{\link{LD.1KG}()},
\code{\link{LD.LD_blocks}()},
\code{\link{LD.UKBiobank}()},
\code{\link{LD.calculate_LD}()},
\code{\link{LD.construct_subset_vcf_name}()},
\code{\link{LD.custom_panel}()},
\code{\link{LD.dprime_table}()},
\code{\link{LD.filter_vcf_gaston}()},
\code{\link{LD.filter_vcf}()},
\code{\link{LD.get_locus_vcf_folder}()},
\code{\link{LD.get_plink_MAF}()},
\code{\link{LD.index_vcf}()},
\code{\link{LD.load_or_create}()},
\code{\link{LD.plink_LD}()},
\code{\link{LD.plink_file}()},
\code{\link{LD.plot_LD}()},
\code{\link{LD.query_vcf}()},
\code{\link{LD.rds_to_npz}()},
\code{\link{LD.read_bin}()},
\code{\link{LD.read_ld_table}()},
\code{\link{LD.run_plink_LD}()},
\code{\link{LD.run_snpstats_LD}()},
\code{\link{LD.save_LD_matrix}()},
\code{\link{LD.translate_population}()},
\code{\link{LD.vcf_to_bed}()},
\code{\link{saveSparse}()}
}
\concept{LD}
\keyword{internal}
