% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LD.R
\name{LD.load_or_create}
\alias{LD.load_or_create}
\title{Procure an LD matrix for fine-mapping}
\usage{
LD.load_or_create(
  locus_dir,
  subset_DT,
  locus,
  force_new_LD = F,
  LD_reference = "1KG_Phase1",
  superpopulation = "EUR",
  download_reference = T,
  download_method = "direct",
  min_r2 = 0,
  LD_block = F,
  LD_block_size = 0.7,
  min_Dprime = F,
  remove_correlates = F,
  fillNA = 0,
  verbose = T,
  server = F,
  remove_tmps = T,
  nThreads = 4
)
}
\arguments{
\item{subset_DT}{The locus subset of the full summary stats file.}

\item{force_new_LD}{By default, if an LD matrix file for a given locus is already present,
then \emph{echolocatoR} will just use the preexisting file.
Set \code{force_new_LD=T} to override this and extract a new subset.}

\item{download_reference}{When acquiring LD matrixes,
the default is to delete the full vcf or npz files after \emph{echolocator} has extracted the necssary subset.
However, if you wish to keep these full files (which can be quite large) set \code{download_reference=T}.}

\item{min_r2}{Remove any SNPs are below the LD r2 threshold with the lead SNP within their respective locus.}

\item{LD_block}{Calculate LD blocks with \emph{plink} and only include the block to which the lead SNP belongs.}

\item{LD_block_size}{Adjust the granularity of block sizes when \code{LD_block=T}.}

\item{min_Dprime}{Remove any SNPs are below the LD D' threshold with the lead SNP within their respective locus.
This is paramter currently only works when \code{LD_reference!="UKB"}.}

\item{remove_correlates}{If \code{remove_correlates} is set to a value between 0-1,
removes any SNPs that are in LD with any of the \code{remove_variants} above the threshold provided by \code{remove_correlates}.}

\item{verbose}{Whether \emph{echolocatoR} should be verbose or silent.}

\item{server}{Whether \emph{echolocatoR} is being run on a computing cluster/server or on a local machine.}

\item{remove_tmps}{Whether to remove any temporary files (e.g. FINEMAP output files) after the pipeline is done running.}
}
\value{
A symmetric LD matrix of pairwise \emph{r} values.
}
\description{
Calculate and/or query linkage disequilibrium (LD) from reference panels (UK Biobank, 1000 Genomes),
or user-supplied datasets.
}
\details{
Options:
\itemize{
\item Download pre-computed LD matrix from UK Biobank.
\item Download raw vcf file from 1KG and compute LD on the fly.
\item Compute LD fon the fly from a user-supplied vcf file.
\item Use a user-supplied pre-computed LD-matrix.
}
}
\examples{
\dontrun{
data("BST1"); data("locus_dir");
locus_dir <- file.path("~/Desktop",locus_dir)
# UK Biobank LD
LD.load_or_create(locus_dir=locus_dir, subset_DT=BST1, locus="BST1", )
}
}
