% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LD.R
\name{LD.plink_LD}
\alias{LD.plink_LD}
\title{Calculate LD}
\usage{
LD.plink_LD(
  leadSNP,
  plink_folder,
  min_r2 = F,
  min_Dprime = F,
  remove_correlates = F,
  fillNA = 0
)
}
\description{
Use \emph{plink} to calculate LD from a vcf.
}
\seealso{
Other LD: 
\code{\link{LD.1KG_download_vcf}()},
\code{\link{LD.1KG}()},
\code{\link{LD.LD_blocks}()},
\code{\link{LD.UKBiobank}()},
\code{\link{LD.calculate_LD}()},
\code{\link{LD.dprime_table}()},
\code{\link{LD.filter_vcf_population}()},
\code{\link{LD.filter_vcf_rsamtools}()},
\code{\link{LD.filter_vcf}()},
\code{\link{LD.leadSNP_block}()},
\code{\link{LD.load_or_create}()},
\code{\link{LD.plot}()},
\code{\link{LD.read_bin}()},
\code{\link{LD.read_ld_table}()},
\code{\link{LD.run_plink_LD}()},
\code{\link{LD.translate_population}()},
\code{\link{LD.vcf_to_bed}()},
\code{\link{get_vcf_folder}()}
}
\concept{LD}
\keyword{internal}
