% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LD.R
\name{LD.plot}
\alias{LD.plot}
\title{Plot a subset of the LD matrix}
\usage{
LD.plot(LD_matrix, subset_DT, span = 10)
}
\arguments{
\item{span}{This is very computationally intensive,
so you need to limit the number of SNPs with span.
If \code{span=10}, only 10 SNPs upstream and 10 SNPs downstream of the lead SNP will be plotted.}
}
\description{
Uses \gaston{gaston} to plot a SNP-annotated LD matrix.
}
