% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LD.R
\name{LD.plot_LD}
\alias{LD.plot_LD}
\title{Plot a subset of the LD matrix}
\usage{
LD.plot_LD(
  LD_matrix,
  subset_DT,
  span = 10,
  method = c("gaston", "heatmap", "image")
)
}
\arguments{
\item{span}{This is very computationally intensive,
so you need to limit the number of SNPs with span.
If \code{span=10}, only 10 SNPs upstream and 10 SNPs downstream of the lead SNP will be plotted.}
}
\description{
Uses \code{gaston} to plot a SNP-annotated LD matrix.
}
\examples{
\dontrun{
data("BST1");
LD_matrix <- readRDS("/Volumes/Steelix/fine_mapping_files/GWAS/Nalls23andMe_2019/BST1/plink/UKB_LD.RDS")
LD.plot_LD(LD_matrix=LD_matrix, subset_DT=BST1)
}
}
\seealso{
Other LD: 
\code{\link{LD.1KG_download_vcf}()},
\code{\link{LD.1KG}()},
\code{\link{LD.LD_blocks}()},
\code{\link{LD.UKBiobank}()},
\code{\link{LD.calculate_LD}()},
\code{\link{LD.construct_subset_vcf_name}()},
\code{\link{LD.custom_panel}()},
\code{\link{LD.dprime_table}()},
\code{\link{LD.filter_vcf_gaston}()},
\code{\link{LD.filter_vcf}()},
\code{\link{LD.get_locus_vcf_folder}()},
\code{\link{LD.index_vcf}()},
\code{\link{LD.leadSNP_block}()},
\code{\link{LD.load_or_create}()},
\code{\link{LD.plink_LD}()},
\code{\link{LD.plink_file}()},
\code{\link{LD.query_vcf}()},
\code{\link{LD.rds_to_npz}()},
\code{\link{LD.read_bin}()},
\code{\link{LD.read_ld_table}()},
\code{\link{LD.run_plink_LD}()},
\code{\link{LD.save_LD_matrix}()},
\code{\link{LD.translate_population}()},
\code{\link{LD.vcf_to_bed}()},
\code{\link{saveSparse}()}
}
\concept{LD}
\keyword{internal}
