% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LD.R
\name{LD.query_vcf}
\alias{LD.query_vcf}
\title{Query vcf file}
\usage{
LD.query_vcf(
  subset_DT,
  vcf_URL,
  locus_dir,
  LD_reference,
  whole_vcf = F,
  force_new_vcf = F,
  remove_original_vcf = F,
  download_method = "wget",
  query_by_regions = F,
  nThread = 4,
  conda_env = "echoR",
  verbose = T
)
}
\description{
Query vcf file
}
\examples{
data("locus_dir"); data("BST1");

# Custom
LD_reference <- "~/Desktop/results/Reference/custom_panel_chr4.vcf"
vcf_file <- LD.index_vcf(vcf_file=LD_reference)
vcf_subset <- LD.query_vcf(subset_DT=BST1, locus_dir=locus_dir, vcf_URL=vcf_file, LD_reference=LD_reference, force_new_vcf=T)
}
\seealso{
Other LD: 
\code{\link{LD.1KG_download_vcf}()},
\code{\link{LD.1KG}()},
\code{\link{LD.LD_blocks}()},
\code{\link{LD.UKBiobank}()},
\code{\link{LD.calculate_LD}()},
\code{\link{LD.construct_subset_vcf_name}()},
\code{\link{LD.custom_panel}()},
\code{\link{LD.dprime_table}()},
\code{\link{LD.filter_LD}()},
\code{\link{LD.filter_vcf_gaston}()},
\code{\link{LD.filter_vcf}()},
\code{\link{LD.get_locus_vcf_folder}()},
\code{\link{LD.index_vcf}()},
\code{\link{LD.leadSNP_block}()},
\code{\link{LD.load_or_create}()},
\code{\link{LD.plink_LD}()},
\code{\link{LD.plink_file}()},
\code{\link{LD.plot_LD}()},
\code{\link{LD.rds_to_npz}()},
\code{\link{LD.read_bin}()},
\code{\link{LD.read_ld_table}()},
\code{\link{LD.run_plink_LD}()},
\code{\link{LD.save_LD_matrix}()},
\code{\link{LD.snpstats_get_LD}()},
\code{\link{LD.snpstats_get_MAF}()},
\code{\link{LD.translate_population}()},
\code{\link{LD.vcf_to_bed}()},
\code{\link{LDlinkR.LDproxy_batch}()},
\code{\link{popDat_1KGphase1}},
\code{\link{popDat_1KGphase3}},
\code{\link{saveSparse}()}
}
\concept{LD}
\keyword{internal}
