% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LD.R
\name{LD.read_bin}
\alias{LD.read_bin}
\title{Create LD matrix from plink output}
\usage{
LD.read_bin(ld.path)
}
\arguments{
\item{ld.path}{Directory that contains the bin/bim files.}
}
\description{
Depending on which parameters you give \emph{plink} when calculating LD, you get different file outputs.
When it produces bin and bim files, use this function to create a proper LD matrix.
For example, this happens when you try to calculate D' with the \code{--r dprime-signed} flag (instead of just r).
}
