% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LD.R
\name{LD.read_ld_table}
\alias{LD.read_ld_table}
\title{Create LD matrix from plink output.}
\usage{
LD.read_ld_table(ld.path, snp.subset = F)
}
\description{
Depending on which parameters you give \emph{plink} when calculating LD, you get different file outputs.
When it produces an LD table, use this function to create a proper LD matrix.
}
