% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LD.R
\name{LD.run_plink_LD}
\alias{LD.run_plink_LD}
\title{Calculate LD (r or r2)}
\usage{
LD.run_plink_LD(bim, plink_folder, r_format = "r")
}
\arguments{
\item{bim}{A bim file produced by \emph{plink}}

\item{plink_folder}{Locus-specific LD output folder.}

\item{r_format}{Whether to fill the matrix with \code{r} or \code{r2}.}
}
\description{
This appriach computes and LD matrix of r or r2 (instead of D') from a vcf.
See \code{\link{LD.dprime_table}} for a slower (but more flexible) alternative to computing LD.
}
\seealso{
Other LD: 
\code{\link{LD.1KG_download_vcf}()},
\code{\link{LD.1KG}()},
\code{\link{LD.LD_blocks}()},
\code{\link{LD.UKBiobank}()},
\code{\link{LD.calculate_LD}()},
\code{\link{LD.dprime_table}()},
\code{\link{LD.filter_vcf_population}()},
\code{\link{LD.filter_vcf_rsamtools}()},
\code{\link{LD.filter_vcf}()},
\code{\link{LD.leadSNP_block}()},
\code{\link{LD.load_or_create}()},
\code{\link{LD.plink_LD}()},
\code{\link{LD.plot}()},
\code{\link{LD.read_bin}()},
\code{\link{LD.read_ld_table}()},
\code{\link{LD.translate_population}()},
\code{\link{LD.vcf_to_bed}()},
\code{\link{get_vcf_folder}()}
}
\concept{LD}
\keyword{internal}
