% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LD.R
\name{LD.snpstats_get_LD}
\alias{LD.snpstats_get_LD}
\title{Get LD using \pkg{snpStats} package}
\source{
\href{https://www.bioconductor.org/packages/release/bioc/html/snpStats.html}{snpStats Bioconductor page}
\href{https://www.bioconductor.org/packages/release/bioc/vignettes/snpStats/inst/doc/ld-vignette.pdf}{LD tutorial}
}
\usage{
LD.snpstats_get_LD(
  LD_folder,
  plink_prefix = "plink",
  select.snps = NULL,
  stats = c("R"),
  symmetric = T,
  depth = "max",
  nThread = 4,
  verbose = T
)
}
\arguments{
\item{LD_folder}{Locus-specific LD output folder.}

\item{stats}{
    A character vector specifying the linkage disequilibrium measures to
    be calculated. This should contain one or more of the strings: \code{"LLR"},
    \code{"OR"}, \code{"Q"}, \code{"Covar"}, \code{"D.prime"},
    \code{"R.squared"}, ad \code{"R"}
  }

\item{symmetric}{
    When no \code{y} argument is supplied this argument controls the format of
    the output band matrices. If \code{TRUE}, symmetric matrices are
    returned and, otherwise, an upper triangular matrices are returned
  }

\item{depth}{
    When \code{y} is not supplied, this parameter is mandatory and
    controls the maximum lag between columns of \code{x}
    considered. Thus, LD statistics are calculated between \code{x[,i]}
    and \code{x[,j]} only if \code{i} and \code{j} differ by no more
    than \code{depth}
  }
}
\description{
Get LD using \pkg{snpStats} package
}
\examples{
subset_DT <- data.table::fread("/pd-omics/brian/Fine_Mapping/Data/GWAS/Kunkle_2019/ABCA7/Multi-finemap/ABCA7.Kunkle_2019.1KGphase3_LD.Multi-finemap.tsv.gz")
LD_folder <- "/pd-omics/brian/Fine_Mapping/Data/GWAS/Kunkle_2019/ABCA7/LD"
LD_matrix <- LD.snpstats_get_LD(LD_folder=LD_folder, select.snps=subset_DT$SNP)
}
\seealso{
Other LD: 
\code{\link{LD.1KG_download_vcf}()},
\code{\link{LD.1KG}()},
\code{\link{LD.LD_blocks}()},
\code{\link{LD.UKBiobank}()},
\code{\link{LD.calculate_LD}()},
\code{\link{LD.construct_subset_vcf_name}()},
\code{\link{LD.custom_panel}()},
\code{\link{LD.dprime_table}()},
\code{\link{LD.filter_LD}()},
\code{\link{LD.filter_vcf_gaston}()},
\code{\link{LD.filter_vcf}()},
\code{\link{LD.get_locus_vcf_folder}()},
\code{\link{LD.index_vcf}()},
\code{\link{LD.leadSNP_block}()},
\code{\link{LD.load_or_create}()},
\code{\link{LD.plink_LD}()},
\code{\link{LD.plink_file}()},
\code{\link{LD.plot_LD}()},
\code{\link{LD.query_vcf}()},
\code{\link{LD.rds_to_npz}()},
\code{\link{LD.read_bin}()},
\code{\link{LD.read_ld_table}()},
\code{\link{LD.run_plink_LD}()},
\code{\link{LD.save_LD_matrix}()},
\code{\link{LD.snpstats_get_MAF}()},
\code{\link{LD.translate_population}()},
\code{\link{LD.vcf_to_bed}()},
\code{\link{saveSparse}()}
}
\concept{LD}
\keyword{internal}
