% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LD.R
\name{LD.snpstats_get_MAF}
\alias{LD.snpstats_get_MAF}
\title{Get MAF using \pkg{snpStats} package}
\source{
\href{https://www.bioconductor.org/packages/release/bioc/html/snpStats.html}{snpStats Bioconductor page}
}
\usage{
LD.snpstats_get_MAF(
  subset_DT,
  LD_folder,
  plink_prefix = "plink",
  force_new_MAF = F,
  nThread = 4,
  verbose = T
)
}
\arguments{
\item{LD_folder}{Locus-specific LD output folder.}
}
\description{
Get MAF using \pkg{snpStats} package
}
\seealso{
Other LD: 
\code{\link{LD.1KG_download_vcf}()},
\code{\link{LD.1KG}()},
\code{\link{LD.LD_blocks}()},
\code{\link{LD.UKBiobank}()},
\code{\link{LD.calculate_LD}()},
\code{\link{LD.construct_subset_vcf_name}()},
\code{\link{LD.custom_panel}()},
\code{\link{LD.dprime_table}()},
\code{\link{LD.filter_vcf_gaston}()},
\code{\link{LD.filter_vcf}()},
\code{\link{LD.get_locus_vcf_folder}()},
\code{\link{LD.index_vcf}()},
\code{\link{LD.leadSNP_block}()},
\code{\link{LD.load_or_create}()},
\code{\link{LD.plink_LD}()},
\code{\link{LD.plink_file}()},
\code{\link{LD.plot_LD}()},
\code{\link{LD.query_vcf}()},
\code{\link{LD.rds_to_npz}()},
\code{\link{LD.read_bin}()},
\code{\link{LD.read_ld_table}()},
\code{\link{LD.run_plink_LD}()},
\code{\link{LD.save_LD_matrix}()},
\code{\link{LD.snpstats_get_LD}()},
\code{\link{LD.translate_population}()},
\code{\link{LD.vcf_to_bed}()},
\code{\link{saveSparse}()}
}
\concept{LD}
\keyword{internal}
