% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LD.R
\name{LD.translate_population}
\alias{LD.translate_population}
\title{Translate superopulation acronyms}
\usage{
LD.translate_population(superpopulation)
}
\arguments{
\item{superpopulation}{Three-letter superpopulation name.}
}
\description{
Ensures a common ontology for synonynmous superpopulation names.
}
