% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UKBiobank_LD.R
\name{LD.UKBiobank}
\alias{LD.UKBiobank}
\title{Download LD matrices from UK Biobank}
\usage{
LD.UKBiobank(
  subset_DT = NULL,
  locus_dir,
  sumstats_path = NULL,
  chrom = NULL,
  min_pos = NULL,
  force_new_LD = F,
  chimera = F,
  server = T,
  download_full_ld = F,
  download_method = "direct",
  fillNA = 0,
  nThread = 4,
  return_matrix = F,
  conda_env = "echoR",
  remove_tmps = T,
  verbose = T
)
}
\description{
Download pre-computed LD matrices from UK Biobank in 3Mb windows,
then subset to the region that overlaps with \code{subset_DT}.
}
\seealso{
Other LD: 
\code{\link{LD.1KG_download_vcf}()},
\code{\link{LD.1KG}()},
\code{\link{LD.LD_blocks}()},
\code{\link{LD.calculate_LD}()},
\code{\link{LD.construct_subset_vcf_name}()},
\code{\link{LD.custom_panel}()},
\code{\link{LD.dprime_table}()},
\code{\link{LD.filter_vcf_gaston}()},
\code{\link{LD.filter_vcf}()},
\code{\link{LD.get_locus_vcf_folder}()},
\code{\link{LD.index_vcf}()},
\code{\link{LD.leadSNP_block}()},
\code{\link{LD.load_or_create}()},
\code{\link{LD.plink_LD}()},
\code{\link{LD.plink_file}()},
\code{\link{LD.plot_LD}()},
\code{\link{LD.query_vcf}()},
\code{\link{LD.rds_to_npz}()},
\code{\link{LD.read_bin}()},
\code{\link{LD.read_ld_table}()},
\code{\link{LD.run_plink_LD}()},
\code{\link{LD.save_LD_matrix}()},
\code{\link{LD.translate_population}()},
\code{\link{LD.vcf_to_bed}()},
\code{\link{saveSparse}()}
}
\concept{LD}
\keyword{internal}
