% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LD.R
\name{LD.vcf_to_bed}
\alias{LD.vcf_to_bed}
\title{Convert vcf file to BED file}
\usage{
LD.vcf_to_bed(vcf.gz.subset, locus_dir)
}
\arguments{
\item{vcf.gz.subset}{Path to the gzipped locus subset vcf.}

\item{locus_dir}{Locus-specific results directory.}
}
\description{
Uses plink to convert vcf to BED.
}
