% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LD.R
\name{LDlinkR.LDproxy_batch}
\alias{LDlinkR.LDproxy_batch}
\title{Extract LD matrix}
\source{
\href{https://www.rdocumentation.org/packages/LDlinkR/versions/1.0.2}{website}
}
\usage{
LDlinkR.LDproxy_batch(
  snp,
  pop = "CEU",
  r2d = "r2",
  threshold = 0.8,
  verbose = T
)
}
\description{
Wrapper for \code{LDlinkR::LDproxy_batch}
}
\examples{
data("merged_DT")
lead.snps <- setNames(subset(merged_DT, leadSNP)$Locus, subset(merged_DT, leadSNP)$SNP)
proxies <- LDlinkR.LDproxy_batch(snp=lead.snps)
}
