% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{LIFTOVER}
\alias{LIFTOVER}
\title{Lift genome across builds}
\usage{
LIFTOVER(
  dat,
  build.conversion = "hg19.to.hg38",
  chrom_col = "CHR",
  start_col = "POS",
  end_col = "POS",
  verbose = T
)
}
\arguments{
\item{build_conversion}{"hg19.to.hg38" (\emph{default}) or "hg38.to.hg19.}
}
\description{
Lift genome across builds
}
\examples{
data("BST1")
gr.lifted <- LIFTOVER(dat=BST1, build.conversion="hg19.to.hg38")
}
\concept{utils}
