% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{locus_dir}
\alias{locus_dir}
\title{Example results path for BST1 locus}
\format{
An object of class \code{character} of length 1.
}
\usage{
locus_dir
}
\description{
Example results path for BST1 locus
}
\examples{
\dontrun{
locus_dir <- "results/GWAS/Nalls23andMe_2019/BST1"
usethis::use_data(locus_dir, overwrite=T)
}
}
\keyword{datasets}
