% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_dataset_dir}
\alias{make_dataset_dir}
\title{Make paths for results and subsets}
\usage{
make_dataset_dir(
  results_dir = "./results",
  dataset_type = "dataset_type",
  dataset_name = "dataset_name",
  locus
)
}
\description{
Make paths for results and subsets
}
\examples{
dataset_dir <- make_dataset_dir(results_dir="./results", dataset_type="GWAS", dataset_name="Nalls23andMe_2019", locus="BST1")
}
\seealso{
Other directory functions: 
\code{\link{get_locus_dir}()},
\code{\link{get_subset_path}()}
}
\concept{directory functions}
\keyword{internal}
