% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_locus_dir.R
\name{make_locus_dir}
\alias{make_locus_dir}
\title{Make locus-specific results folder}
\usage{
make_locus_dir(
  results_dir,
  dataset_type = "dataset_type",
  dataset_name = "dataset_name",
  locus
)
}
\description{
Make locus-specific results folder
}
\examples{
locus_dir <- echolocatoR:::make_locus_dir(results_dir = file.path(tempdir(),
                                                                  "results"),
                                          dataset_type="GWAS",
                                          dataset_name="Nalls23andMe_2019",
                                          locus="BST1")
}
\seealso{
Other directory functions: 
\code{\link{get_locus_dir}()},
\code{\link{get_study_dir}()},
\code{\link{get_subset_path}()}
}
\concept{directory functions}
\keyword{internal}
