% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_results_path}
\alias{make_results_path}
\title{Make paths for results and subsets}
\usage{
make_results_path(dataset_name, dataset_type, locus)
}
\description{
Make paths for results and subsets
}
\concept{directory functions}
\keyword{internal}
