% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MARGE.R
\name{MARGE.find_motifs}
\alias{MARGE.find_motifs}
\title{Run \emph{HOMER} TFBM enrichment}
\source{
\url{https://github.com/robertamezquita/marge}
}
\usage{
MARGE.find_motifs(finemap_dat, output_dir = "./HOMER_results", genome = "hg19")
}
\description{
Run HOMER transcription factor binding motif enrichment analysis
via the R package \code{marge}
}
\examples{
\dontrun{
# devtools::install_github('robertamezquita/marge', ref = 'master')
## WARNING!: Must install HOMER separately first.
## ml homer R/3.6.0
library(marge)
data("merged_DT")
finemap_dat <- subset(merged_DT, Support>0 | leadSNP)
MARGE.find_motifs()
}
}
\concept{TFBM}
