% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate.R
\name{merge_celltype_specific_epigenomics}
\alias{merge_celltype_specific_epigenomics}
\title{Merge all cell-type-specific epigenomics}
\usage{
merge_celltype_specific_epigenomics(keep_extra_cols = F)
}
\description{
Merges multiple cell-type-specific epigenomic datasets (Nott 2019, Corces 2020) into a single \code{GRAnges} object.
}
\examples{
gr.merged <- merge_celltype_specific_epigenomics()
}
