% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{merge_coloc_results}
\alias{merge_coloc_results}
\title{Convert Jack's coloc results to \emph{echolocatoR} format}
\usage{
merge_coloc_results(
  all_obj,
  results_level = c("summary"),
  nThread = 4,
  verbose = T,
  save_path = F
)
}
\arguments{
\item{all_obj}{Nested list object created by Jack.}

\item{results_level}{Return coloc results at the
"summary" (one row per Locus:eGene pair) or "snp" level (one row per SNP).}

\item{save_path}{File where you want the merged results to be saved.}
}
\description{
Convert Jack's coloc results to \emph{echolocatoR} format
}
\examples{
\dontrun{
load("/sc/hydra/projects/ad-omics/microglia_omics/COLOC/Kunkle_2019/Microglia_all_regions_Kunkle_2019_COLOC.RData")
merged_results <- merge_coloc_results(all_obj=all_obj, results_level="snp", save_path="~/Desktop")
}
}
\seealso{
Other coloc: 
\code{\link{merge_coloc_results_each}()}
}
\concept{coloc}
\keyword{internal}
