% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{merge_coloc_results_each}
\alias{merge_coloc_results_each}
\title{Iterate merging of coloc results across files}
\usage{
merge_coloc_results_each(
  root_dir,
  save_path = F,
  results_level = "summary",
  ppH4_thresh = 0.5,
  filter = F,
  force_new = T,
  nThread = 4,
  verbose = T
)
}
\description{
Iterate merging of coloc results across files
}
\examples{
root_dir <- "/sc/hydra/projects/ad-omics/microglia_omics/COLOC"
# Summary-level
coloc_summary <- merge_coloc_results_each(root_dir=root_dir, save_path="/sc/hydra/projects/pd-omics/brian/all_COLOC_results_merged.summary-level.tsv.gz", results_level="summary")
# SNP-level
coloc_snp <- coloc_summary <- merge_coloc_results_each(root_dir=root_dir, save_path="/sc/hydra/projects/pd-omics/brian/all_COLOC_results_merged.snp-level.tsv.gz", results_level="snp", filter="leadSNP==T")
}
\seealso{
Other coloc: 
\code{\link{merge_coloc_results}()}
}
\concept{coloc}
\keyword{internal}
