% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{merged_DT}
\alias{merged_DT}
\title{\emph{echolocatoR} output example (all loci)}
\format{
data.table
\describe{
  \item{SNP}{SNP RSID}
  \item{CHR}{Chromosome}
  \item{POS}{Genomic positiion (in basepairs)}
  ...
}
}
\source{
\url{https://www.biorxiv.org/content/10.1101/388165v3}
}
\usage{
merged_DT
}
\description{
An example results file after running \code{\link{finemap_loci}}
 on all Parkinson's disease (PD)-associated loci.
}
\details{
Data originally comes from the PD GWAS
by \href{https://www.biorxiv.org/content/10.1101/388165v3}{Nalls et al. (bioRxiv)}.
}
\examples{
\dontrun{
merged_DT <- merge_finemapping_results(dataset = "../Fine_Mapping/Data/GWAS/Nalls23andMe_2019")
}
}
\keyword{datasets}
