% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motifbreakR.R
\name{MOTIFBREAKR.filter_by_metadata}
\alias{MOTIFBREAKR.filter_by_metadata}
\title{Filter by motif database metadata}
\usage{
MOTIFBREAKR.filter_by_metadata(mb.results, Organism = "Hsapiens")
}
\description{
Filter by motif database metadata
}
\seealso{
Other motifbreakR: 
\code{\link{MOTIFBREAKR}()}
}
\concept{motifbreakR}
\keyword{internal}
