% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motifbreakR.R
\name{MOTIFBREAKR.plot}
\alias{MOTIFBREAKR.plot}
\title{Plot \code{\link{motifbreakR}} results}
\source{
\strong{Publication:}
\url{https://pubmed.ncbi.nlm.nih.gov/26272984/}

\strong{GitHub:}
\url{https://github.com/Simon-Coetzee/MotifBreakR}
}
\usage{
MOTIFBREAKR.plot(motifbreakr.results, rsid = NULL)
}
\description{
Plot \code{\link{motifbreakR}} results
}
\examples{
\dontrun{
# motifbreakr.results <- readRDS("/sc/arion/projects/pd-omics/brian/motifbreakR/motifbreakR_results.rds")
motifbreakr.results <- readRDS("~/Desktop/Fine_Mapping/Data/GWAS/Nalls23andMe_2019/_genome_wide/motifbreakR/motifbreakR_results.rds")
MOTIFBREAKR.plot(motifbreakr.results=motifbreakr.results, rsid="rs114528427")
}
}
