% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motifbreakR.R
\name{MOTIFBREAKR}
\alias{MOTIFBREAKR}
\title{Run \code{\link{motifbreakR}}}
\source{
\strong{Publication:}
\url{https://pubmed.ncbi.nlm.nih.gov/26272984/}
\strong{GitHub:}
\url{https://github.com/Simon-Coetzee/MotifBreakR}
\strong{Vignette:}
\url{http://simon-coetzee.github.io/motifBreakR}
}
\usage{
MOTIFBREAKR(snp_list, save_rds = T, dataset_dir)
}
\value{
motifbreakr results
}
\description{
\code{\link{motifbreakR}} is a package to predict how much a SNP will disrupt
a transcription factor binding motif (if it falls witihn one).
}
\examples{
\dontrun{
data("merged_DT")
library('BSgenome')
# merged_DT <- openxlsx::read.xlsx("/sc/arion/projects/pd-omics/brian/Fine_Mapping/Data/Nalls23andMe_2019_results.xlsx")
snp_list <- unique(merged_DT$SNP)
motifbreakr.results <- MOTIFBREAKR(snp_list=snp_list)
}

}
