% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motifbreakR.R
\name{MOTIFBREAKR.summarize}
\alias{MOTIFBREAKR.summarize}
\title{Summarise \code{\link{motifbreakR}} + \code{\link{echolocatoR}} results}
\usage{
MOTIFBREAKR.summarize(merged_DT, mb.results, no_no_loci = NULL)
}
\description{
For each SNP we have at least one allele achieving a p-value below 1e-4 threshold that we required.
The seqMatch column shows what the reference genome sequence is at that location,
with the variant position appearing in an uppercase letter.
pctRef and pctAlt display the the score for the motif in the sequence
as a percentage of the best score that motif could achieve on an ideal sequence.
In other words (scoreVariant−minscorePWM)/(maxscorePWM−minscorePWM).
We can also see the absolute scores for our method in scoreRef and scoreAlt
and thier respective p-values.
}
\examples{
data("merged_DT")
}
