% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MungeSumstats.R
\name{MUNGESUMSTATS.check_syn}
\alias{MUNGESUMSTATS.check_syn}
\title{Check whether a table has headers that can be mapped to an attribute in \pkg{MungeSumstats}}
\usage{
MUNGESUMSTATS.check_syn(dat, col_name = "P-value", col_type = "P")
}
\description{
Check whether a table has headers that can be mapped to an attribute in \pkg{MungeSumstats}
}
\concept{internal}
