% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MungeSumstats.R
\name{MUNGESUMSTATS.col_map}
\alias{MUNGESUMSTATS.col_map}
\title{Column name mappings from \pkg{MungeSumstats} to \pkg{echolocatoR}}
\source{
https://github.com/neurogenomics/MungeSumstats
}
\usage{
MUNGESUMSTATS.col_map()
}
\description{
Column name mappings from \pkg{MungeSumstats} to \pkg{echolocatoR}
}
