% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MungeSumstats.R
\name{MUNGESUMSTATS.to_echolocatoR}
\alias{MUNGESUMSTATS.to_echolocatoR}
\title{Convert from \pkg{MungeSumstats} to \pkg{echolocatoR} format}
\usage{
MUNGESUMSTATS.to_echolocatoR(dat)
}
\description{
Convert from \pkg{MungeSumstats} to \pkg{echolocatoR} format
}
\examples{
\dontrun{
eduAttainOkbayPth <- system.file("extdata","eduAttainOkbay.txt", package="MungeSumstats")
reformatted <- MungeSumstats::format_sumstats(path=eduAttainOkbayPth, ref_genome="GRCh37", return_data = TRUE)
dat_echoR <- MUNGESUMSTATS.to_echolocatoR(dat=reformatted)
}
}
