% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Nalls_top_SNPs}
\alias{Nalls_top_SNPs}
\title{TopSS example file}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 107 rows and 31 columns.
}
\source{
\url{https://www.biorxiv.org/content/10.1101/388165v3}
}
\usage{
Nalls_top_SNPs
}
\description{
Summary stats of the top SNP(s) per locus.
Used to query locus subsets.for fine-mapping.
}
\details{
Data from \href{https://www.biorxiv.org/content/10.1101/388165v3}{Nalls et al. (bioRxiv)}, Table S2.
}
\keyword{datasets}
