% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NOTT_2019.bigwig_metadata}
\alias{NOTT_2019.bigwig_metadata}
\title{Metadata and links to data}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 18 rows and 12 columns.
}
\source{
\url{https://science.sciencemag.org/content/366/6469/1134}
}
\usage{
NOTT_2019.bigwig_metadata
}
\description{
Metadata for cell type-specific epigenomic bigWig files hosted on UCSC Genome Browser.
bigWig files contain the genomic ranges from each epigenomic assay,
as well as a Score column which describes the peaks of the aggregate reads.
}
\seealso{
Other NOTT_2019: 
\code{\link{NOTT_2019.interactome}},
\code{\link{NOTT_2019.superenhancer_interactome}}
}
\concept{NOTT_2019}
\keyword{datasets}
