% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Nott_2019.R
\name{NOTT_2019.epigenomic_histograms}
\alias{NOTT_2019.epigenomic_histograms}
\title{Plot brain cell-specific epigenomic data}
\source{
\href{https://science.sciencemag.org/content/366/6469/1134}{Nott et al. (2019)}
\url{https://genome.ucsc.edu/cgi-bin/hgTracks?db=hg19&lastVirtModeType=default&lastVirtModeExtraState=&virtModeType=default&virtMode=0&nonVirtPosition=&position=chr2:127770344-127983251&hgsid=778249165_ySowqECRKNxURRn6bafH0yewAiuf}
}
\usage{
NOTT_2019.epigenomic_histograms(
  finemap_dat,
  locus_dir,
  show_plot = T,
  save_plot = T,
  full_data = T,
  return_assay_track = F,
  binwidth = 200,
  plot.zoom = "1x",
  geom = "histogram",
  plot_formula = "Assay + Cell_type ~.",
  bigwig_dir = NULL,
  nThread = 4,
  save_annot = F,
  verbose = T
)
}
\arguments{
\item{plot.zoom}{Zoom into the center of the locus when plotting (without editing the fine-mapping results file).
You can provide either:
\itemize{
\item{The size of your plot window in terms of basepairs (e.g. \code{plot.zoom=50000} for a 50kb window)}.
\item{How much you want to zoom in (e.g. \code{plot.zoom="1x"} for the full locus, \code{plot.zoom="2x"} for 2x zoom into the center of the locus, etc.)}.
}
You can pass a list of window sizes (e.g. \code{c(50000,100000,500000)}) to automatically generate
multiple views of each locus.
This can even be a mix of different style inputs: e.g. \code{c("1x","4.5x",25000)}.}

\item{verbose}{Whether \pkg{echolocatoR} should be verbose or silent.}
}
\description{
Brain cell-specific epigenomic data from Nott et al. (2019).
}
\examples{
data("BST1"); data("locus_dir");
 track.Nott_histo <- NOTT_2019.epigenomic_histograms(finemap_dat = BST1, locus_dir = locus_dir, save_plot=F, return_assay_track=T, save_annot=F)
}
\seealso{
Other NOTT_2019: 
\code{\link{NOTT_2019.bigwig_metadata}},
\code{\link{NOTT_2019.get_epigenomic_peaks}()},
\code{\link{NOTT_2019.get_interactions}()},
\code{\link{NOTT_2019.get_interactome}()},
\code{\link{NOTT_2019.get_promoter_celltypes}()},
\code{\link{NOTT_2019.get_promoter_interactome_data}()},
\code{\link{NOTT_2019.get_regulatory_regions}()},
\code{\link{NOTT_2019.interactome}},
\code{\link{NOTT_2019.plac_seq_plot}()},
\code{\link{NOTT_2019.superenhancer_interactome}},
\code{\link{NOTT_2019.superenhancers}()}
}
\concept{NOTT_2019}
\keyword{internal}
