% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Nott_2019.R
\name{NOTT_2019.get_promoter_celltypes}
\alias{NOTT_2019.get_promoter_celltypes}
\title{Get promoter cell types}
\usage{
NOTT_2019.get_promoter_celltypes(annot_sub, marker_key)
}
\description{
Brain cell-specific epigenomic data from Nott et al. (2019).
}
\seealso{
Other NOTT_2019: 
\code{\link{NOTT_2019.bigwig_metadata}},
\code{\link{NOTT_2019.epigenomic_histograms}()},
\code{\link{NOTT_2019.get_epigenomic_peaks}()},
\code{\link{NOTT_2019.get_interactions}()},
\code{\link{NOTT_2019.get_interactome}()},
\code{\link{NOTT_2019.get_promoter_interactome_data}()},
\code{\link{NOTT_2019.get_regulatory_regions}()},
\code{\link{NOTT_2019.interactome}},
\code{\link{NOTT_2019.plac_seq_plot}()},
\code{\link{NOTT_2019.superenhancer_interactome}},
\code{\link{NOTT_2019.superenhancers}()}
}
\concept{NOTT_2019}
\keyword{internal}
