% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NOTT_2019.superenhancer_interactome}
\alias{NOTT_2019.superenhancer_interactome}
\title{Brain cell type-specific interactomes with superenhancers}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 2954 rows and 29 columns.
}
\source{
\url{https://science.sciencemag.org/content/366/6469/1134}
}
\usage{
NOTT_2019.superenhancer_interactome
}
\description{
Originally from \href{https://science.sciencemag.org/content/366/6469/1134}{Nott et al. (2019)}.
Specifically: \emph{aay0793-Nott-Table-S6.xlsx}.
}
\seealso{
Other NOTT_2019: 
\code{\link{NOTT_2019.bigwig_metadata}},
\code{\link{NOTT_2019.interactome}}
}
\concept{NOTT_2019}
\keyword{datasets}
