% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order_loci.R
\name{order_loci}
\alias{order_loci}
\title{Order loci by UCS size, or alphabetically}
\usage{
order_loci(dat, merged_dat, by_UCS_size = F, descending = T, verbose = F)
}
\description{
Order loci by UCS size, or alphabetically
}
\examples{
data("merged_DT");
... by UCS size ...
merged_dat <- order_loci(dat=merged_DT, merged_dat=merged_DT, descending=F)
... alphabetically ...
}
