% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PAINTOR.R
\name{PAINTOR.create_locusFile.QTL}
\alias{PAINTOR.create_locusFile.QTL}
\title{Create QTL locus file for PAINTOR}
\usage{
PAINTOR.create_locusFile.QTL(
  finemap_dat,
  GWAS_datasets,
  QTL_datasets = NULL,
  PT_results_path,
  locus_name,
  metric_suffix = ".t_stat",
  NA_method = c("fill", "drop"),
  force_new_zscore = FALSE
)
}
\description{
Create QTL locus file for PAINTOR
}
\keyword{internal}
