% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PAINTOR.R
\name{PAINTOR.create_locusFile}
\alias{PAINTOR.create_locusFile}
\title{Create PAINTOR locus file}
\usage{
PAINTOR.create_locusFile(
  subset_path,
  fullSS,
  PT_results_path,
  GWAS_datasets = "Nalls23andMe_2019",
  QTL_datasets = c("Fairfax_2014_CD14", "Fairfax_2014_IFN", "Fairfax_2014_LPS2",
    "Fairfax_2014_LPS24"),
  locus,
  locus_name
)
}
\description{
Create PAINTOR locus file
}
\keyword{internal}
