% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PAINTOR.R
\name{PAINTOR.datatype_handler}
\alias{PAINTOR.datatype_handler}
\title{Determine data types (GWAS, QTL)}
\usage{
PAINTOR.datatype_handler(GWAS_datasets = NULL, QTL_datasets = NULL, locus)
}
\description{
Determine data types (GWAS, QTL)
}
\keyword{internal}
