% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PAINTOR.R
\name{PAINTOR.download_annotations}
\alias{PAINTOR.download_annotations}
\title{Download annotations for PAINTOR}
\usage{
PAINTOR.download_annotations(
  PT_results_path,
  locus_name,
  locus_DT,
  XGR_dataset = NA,
  ROADMAP_search = NA,
  chromatin_state = "TssA",
  no_annotations = FALSE
)
}
\description{
Download annotations for PAINTOR
}
\keyword{internal}
