% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PAINTOR.R
\name{PAINTOR.import_QTL_DT}
\alias{PAINTOR.import_QTL_DT}
\title{Import QTL data for PAINTOR}
\usage{
PAINTOR.import_QTL_DT(
  QTL_datasets,
  locus,
  trim_gene_limits = FALSE,
  force_new_subset = FALSE,
  metric = "t_stat"
)
}
\description{
@keywords internal
}
