% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PAINTOR.R
\name{PAINTOR.list_paintor_annotations}
\alias{PAINTOR.list_paintor_annotations}
\title{List available PAINTOR annotations}
\usage{
PAINTOR.list_paintor_annotations(
  annotations_dir = file.path("/sc/orga/projects/pd-omics/brian/PAINTOR_V3.0",
    "Annotation_directory/Functional_Annotations")
)
}
\description{
List available PAINTOR annotations
}
\keyword{internal}
