% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PAINTOR.R
\name{PAINTOR.locusName_handler}
\alias{PAINTOR.locusName_handler}
\title{Construct locus name for PAINTOR}
\usage{
PAINTOR.locusName_handler(
  locus_name = NULL,
  locus,
  GWAS_datasets = NULL,
  QTL_datasets = NULL
)
}
\description{
Construct locus name for PAINTOR
}
\keyword{internal}
