% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PAINTOR.R
\name{PAINTOR.merge_results}
\alias{PAINTOR.merge_results}
\title{Merge PAINTOR results}
\usage{
PAINTOR.merge_results(
  finemap_dat,
  paintor.results,
  PP_threshold = 0.5,
  multi_finemap_col_name = "PAINTOR"
)
}
\description{
@keywords internal
}
