% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PAINTOR.R
\name{PAINTOR.prepare_annotations}
\alias{PAINTOR.prepare_annotations}
\title{Prepare PAINTOR annotations}
\usage{
PAINTOR.prepare_annotations(
  paintor_path = "./echolocatoR/tools/PAINTOR_V3.0",
  BED_paths,
  PT_results_path,
  locus_name,
  remove_BED = F
)
}
\description{
Prepare PAINTOR annotations
}
\keyword{internal}
