% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PAINTOR.R
\name{PAINTOR.prepare_LD}
\alias{PAINTOR.prepare_LD}
\title{Prepare LD file for PAINTOR}
\usage{
PAINTOR.prepare_LD(
  subset_path,
  PT_results_path,
  locus_name,
  locus_DT,
  locus,
  LD_matrix = NULL
)
}
\description{
Prepare LD file for PAINTOR
}
\keyword{internal}
