% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PAINTOR.R
\name{PAINTOR}
\alias{PAINTOR}
\title{Run full PAINTOR pipeline}
\source{
\href{https://github.com/gkichaev/PAINTOR_V3.0}{GitHub}
\href{https://github.com/gkichaev/PAINTOR_V3.0/wiki/2a.-Computing-1000-genomes-LD}{LD Tutorial}
}
\usage{
PAINTOR(
  finemap_dat = NULL,
  paintor_path = NULL,
  GWAS_datasets = NULL,
  QTL_datasets = NULL,
  locus,
  locus_name = NULL,
  n_causal = 5,
  XGR_dataset = NA,
  ROADMAP_search = NA,
  chromatin_state = "TssA",
  use_annotations = FALSE,
  PP_threshold = 0.95,
  consensus_thresh = 2,
  multi_finemap_col_name = "PAINTOR",
  trim_gene_limits = FALSE,
  force_new_subset = FALSE,
  GWAS_populations = "EUR",
  QTL_populations = "EUR",
  LD_reference = "1KG_Phase1",
  force_new_LD = FALSE,
  LD_matrix = NULL,
  force_reinstall = FALSE
)
}
\description{
Run full PAINTOR pipeline
}
\keyword{internal}
