% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PAINTOR.R
\name{PAINTOR.run}
\alias{PAINTOR.run}
\title{Run PAINTOR}
\usage{
PAINTOR.run(
  paintor_path = NULL,
  PT_results_path,
  .LD_file.paths,
  n_datasets,
  method = "enumerate",
  n_causal = 2
)
}
\description{
@keywords internal
}
