% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PAINTOR.R
\name{PAINTOR.survey_annotation}
\alias{PAINTOR.survey_annotation}
\title{Report the number of annotations > 0}
\usage{
PAINTOR.survey_annotation(PT_results_path, locus_name = "Locus1")
}
\description{
@keywords internal
}
