% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLOT.locus.R
\name{PLOT.add_multitrack_lines}
\alias{PLOT.add_multitrack_lines}
\title{Add vertical lines}
\usage{
PLOT.add_multitrack_lines(
  TRKS,
  finemap_dat,
  snp_groups = c("Lead", "UCS", "Consensus"),
  line_alpha = 1,
  line_size = 0.3,
  remove_duplicated_UCS_Consensus = T,
  verbose = F
)
}
\description{
Adds vertical lines indicates key SNPs
}
\seealso{
Other plot: 
\code{\link{PLOT.dot_summary}()},
\code{\link{PLOT.get_window_limits}()},
\code{\link{PLOT.get_window_suffix}()}
}
\concept{plot}
\keyword{internal}
