% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLOT.locus.R
\name{PLOT.dot_summary}
\alias{PLOT.dot_summary}
\title{Multi-fine-map summary dot plot}
\usage{
PLOT.dot_summary(finemap_dat, PP_threshold = 0.95, show_plot = T)
}
\description{
Multi-fine-map summary dot plot
}
\examples{
data("BST1")
gp <- PLOT.dot_summary(finemap_dat=BST1)
}
\seealso{
Other plot: 
\code{\link{PLOT.add_multitrack_lines}()},
\code{\link{PLOT.get_window_limits}()},
\code{\link{PLOT.get_window_suffix}()}
}
\concept{plot}
