% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLOT.locus.R
\name{PLOT.get_window_limits}
\alias{PLOT.get_window_limits}
\title{Get window size limits for plot}
\usage{
PLOT.get_window_limits(
  finemap_dat,
  index_as_center = T,
  plot.zoom = NULL,
  genomic_units = "Mb",
  verbose = T
)
}
\description{
Get window size limits for plot
}
\examples{
data("BST1");
xlims <- PLOT.get_window_limits(finemap_dat=BST1, plot.zoom=50000)
xlims <- PLOT.get_window_limits(finemap_dat=BST1, plot.zoom="all")
xlims <- PLOT.get_window_limits(finemap_dat=BST1, plot.zoom="5x")
}
\seealso{
Other plot: 
\code{\link{PLOT.add_multitrack_lines}()},
\code{\link{PLOT.dot_summary}()},
\code{\link{PLOT.get_window_suffix}()}
}
\concept{plot}
\keyword{internal}
