% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLOT.locus.R
\name{PLOT.get_window_suffix}
\alias{PLOT.get_window_suffix}
\title{Determine the plot suffix indicating its window size}
\usage{
PLOT.get_window_suffix(finemap_dat, plot.zoom, verbose = T)
}
\description{
Determine the plot suffix indicating its window size
}
\examples{
data("BST1")
window_suffix <- PLOT.get_window_suffix(finemap_dat=BST1, plot.zoom=1000)
window_suffix <- PLOT.get_window_suffix(finemap_dat=BST1, plot.zoom=NULL)
window_suffix <- PLOT.get_window_suffix(finemap_dat=BST1, plot.zoom="all")
window_suffix <- PLOT.get_window_suffix(finemap_dat=BST1, plot.zoom="2x")
}
\seealso{
Other plot: 
\code{\link{PLOT.add_multitrack_lines}()},
\code{\link{PLOT.dot_summary}()},
\code{\link{PLOT.get_window_limits}()}
}
\concept{plot}
\keyword{internal}
