% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggbio.R
\name{PLOT.multitrack}
\alias{PLOT.multitrack}
\title{Remake of multitrack plot}
\usage{
PLOT.multitrack(
  finemap_dat,
  locus_dir,
  LD_matrix = NULL,
  LD_reference = NULL,
  dataset_type = "GWAS",
  color_r2 = T,
  method_list = c("ABF", "FINEMAP", "SUSIE", "POLYFUN_SUSIE"),
  dot_summary = F,
  QTL_prefixes = NULL,
  mean.PP = T,
  PP_threshold = 0.95,
  consensus_threshold = 2,
  sig_cutoff = 5e-08,
  gene_track = T,
  point_size = 1,
  point_alpha = 0.6,
  snp_group_lines = c("Lead", "UCS", "Consensus"),
  XGR_libnames = c("ENCODE_TFBS_ClusteredV3_CellTypes",
    "ENCODE_DNaseI_ClusteredV3_CellTypes", "Broad_Histone"),
  n_top_xgr = 5,
  Roadmap = F,
  Roadmap_query = NULL,
  n_top_roadmap = 7,
  annot_overlap_threshold = 5,
  Nott_epigenome = F,
  Nott_regulatory_rects = T,
  Nott_show_placseq = T,
  Nott_binwidth = 200,
  Nott_bigwig_dir = NULL,
  save_plot = T,
  show_plot = T,
  max_transcripts = 1,
  plot.zoom = "1x",
  dpi = 300,
  height = 12,
  width = 10,
  nThread = 4,
  verbose = T
)
}
\description{
0. Remove labels from zoomed out GWAS
1. Remove tool PP and use labels/symbols in mean PP
2. Add multiple transcripts
3. Zoom out/reduce alpha of PLAC-seq
}
\examples{
data("LRRK2")

locus_dir <- "/pd-omics/brian/Fine_Mapping/Data/GWAS/Nalls23andMe_2019/LRRK2"
finemap_dat <- LRRK2
LD_mat <- data.table::fread(file.path(locus_dir,"LD/LRRK2.leadSNP_LD.csv.gz"))

}
