% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggbio.R
\name{PLOT.transcript_model_track}
\alias{PLOT.transcript_model_track}
\title{Plot gene/transcript models}
\source{
\href{https://bioconductor.org/packages/release/bioc/vignettes/ensembldb/inst/doc/ensembldb.html}{ensembld tutorial}
\href{https://bioconductor.org/packages/devel/bioc/vignettes/Gviz/inst/doc/Gviz.html#45_GeneRegionTrack}{Gvix tutorial}
\href{http://bioconductor.org/packages/devel/bioc/vignettes/ggbio/inst/doc/ggbio.pdf}{ggbio tutorial}
}
\usage{
PLOT.transcript_model_track(
  finemap_dat,
  max_transcripts = 1,
  remove_RP11 = T,
  show.legend = T,
  show_plot = F,
  fill = "skyblue",
  shape = c("arrow", "box", "ellipse", "smallArrow"),
  transcriptAnnotation = c("symbol", "transcript"),
  collapseTranscripts = c(F, T, "longest"),
  stacking = c("squish", "hide", "dense", "pack", "full"),
  method = "ggbio",
  verbose = T
)
}
\description{
Plot gene/transcript models
}
