% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/POLYFUN.R
\name{POLYFUN.h2_enrichment}
\alias{POLYFUN.h2_enrichment}
\title{Run heritability enrichment tests}
\source{
https://www.biorxiv.org/content/10.1101/807792v3
}
\usage{
POLYFUN.h2_enrichment(h2_df, target_SNPs = NULL, fillNA = TRUE)
}
\description{
Run heritability enrichment tests
}
\seealso{
Other polyfun: 
\code{\link{POLYFUN.compute_priors}()},
\code{\link{POLYFUN.download_ref_files}()},
\code{\link{POLYFUN.find_polyfun_folder}()},
\code{\link{POLYFUN.finemapper}()},
\code{\link{POLYFUN.functional_enrichment}()},
\code{\link{POLYFUN.gather_annotations}()},
\code{\link{POLYFUN.gather_ldscores}()},
\code{\link{POLYFUN.get_precomputed_priors}()},
\code{\link{POLYFUN.h2_enrichment_SNPgroups_plot}()},
\code{\link{POLYFUN.h2_enrichment_SNPgroups}()},
\code{\link{POLYFUN.help}()},
\code{\link{POLYFUN.initialize}()},
\code{\link{POLYFUN.ldsc_annot_enrichment}()},
\code{\link{POLYFUN.munge_summ_stats}()},
\code{\link{POLYFUN.plot}()},
\code{\link{POLYFUN.prepare_snp_input}()},
\code{\link{POLYFUN.read_parquet}()},
\code{\link{POLYFUN.run_ldsc}()},
\code{\link{POLYFUN_SUSIE}()}
}
\concept{polyfun}
\keyword{internal}
