% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/POLYFUN.R
\name{POLYFUN.h2_enrichment_SNPgroups}
\alias{POLYFUN.h2_enrichment_SNPgroups}
\title{Run heritability enrichment tests across SNP groups}
\source{
https://www.biorxiv.org/content/10.1101/807792v3
}
\usage{
POLYFUN.h2_enrichment_SNPgroups(
  merged_dat,
  chrom = "*",
  ldsc_suffix = "*.snpvar_constrained.gz",
  subtitle = "",
  show_plot = T,
  save_plot = "./h2_enrichment.png",
 
    out.path = "/sc/arion/projects/pd-omics/brian/Fine_Mapping/Data/GWAS/Nalls23andMe_2019/_genome_wide/PolyFun/output",
  nThread = 4
)
}
\description{
Run heritability enrichment tests across SNP groups
}
\examples{
# IMPORTANT! For this to make sense, you need to merge the full data ("merged_DT" only includes Support>0 and leadSNPs)
merged_dat_full <- merge_finemapping_results(dataset = "/sc/arion/projects/pd-omics/brian/Fine_Mapping/Data/GWAS/Nalls23andMe_2019", minimum_support = 0)
res <- POLYFUN.h2_enrichment_SNPgroups(merged_dat=merged_DT, out.path="/sc/arion/projects/pd-omics/brian/Fine_Mapping/Data/GWAS/Nalls23andMe_2019/_genome_wide/PolyFun/output")
}
\seealso{
Other polyfun: 
\code{\link{POLYFUN.compute_priors}()},
\code{\link{POLYFUN.download_ref_files}()},
\code{\link{POLYFUN.find_polyfun_folder}()},
\code{\link{POLYFUN.finemapper}()},
\code{\link{POLYFUN.functional_enrichment}()},
\code{\link{POLYFUN.gather_annotations}()},
\code{\link{POLYFUN.gather_ldscores}()},
\code{\link{POLYFUN.get_precomputed_priors}()},
\code{\link{POLYFUN.h2_enrichment}()},
\code{\link{POLYFUN.help}()},
\code{\link{POLYFUN.initialize}()},
\code{\link{POLYFUN.ldsc_annot_enrichment}()},
\code{\link{POLYFUN.munge_summ_stats}()},
\code{\link{POLYFUN.plot}()},
\code{\link{POLYFUN.prepare_snp_input}()},
\code{\link{POLYFUN.read_parquet}()},
\code{\link{POLYFUN.run_ldsc}()},
\code{\link{POLYFUN_SUSIE}()}
}
\concept{polyfun}
\keyword{internal}
