% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/POLYFUN.R
\name{POLYFUN.ldsc_annot_enrichment}
\alias{POLYFUN.ldsc_annot_enrichment}
\title{Run and plot heritability enrichment tests}
\source{
https://www.biorxiv.org/content/10.1101/807792v3
}
\usage{
POLYFUN.ldsc_annot_enrichment(
 
    .results = "Data/GWAS/Nalls23andMe_2019/_genome_wide/PolyFun/output/PD_GWAS_LDSC/PD_GWAS_LDSC.results",
  show_plot = TRUE,
  save_plot = FALSE,
  title = "LDSC Heritability Enrichment",
  subtitle = "PD GWAS"
)
}
\description{
Run and plot heritability enrichment tests
}
\seealso{
Other polyfun: 
\code{\link{POLYFUN.compute_priors}()},
\code{\link{POLYFUN.download_ref_files}()},
\code{\link{POLYFUN.find_polyfun_folder}()},
\code{\link{POLYFUN.finemapper}()},
\code{\link{POLYFUN.functional_enrichment}()},
\code{\link{POLYFUN.gather_annotations}()},
\code{\link{POLYFUN.gather_ldscores}()},
\code{\link{POLYFUN.get_precomputed_priors}()},
\code{\link{POLYFUN.h2_enrichment_SNPgroups_plot}()},
\code{\link{POLYFUN.h2_enrichment_SNPgroups}()},
\code{\link{POLYFUN.h2_enrichment}()},
\code{\link{POLYFUN.help}()},
\code{\link{POLYFUN.initialize}()},
\code{\link{POLYFUN.munge_summ_stats}()},
\code{\link{POLYFUN.plot}()},
\code{\link{POLYFUN.prepare_snp_input}()},
\code{\link{POLYFUN.read_parquet}()},
\code{\link{POLYFUN.run_ldsc}()},
\code{\link{POLYFUN_SUSIE}()}
}
\concept{polyfun}
\keyword{internal}
