% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/POLYFUN.R
\name{POLYFUN.munge_summ_stats}
\alias{POLYFUN.munge_summ_stats}
\title{Munge summary stats}
\usage{
POLYFUN.munge_summ_stats(
  polyfun = NULL,
  fullSS_path,
  locus_dir,
  sample_size = 1474097,
  min_INFO = 0,
  min_MAF = 0.001,
  force_new_munge = F
)
}
\description{
Munge summary stats
}
\examples{
\dontrun{
data("genome_wide_dir");
fullSS_path <- "/Users/schilder/Desktop/echolocatoR/results/GWAS/Nalls23andMe_2019/_genome_wide/nallsEtAl2019_allSamples_allVariants.mod.txt.gz"
munged_path <- POLYFUN.munge_summ_stats(fullSS_path=fullSS_path, locus_dir=locus_dir_genome_wide)
}
}
\seealso{
Other polyfun: 
\code{\link{POLYFUN.compute_priors}()},
\code{\link{POLYFUN.conda_from_list}()},
\code{\link{POLYFUN.conda_from_yaml}()},
\code{\link{POLYFUN.download_ref_files}()},
\code{\link{POLYFUN.find_polyfun_folder}()},
\code{\link{POLYFUN.finemapper}()},
\code{\link{POLYFUN.functional_enrichment}()},
\code{\link{POLYFUN.gather_annotations}()},
\code{\link{POLYFUN.gather_ldscores}()},
\code{\link{POLYFUN.get_precomputed_priors}()},
\code{\link{POLYFUN.h2_enrichment_SNPgroups}()},
\code{\link{POLYFUN.h2_enrichment}()},
\code{\link{POLYFUN.help}()},
\code{\link{POLYFUN.initialize}()},
\code{\link{POLYFUN.install_conda}()},
\code{\link{POLYFUN.install_dependencies}()},
\code{\link{POLYFUN.ldsc_annot_enrichment}()},
\code{\link{POLYFUN.load_conda}()},
\code{\link{POLYFUN.plot}()},
\code{\link{POLYFUN.prepare_snp_input}()},
\code{\link{POLYFUN.read_parquet}()},
\code{\link{POLYFUN.run_ldsc}()},
\code{\link{POLYFUN_SUSIE}()}
}
\concept{polyfun}
\keyword{internal}
