% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/POLYFUN.R
\name{POLYFUN.prepare_snp_input}
\alias{POLYFUN.prepare_snp_input}
\title{Prepare SNP input for PolyFun}
\usage{
POLYFUN.prepare_snp_input(
  PF.output.path,
  locus_dir,
  finemap_dat = NULL,
  nThread = 1
)
}
\description{
PolyFun requires a space-delimited (gzipped or not) file with these columns:
\itemize{
\item{CHR}
\item{BP}
\item{A1}
\item{A2}
}
}
\examples{
BST1 <- echodata::BST1; locus_dir <- echodata::locus_dir;
finemap_dat <- BST1
PF.output.path <- file.path(locus_dir, "PolyFun")
POLYFUN.prepare_snp_input(PF.output.path=PF.output.path, locus_dir=locus_dir, finemap_dat=finemap_dat)
}
\seealso{
Other polyfun: 
\code{\link{POLYFUN.compute_priors}()},
\code{\link{POLYFUN.download_ref_files}()},
\code{\link{POLYFUN.find_polyfun_folder}()},
\code{\link{POLYFUN.finemapper}()},
\code{\link{POLYFUN.functional_enrichment}()},
\code{\link{POLYFUN.gather_annotations}()},
\code{\link{POLYFUN.gather_ldscores}()},
\code{\link{POLYFUN.get_precomputed_priors}()},
\code{\link{POLYFUN.h2_enrichment_SNPgroups_plot}()},
\code{\link{POLYFUN.h2_enrichment_SNPgroups}()},
\code{\link{POLYFUN.h2_enrichment}()},
\code{\link{POLYFUN.help}()},
\code{\link{POLYFUN.initialize}()},
\code{\link{POLYFUN.ldsc_annot_enrichment}()},
\code{\link{POLYFUN.munge_summ_stats}()},
\code{\link{POLYFUN.plot}()},
\code{\link{POLYFUN.read_parquet}()},
\code{\link{POLYFUN.run_ldsc}()},
\code{\link{POLYFUN_SUSIE}()}
}
\concept{polyfun}
\keyword{internal}
