% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/POLYFUN.R
\name{POLYFUN.read_parquet}
\alias{POLYFUN.read_parquet}
\title{Read parquet file as data.frame}
\usage{
POLYFUN.read_parquet(parquet_path, conda_env = "echoR", method = "pandas")
}
\arguments{
\item{method}{Specify whether to use \code{\link{SparkR}} (R)
or \code{pandas} (Python + \code{reticulate}).}
}
\description{
Read parquet file as data.frame
}
\examples{
\dontrun{
parquet_path <- system.file("tools/polyfun/example_data/weights.10.l2.ldscore.parquet", package = "echolocatoR")
# Using python (pandas) - default
dat <- POLYFUN.read_parquet(parquet_path=parquet_path, method="pandas")
# Using R (SparkR)
dat <- POLYFUN.read_parquet(parquet_path=parquet_path, method="sparkR")
}
}
\seealso{
Other polyfun: 
\code{\link{POLYFUN.compute_priors}()},
\code{\link{POLYFUN.download_ref_files}()},
\code{\link{POLYFUN.find_polyfun_folder}()},
\code{\link{POLYFUN.finemapper}()},
\code{\link{POLYFUN.functional_enrichment}()},
\code{\link{POLYFUN.gather_annotations}()},
\code{\link{POLYFUN.gather_ldscores}()},
\code{\link{POLYFUN.get_precomputed_priors}()},
\code{\link{POLYFUN.h2_enrichment_SNPgroups_plot}()},
\code{\link{POLYFUN.h2_enrichment_SNPgroups}()},
\code{\link{POLYFUN.h2_enrichment}()},
\code{\link{POLYFUN.help}()},
\code{\link{POLYFUN.initialize}()},
\code{\link{POLYFUN.ldsc_annot_enrichment}()},
\code{\link{POLYFUN.munge_summ_stats}()},
\code{\link{POLYFUN.plot}()},
\code{\link{POLYFUN.prepare_snp_input}()},
\code{\link{POLYFUN.run_ldsc}()},
\code{\link{POLYFUN_SUSIE}()}
}
\concept{polyfun}
\keyword{internal}
