% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/POLYFUN.R
\name{POLYFUN_SUSIE}
\alias{POLYFUN_SUSIE}
\title{Run PolyFun+SUSIE fine-mapping pipeline}
\source{
https://www.biorxiv.org/content/10.1101/807792v3
}
\usage{
POLYFUN_SUSIE(
  locus_dir,
  polyfun = NULL,
  finemap_dat = NULL,
  LD_matrix = NULL,
  polyfun_approach = "non-parametric",
  dataset_type = "GWAS",
  max_causal = 5,
  sample_size = NULL,
  server = FALSE,
  PP_threshold = 0.95,
  conda_env = "echoR"
)
}
\description{
Uses echolocatoR wrapper for SUSIE instead of the \code{\link{POLYFUN.finemapper}}
function which uses a python script provided with PolyFun.
}
\seealso{
Other polyfun: 
\code{\link{POLYFUN.compute_priors}()},
\code{\link{POLYFUN.download_ref_files}()},
\code{\link{POLYFUN.find_polyfun_folder}()},
\code{\link{POLYFUN.finemapper}()},
\code{\link{POLYFUN.functional_enrichment}()},
\code{\link{POLYFUN.gather_annotations}()},
\code{\link{POLYFUN.gather_ldscores}()},
\code{\link{POLYFUN.get_precomputed_priors}()},
\code{\link{POLYFUN.h2_enrichment_SNPgroups_plot}()},
\code{\link{POLYFUN.h2_enrichment_SNPgroups}()},
\code{\link{POLYFUN.h2_enrichment}()},
\code{\link{POLYFUN.help}()},
\code{\link{POLYFUN.initialize}()},
\code{\link{POLYFUN.ldsc_annot_enrichment}()},
\code{\link{POLYFUN.munge_summ_stats}()},
\code{\link{POLYFUN.plot}()},
\code{\link{POLYFUN.prepare_snp_input}()},
\code{\link{POLYFUN.read_parquet}()},
\code{\link{POLYFUN.run_ldsc}()}
}
\concept{polyfun}
\keyword{internal}
