% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IMPACT.R
\name{prepare_mat_meta}
\alias{prepare_mat_meta}
\title{Prepare IMPACT data for for ComplexHeatmap}
\usage{
prepare_mat_meta(
  TOP_IMPACT,
  TOP_IMPACT_all,
  snp.group = "Consensus",
  value.var = "mean_IMPACT",
  fill_na = 0
)
}
\description{
Prepare IMPACT data for for ComplexHeatmap
}
\keyword{internal}
