% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{printer}
\alias{printer}
\title{printer}
\usage{
printer(..., v = T)
}
\value{
character
}
\description{
Concatenate and print any number of items.
}
\examples{
n.snps <- 50
printer("echolocatoR::","Processing",n.snps,"SNPs...")
}
