% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COJO.R
\name{process_COJO_results}
\alias{process_COJO_results}
\title{Gather all GCTA-COJO results}
\usage{
process_COJO_results(subset_DT, locus_dir, freq_cutoff = 0.1)
}
\description{
Gather all GCTA-COJO results
}
\seealso{
Other COJO
\url{https://www.nature.com/articles/ng.2213}
\url{https://www.cell.com/ajhg/fulltext/S0002-9297(10)00598-7}
\url{https://cnsgenomics.com/software/gcta/#Overview}: 
\code{\link{COJO.stepwise}()},
\code{\link{COJO}()},
\code{\link{get_stepwise_results}()}
}
\concept{COJO
\url{https://www.nature.com/articles/ng.2213}
\url{https://www.cell.com/ajhg/fulltext/S0002-9297(10)00598-7}
\url{https://cnsgenomics.com/software/gcta/#Overview}}
