% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{query_by_coordinates_merged}
\alias{query_by_coordinates_merged}
\title{Use \emph{awk} to query locus subsets.}
\usage{
query_by_coordinates_merged(
  top_SNPs,
  fullSS_path,
  subset_path,
  locus,
  chrom_col,
  file_sep = " ",
  location_sep = ":",
  min_POS,
  max_POS,
  bp_distance
)
}
\arguments{
\item{top_SNPs}{A data.frame with the genomic coordinates of the lead SNP
for each locus.
The lead SNP will be used as the center of the window when extracting
subset from the full GWAS/QTL summary statistics file.
Only one SNP per \strong{Locus} should be included.
At minimum, \code{top_SNPs} should include the following columns:
\describe{
\item{\emph{Locus}}{A unique name for each locus. Often,
 loci are named after a relevant gene (e.g. LRRK2) or based on
  the name/coordinates of the lead SNP (e.g. locus_chr12_40734202) }
\item{\emph{CHR}}{The chromosome that the SNP is on.
 Can be "chr12" or "12" format.}
\item{\emph{POS}}{The genomic position of the SNP (in basepairs)}
}}

\item{fullSS_path}{Path to the full summary statistics file (GWAS or QTL)
that you want to fine-map.
It is usually best to provide the absolute path rather
than the relative path.}

\item{chrom_col}{Name of the chromosome column in the full summary stats
file.
Can be "chr1" or "1" format.
(\emph{default: ="CHR"})}

\item{location_sep}{The separator character when \strong{CHR} and \strong{POS} are merged into one column (e.g. ":" when formatted like chr12:12209944).}

\item{min_POS}{Minimum genomic position to include.}

\item{max_POS}{Maximum genomic position to include.}

\item{bp_distance}{Distance around the lead SNP to include.}
}
\description{
Search full summary stats file by genomics coordinates.
To be used when \strong{CHR} and \strong{POS} are merged into one column (e.g. chr12:12209944).
}
\seealso{
Other query functions: 
\code{\link{extract_snp_subset}()},
\code{\link{query_by_coordinates}()},
\code{\link{query_by_gene}()},
\code{\link{query_by_probe}()},
\code{\link{query_fullSS}()},
\code{\link{query_handler}()}
}
\concept{query functions}
\keyword{internal}
