% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{query_by_probe}
\alias{query_by_probe}
\title{Use \emph{awk} to query locus subsets.}
\usage{
query_by_probe(
  fullSS_path,
  subset_path,
  gene,
  gene_col,
  chrom_col,
  file_sep,
  probe_path,
  coordinates_merged = T,
  location_sep = ":"
)
}
\arguments{
\item{coordinates_merged}{Whether \strong{CHR} and \strong{POS} are merged into one column (e.g. chr12:12209944).}

\item{location_sep}{The separator character when \strong{CHR} and \strong{POS} are merged into one column (e.g. ":" when formatted like chr12:12209944).}
}
\description{
To be used when probe name (but not gene name) is a column in the full summary stats file.
Requires a gene-probe key file (\strong{probe_path}).
More commonly useful for QTL full summary stats files.
}
