% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{rbind.file.list}
\alias{rbind.file.list}
\title{Bind stored files}
\usage{
\method{rbind}{file.list}(file.list, verbose = T, nCores = 4)
}
\description{
Rapidly read a list of files from storage and concatenate them by rows.
}
\examples{
file.list <- c("data/file1.tsv", "data/file2.tsv","new_data/file3/tsv")
dat <- rbind.file.list(file.list = file.list)
}
