% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbind_GRanges.R
\name{rbind_GRanges}
\alias{rbind_GRanges}
\title{Bind GRanges with different mcols}
\usage{
rbind_GRanges(gr1, gr2)
}
\description{
Bind GRanges with different mcols
}
\examples{
data("merged_DT")
gr.hits <- CORCES_2020.get_ATAC_peak_overlap(finemap_dat = merged_DT)
gr.hits$extra_col <- "Extra"
gr.anchor_hits <- CORCES_2020.get_HiChIP_FitHiChIP_overlap(finemap_dat = merged_DT)
try({gr.bind <- c(gr.hits, gr.anchor_hits)})
gr.bound <- rbind_GRanges(gr1, gr2)
}
\seealso{
Other utils: 
\code{\link{LIFTOVER}()},
\code{\link{dataframe_2_vcf}()}
}
\concept{utils}
\keyword{internal}
