% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise.R
\name{results_report}
\alias{results_report}
\title{Give a quick summary report of the fine-mapping results}
\usage{
results_report(merged_dat)
}
\description{
Give a quick summary report of the fine-mapping results
}
\examples{
data("merged_DT");
results_report(merged_DT)
}
\seealso{
Other summarise: 
\code{\link{SUMMARISE.CS_bin_plot}()},
\code{\link{SUMMARISE.CS_counts_plot}()},
\code{\link{SUMMARISE.get_CS_bins}()},
\code{\link{SUMMARISE.get_CS_counts}()},
\code{\link{SUMMARISE.get_SNPgroup_counts}()},
\code{\link{SUMMARISE.peak_overlap_plot}()}
}
\concept{summarise}
