% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Roadmap.R
\name{ROADMAP.merge_and_process_grl}
\alias{ROADMAP.merge_and_process_grl}
\title{Standardize Roadmap query}
\usage{
ROADMAP.merge_and_process_grl(
  grl.roadmap,
  gr.snp,
  n_top_tissues = 5,
  sep = " "
)
}
\arguments{
\item{grl.roadmap}{Roadmap query results}

\item{n_top_tissues}{The number of top tissues to include,
sorted by greatest number of rows
(i.e. the number of genomic ranges within the window).}
}
\description{
Standardize Roadmap query
}
\seealso{
Other ROADMAP: 
\code{\link{ROADMAP.construct_reference}()},
\code{\link{ROADMAP.query_and_plot}()},
\code{\link{ROADMAP.query}()},
\code{\link{ROADMAP.tabix}()}
}
\concept{ROADMAP}
