% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Roadmap.R
\name{ROADMAP.query}
\alias{ROADMAP.query}
\title{Query Roadmap by genomic coordinates}
\usage{
ROADMAP.query(results_path, gr.snp, keyword_query = NULL, limit_files = NULL)
}
\arguments{
\item{results_path}{Where to store query results.}

\item{gr.snp}{\code{\link[GenomicRanges]{GRanges}} object of SNPs to query Roadmap with.}

\item{limit_files}{Limit the number of annotation files queried (for faster testing).}
}
\description{
Query Roadmap by genomic coordinates
}
\examples{
data("BST1")
finemap_DT <- BST1
gr.snp <- DT_to_GRanges(subset_DT = finemap_DT)
grl.roadmap <- ROADMAP.query(results_path="./Roadmap", gr.snp=gr.snp, keyword_query="placenta")
}
\seealso{
Other ROADMAP: 
\code{\link{ROADMAP.construct_reference}()},
\code{\link{ROADMAP.merge_and_process_grl}()},
\code{\link{ROADMAP.query_and_plot}()},
\code{\link{ROADMAP.tabix}()}
}
\concept{ROADMAP}
