% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Roadmap.R
\name{ROADMAP.tabix}
\alias{ROADMAP.tabix}
\title{Query Roadmap API}
\usage{
ROADMAP.tabix(
  results_path,
  chrom,
  min_pos,
  max_pos,
  eid,
  convert_to_GRanges = T,
  conda_env = "echoR"
)
}
\arguments{
\item{results_path}{Where to store query results.}

\item{chrom}{Chromosome to query}

\item{min_pos}{Minimum genomic position}

\item{max_pos}{Maximum genomic position}

\item{eid}{Roadmap annotation ID}

\item{convert_to_GRanges}{Whether to return query
as a \code{data.frame} or \code{\link[GenomicRanges]{GRanges}}.}
}
\description{
Query Roadmap epigenomic annotations (chromatin marks)
using a range of genomic coordinates.
}
\seealso{
Other ROADMAP: 
\code{\link{ROADMAP.construct_reference}()},
\code{\link{ROADMAP.merge_and_process_grl}()},
\code{\link{ROADMAP.query_and_plot}()},
\code{\link{ROADMAP.query}()}
}
\concept{ROADMAP}
