% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpliceAI.R
\name{SPLICEAI.plot}
\alias{SPLICEAI.plot}
\title{Plot \emph{SpliceAI} predictions}
\source{
\href{https://github.com/Illumina/SpliceAI}{GitHub}
\href{https://www.sciencedirect.com/science/article/pii/S0092867418316295}{Publication}
}
\usage{
SPLICEAI.plot(dat_merged)
}
\description{
Plot \emph{SpliceAI} predictions
}
\seealso{
Other SpliceAI: 
\code{\link{SPLICEAI.run}()},
\code{\link{SPLICEAI.snp_probs}()},
\code{\link{SPLICEAI.subset_precomputed_tsv_iterate}()},
\code{\link{SPLICEAI.subset_precomputed_tsv}()},
\code{\link{SPLICEAI.subset_precomputed_vcf}()}
}
\concept{SpliceAI}
