% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpliceAI.R
\name{SPLICEAI.run}
\alias{SPLICEAI.run}
\title{Run pre-trained \emph{SpliceAI} model to get predictions}
\source{
\href{https://github.com/Illumina/SpliceAI}{GitHub}
\href{https://www.sciencedirect.com/science/article/pii/S0092867418316295}{Publication}
}
\usage{
SPLICEAI.run(
  vcf_path = "./GWAS_converted.vcf",
  output_path = "spliceai_predictions.vcf",
  reference_fasta = "/pd-omics/tools/polyfun/reference_fasta/hg19.fa",
  gene_annotation = "./echolocatoR/tools/spliceAI/grch37.txt",
  distance = 50,
  mask = 0
)
}
\description{
Run pre-trained \emph{SpliceAI} model to get predictions
}
\seealso{
Other SpliceAI: 
\code{\link{SPLICEAI.plot}()},
\code{\link{SPLICEAI.snp_probs}()},
\code{\link{SPLICEAI.subset_precomputed_tsv_iterate}()},
\code{\link{SPLICEAI.subset_precomputed_tsv}()},
\code{\link{SPLICEAI.subset_precomputed_vcf}()}
}
\concept{SpliceAI}
