% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpliceAI.R
\name{SPLICEAI.subset_precomputed_tsv}
\alias{SPLICEAI.subset_precomputed_tsv}
\title{Query genome-wide \emph{SpliceAI} results file (tsv format)}
\source{
\href{https://github.com/Illumina/SpliceAI}{GitHub}
\href{https://www.sciencedirect.com/science/article/pii/S0092867418316295}{Publication}
}
\usage{
SPLICEAI.subset_precomputed_tsv(
  dat,
 
    precomputed_path = "/pd-omics/data/spliceAI/whole_genome_filtered_spliceai_scores.tsv.gz",
  merge_data = TRUE,
  drop_na = TRUE,
  filtered = TRUE
)
}
\description{
Query genome-wide \emph{SpliceAI} results file (tsv format)
}
\seealso{
Other SpliceAI: 
\code{\link{SPLICEAI.plot}()},
\code{\link{SPLICEAI.run}()},
\code{\link{SPLICEAI.snp_probs}()},
\code{\link{SPLICEAI.subset_precomputed_tsv_iterate}()},
\code{\link{SPLICEAI.subset_precomputed_vcf}()}
}
\concept{SpliceAI}
