% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpliceAI.R
\name{SPLICEAI.subset_precomputed_tsv_iterate}
\alias{SPLICEAI.subset_precomputed_tsv_iterate}
\title{Make multiple queries to genome-wide \emph{SpliceAI} results file (tsv format)}
\source{
\href{https://github.com/Illumina/SpliceAI}{GitHub}
\href{https://www.sciencedirect.com/science/article/pii/S0092867418316295}{Publication}
}
\usage{
SPLICEAI.subset_precomputed_tsv_iterate(
  sumstats_paths,
  precomputed_path = "/pd-omics/data/spliceAI/spliceai_scores.raw.snv.hg19.tsv.gz",
  nThread = 4,
  merge_data = T,
  drop_na = T,
  filtered = F,
  save_path = "./spliceAI_subset.tsv.gz"
)
}
\description{
Make multiple queries to genome-wide \emph{SpliceAI} results file (tsv format)
}
\examples{
\dontrun{
root.pd <- "/sc/arion/projects/pd-omics"
precomputed_path <- file.path(root.pd,"data/spliceAI/spliceai_scores.raw.snv.hg19.tsv.gz")
sumstats_paths <- list.files(file.path(root.pd, "/brian/Fine_Mapping/Data/GWAS/Nalls23andMe_2019"), pattern = "*.UKB_LD.Multi-finemap.tsv.gz", recursive = T, full.names = T)
DAT <- SPLICEAI.subset_precomputed_tsv_iterate(sumstats_paths=sumstats_paths)
}
}
\seealso{
Other SpliceAI: 
\code{\link{SPLICEAI.plot}()},
\code{\link{SPLICEAI.run}()},
\code{\link{SPLICEAI.snp_probs}()},
\code{\link{SPLICEAI.subset_precomputed_tsv}()},
\code{\link{SPLICEAI.subset_precomputed_vcf}()}
}
\concept{SpliceAI}
