% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpliceAI.R
\name{SPLICEAI.subset_precomputed_vcf}
\alias{SPLICEAI.subset_precomputed_vcf}
\title{Query genome-wide \emph{SpliceAI} results file (vcf format)}
\source{
\href{https://github.com/Illumina/SpliceAI}{GitHub}
\href{https://www.sciencedirect.com/science/article/pii/S0092867418316295}{Publication}
}
\usage{
SPLICEAI.subset_precomputed_vcf(
  dat,
 
    precomputed_path = "/pd-omics/data/spliceAI/whole_genome_filtered_spliceai_scores.vcf.gz",
  subset_vcf = "subset.vcf"
)
}
\description{
Query genome-wide \emph{SpliceAI} results file (vcf format)
}
\seealso{
Other SpliceAI: 
\code{\link{SPLICEAI.plot}()},
\code{\link{SPLICEAI.run}()},
\code{\link{SPLICEAI.snp_probs}()},
\code{\link{SPLICEAI.subset_precomputed_tsv_iterate}()},
\code{\link{SPLICEAI.subset_precomputed_tsv}()}
}
\concept{SpliceAI}
