% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize_gene.R
\name{standardize_gene}
\alias{standardize_gene}
\title{Standardize genes}
\usage{
standardize_gene(dat, gene_col = "gene", verbose = TRUE)
}
\description{
Convert ensembl IDs to gene symbols.
}
\keyword{internal}
