% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize.R
\name{standardize_subset}
\alias{standardize_subset}
\title{Stanardize the locus subset}
\usage{
standardize_subset(
  locus,
  top_SNPs = NULL,
  subset_path = "./Data",
  chrom_col = "CHR",
  position_col = "POS",
  snp_col = "SNP",
  pval_col = "P",
  effect_col = "Effect",
  stderr_col = "StdErr",
  tstat_col = "t_stat",
  MAF_col = "MAF",
  freq_col = "Freq",
  N_cases_col = "N_cases",
  N_controls_col = "N_controls",
  N_cases = NULL,
  N_controls = NULL,
  proportion_cases = "calculate",
  A1_col = "A1",
  A2_col = "A2",
  return_dt = T,
  verbose = T
)
}
\arguments{
\item{top_SNPs}{A data.frame with the genomic coordinates of the lead SNP for each locus.
The lead SNP will be used as the center of the window when extracting subset from the full GWAS/QTL summary statistics file.
Only one SNP per \strong{Locus} should be included.
At minimum, \code{top_SNPs} should include the following columns:
\describe{
\item{Locus}{A unique name for each locus. Often, loci are named after a relevant gene (e.g. LRRK2) or based on the name/coordinates of the lead SNP (e.g. locus_chr12_40734202) }
\item{CHR}{The chromosome that the SNP is on. Can be "chr12" or "12" format.}
\item{POS}{The genomic position of the SNP (in basepairs)}
}}

\item{chrom_col}{Name of the chromosome column in the full summary stats file.}

\item{position_col}{Name of the genomic position column in the full summary stats file.}

\item{snp_col}{Name of the SNP RSID column in the full summary stats file.}

\item{pval_col}{Name of the p-value column in the full summary stats file.
Raw p-values are preferred, but if not available corrected p-values (e.g. FDR) can be used instead.}

\item{effect_col}{Name of the effect size column in the full summary stats file.
Effect size is preferred, but if not available other metrics like Beta for Odds Ratio can be used instead.}

\item{stderr_col}{Name of the standard error  column in the full summary stats file.
You can also set \code{stderr_col="calculate"} to infer standard error using: \code{effect / tstat}.}

\item{tstat_col}{Name of the t-statistic column in the full summary stats file.
This column is not necessary unless \code{stderr_col="calculate"} or the standard error column is missing.}

\item{MAF_col}{Name of the minor allele frequency column in the full summary stats file.
Can be inferred from \strong{freq_col} if missing from the dataset.}

\item{freq_col}{Name of the allele frequency column in the full summary stats file.
Effect allele frequency is preferred, but the non-effect allele can be provided instead (though this may be less accurate).
This column is not necessary unless \code{MAF_col="calculate"} or the MAF column is missing.}

\item{N_cases_col}{Name of the column in the full summary stats that has the number of case subjects in the study.
This can either be per SNP sample sizes, or one number repeated across all rows.
Proxy cases (e.g. relatives of people with the disease being investigated) should be included in this estimate if any were used in the study.
This column is not necesssary if \code{N_cases} parameter is provided.}

\item{N_controls_col}{Name of the column in the full summary stats that has the number of control subjects in the study.
 This can either be per SNP sample sizes, or one number repeated across all rows.
 This column is not necesssary if \code{N_controls} parameter is provided.
  @param N_cases The number of case subjects in the study.
 Instead of providing a reudundant \strong{N_cases_col} column, you can simply enter one value here.
 @param N_controls The number of control subjects in the study.
 Instead of providing a reudundant \strong{N_controls_col} column, you can simply enter one value here.
 @param proportion_cases The proportion of total subjects in the study that were cases.
 if \code{proportion_cases="calculate"} then this is inferred:  \code{N_controls / N_controls}.
 #' @param sample_size The overall sample size of the study.
If none is given, and \strong{N_cases} and \strong{N_controls} columns are present,
then sample_size is inferred to be \code{max(N_cases) + max(N_controls)}.}

\item{A1_col}{Name of the effect/risk allele column in the full summary stats.
Unfortunately, different studies report different kinds of allele information in a non-standardized way.
Meaning that A1/A2 can refer to any number of things:
 \describe{
 \item{effect/other alleles}{in the case of diseases}
 \item{ref/alt alleles}{where ref is the reference genome being used}
 \item{major/minor alleles}{This dichotomy holds true for bi-allelic SNPs but not necessary multi-allelic SNPs}
 }
 This makes comparing summary stats across GWAS/QTL datasets very confusing for several reasons:
 \describe{
 \item{Multi-allelic SNPs}{SNPs can have more than just 2 possible alleles (multi-allelic SNPs). Even if you compare the same SNP between two studies, you may accidentally be comparing totally different alleles.}
 \item{Valence}{The valence (+/-) of per-SNP GWAS effect sizes/beta can be relative to different allele types between studies.
 For example, let's say in one GWAS study your effect size for SNP A is 1.5 relative to the major allele in one study,
  and the minor allele happens to be the one found in the reference genome.
  You then try to compare that effect size to that of the same SNP in another GWAS.
  But, the valence of the effect sizes in the 2nd GWAS study are all relative to the reference genome (instead of the minor allele),
  giving the same SNP a value of -1.2. If you took the effect sizes at face value you'd say the signals are in opposite directions.
  But once you take into account how the valences were determined in each study you realize that they're actually both positive relative to the major allele.}
 }
This process of reversing per-SNP valences based on aligning the alleles is known as allele flipping.
This is important when comparing individual SNPs, but can also have an impact on colocalization results.}

\item{verbose}{Whether \emph{echolocatoR} should be verbose or silent.}
}
\description{
After querying a subset of the full summary statistics,
this function converts it into a standardized format
that the rest of \emph{echolocatoR} can work with.
}
\seealso{
Other standardizing functions: 
\code{\link{get_UKB_MAF}()}
}
\concept{standardizing functions}
