% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startup.R
\name{startup}
\alias{startup}
\title{Startup messages}
\usage{
startup(package = "echolocatoR", add_batty = FALSE)
}
\description{
Startup messages with info on package name, version, citation, GitHub, etc.
}
\keyword{internal}
