% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{startup_image}
\alias{startup_image}
\title{Ensembl IDs -> Gene symbols}
\usage{
startup_image()
}
\description{
Print the echolocatoR image to the R console.
}
\examples{
startup_image()
}
\seealso{
Other general: 
\code{\link{.rbind.file.list}()},
\code{\link{createDT_html}()},
\code{\link{createDT}()},
\code{\link{dt.replace}()},
\code{\link{example_fullSS}()},
\code{\link{fillNA_CS_PP}()},
\code{\link{get_os}()},
\code{\link{get_sample_size}()},
\code{\link{merge_coloc_results}()},
\code{\link{tryFunc}()}
}
\concept{general}
\keyword{internal}
