% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAIN.R
\name{subset_common_snps}
\alias{subset_common_snps}
\title{Subset LD matrix and data.frame to only their shared SNPs}
\usage{
subset_common_snps(LD_matrix, finemap_DT, verbose = F)
}
\value{
data.frame
}
\description{
Find the SNPs that are shared between an LD matrix and another data.frame with a `SNP` column.
Then remove any non-shared SNPs from both objects.
}
