% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise.R
\name{SUMMARISE.cell_type_specificity}
\alias{SUMMARISE.cell_type_specificity}
\title{Get cell-type-specifity score for each cell type}
\usage{
SUMMARISE.cell_type_specificity(
  plot_dat,
  merged_DT,
  min_count = NULL,
  top_celltype_only = F,
  label_yaxis = T,
  y_lab = NULL,
  show_genes = F,
  x_strip_angle = 40,
  show_plot = T
)
}
\description{
Aggregate SNP overlap across various epigenomic datasets
and then identify the number of SNPs overlapping by each cell type
}
