% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise.R
\name{SUMMARISE.coloc_nominated_eGenes}
\alias{SUMMARISE.coloc_nominated_eGenes}
\title{Nominate target genes within each locus}
\usage{
SUMMARISE.coloc_nominated_eGenes(
  coloc_results_path,
  merged_DT,
  label_yaxis = T,
  y_lab = "Locus",
  x_lab = NULL,
  fill_var = "PP.H4",
  text_size = 2,
  PP_threshold = NULL,
  nThread = 4,
  show_plot = T,
  verbose = T
)
}
\description{
Across all GWAS-QTL colocalization tests across all studies,
take the eGene with the highest colocalziation probability (PP.H4)
and assign it as the most likely causal gene in that locus.
}
\details{
eQTL queries and colocalization test done with \pkg{catalogueR}.
}
\examples{
\dontrun{
data("merged_DT")
base_url <- "~/Desktop/Fine_Mapping/Data/GWAS/Nalls23andMe_2019"
coloc_results_path <- file.path(base_url,"_genome_wide/COLOC/coloc.eQTL_Catalogue_ALL.csv.gz")
gg_egene <- SUMMARISE.coloc_nominated_eGenes(coloc_results_path, merged_DT=merged_DT, fill_var=NULL)
}
}
