% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise.R
\name{SUMMARISE.CS_bin_plot}
\alias{SUMMARISE.CS_bin_plot}
\title{Plot CS bin counts}
\usage{
SUMMARISE.CS_bin_plot(merged_DT, show_plot = T)
}
\description{
Plot CS bin counts
}
\examples{
data("merged_DT");
bin_plot <- SUMMARISE.CS_bin_plot(merged_DT=merged_DT)
}
\seealso{
Other summarise: 
\code{\link{SUMMARISE.CS_counts_plot}()},
\code{\link{SUMMARISE.get_CS_bins}()},
\code{\link{SUMMARISE.get_CS_counts}()},
\code{\link{SUMMARISE.get_SNPgroup_counts}()},
\code{\link{SUMMARISE.peak_overlap_plot}()},
\code{\link{results_report}()}
}
\concept{summarise}
