% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise.R
\name{SUMMARISE.CS_counts_plot}
\alias{SUMMARISE.CS_counts_plot}
\title{Bar plot of tool-specific CS sizes}
\usage{
SUMMARISE.CS_counts_plot(
  merged_DT,
  show_numbers = T,
  ylabel = "Locus",
  legend_nrow = 3,
  label_yaxis = T,
  show_plot = T
)
}
\description{
Loci ordered by UCS size (smallest to largest).
}
\examples{
gg_CS <- SUMMARISE.CS_counts_plot(merged_DT=merged_DT)
}
\seealso{
Other summarise: 
\code{\link{SUMMARISE.CS_bin_plot}()},
\code{\link{SUMMARISE.get_CS_bins}()},
\code{\link{SUMMARISE.get_CS_counts}()},
\code{\link{SUMMARISE.get_SNPgroup_counts}()},
\code{\link{SUMMARISE.peak_overlap_plot}()},
\code{\link{results_report}()}
}
\concept{summarise}
