% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise.R
\name{SUMMARISE.get_CS_bins}
\alias{SUMMARISE.get_CS_bins}
\title{Count bins of tool-specific and union CS sizes}
\usage{
SUMMARISE.get_CS_bins(merged_DT)
}
\description{
Count bins of tool-specific and union CS sizes
}
\examples{
data("merged_DT");
bin_counts <- SUMMARISE.get_CS_bins(merged_DT=merged_DT)
}
\seealso{
Other summarise: 
\code{\link{SUMMARISE.CS_bin_plot}()},
\code{\link{SUMMARISE.CS_counts_plot}()},
\code{\link{SUMMARISE.get_CS_counts}()},
\code{\link{SUMMARISE.get_SNPgroup_counts}()},
\code{\link{SUMMARISE.peak_overlap_plot}()},
\code{\link{results_report}()},
\code{\link{super_summary_plot}()}
}
\concept{summarise}
