% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise.R
\name{SUMMARISE.get_CS_counts}
\alias{SUMMARISE.get_CS_counts}
\title{Tally tool-specific and union CS sizes}
\usage{
SUMMARISE.get_CS_counts(merged_dat, top_CS_only = F)
}
\description{
Tally tool-specific and union CS sizes
}
\examples{
data("merged_DT");
locus_order <- SUMMARISE.get_CS_counts(merged_dat=merged_DT)
}
\seealso{
Other summarise: 
\code{\link{SUMMARISE.CS_bin_plot}()},
\code{\link{SUMMARISE.CS_counts_plot}()},
\code{\link{SUMMARISE.get_CS_bins}()},
\code{\link{SUMMARISE.get_SNPgroup_counts}()},
\code{\link{SUMMARISE.peak_overlap_plot}()},
\code{\link{SUMMARISE.plot_dataset_overlap}()},
\code{\link{results_report}()},
\code{\link{super_summary_plot}()}
}
\concept{summarise}
