% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise.R
\name{SUMMARISE.get_SNPgroup_counts}
\alias{SUMMARISE.get_SNPgroup_counts}
\title{Tally locus-specific SNP group sizes}
\usage{
SUMMARISE.get_SNPgroup_counts(merged_DT)
}
\description{
Tally locus-specific SNP group sizes
}
\examples{
data("merged_DT");
snp_groups <- SUMMARISE.get_SNPgroup_counts(merged_DT=merged_DT)
}
\seealso{
Other summarise: 
\code{\link{SUMMARISE.CS_bin_plot}()},
\code{\link{SUMMARISE.CS_counts_plot}()},
\code{\link{SUMMARISE.get_CS_bins}()},
\code{\link{SUMMARISE.get_CS_counts}()},
\code{\link{SUMMARISE.peak_overlap_plot}()}
}
\concept{summarise}
