% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise.R
\name{super_summary_plot}
\alias{super_summary_plot}
\title{Merge all summary plots into one super plot}
\usage{
super_summary_plot(
  merged_DT,
  snp_filter = "Consensus_SNP==T",
  coloc_results_path = NULL,
  show_plot = T,
  save_plot = F,
  height = 15,
  width = 13,
  dpi = 500
)
}
\description{
Merge all summary plots into one super plot
}
\seealso{
Other summarise: 
\code{\link{SUMMARISE.CS_bin_plot}()},
\code{\link{SUMMARISE.CS_counts_plot}()},
\code{\link{SUMMARISE.get_CS_bins}()},
\code{\link{SUMMARISE.get_CS_counts}()},
\code{\link{SUMMARISE.get_SNPgroup_counts}()},
\code{\link{SUMMARISE.peak_overlap_plot}()},
\code{\link{results_report}()}
}
\concept{summarise}
