% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SURE.R
\name{SURE.download_annotations}
\alias{SURE.download_annotations}
\title{Download \emph{SuRE} annotations}
\source{
\href{https://www.nature.com/articles/s41588-019-0455-2}{Publication}
\href{https://github.com/vansteensellab/SuRE-SNV-code}{GitHub}
\href{https://osf.io/w5bzq/wiki/home/?view}{Full SuRE data}
\href{https://sure.nki.nl}{SNP-SuRE data browse}
}
\usage{
SURE.download_annotations(
  URL = "https://osf.io/vxfk3/download",
  output_dir = ".",
  nThread = 1,
  v = verbose,
  conda_env = "echoR"
)
}
\description{
Download \emph{SuRE} annotations
}
