% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SURE.R
\name{SURE.import_annotations}
\alias{SURE.import_annotations}
\title{Download \emph{SuRE} annotations}
\usage{
SURE.import_annotations(
  base_url = "https://osf.io/vxfk3/download",
  nThread = 4
)
}
\description{
Download \emph{SuRE} annotations
}
