% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SURE.R
\name{SURE.merge}
\alias{SURE.merge}
\title{Merge with fine-mapping and \emph{SuRE} results}
\source{
\href{https://www.nature.com/articles/s41588-019-0455-2}{Publication}
\href{https://github.com/vansteensellab/SuRE-SNV-code}{GitHub}
\href{https://osf.io/w5bzq/wiki/home/?view}{Full SuRE data}
\href{https://sure.nki.nl}{SNP-SuRE data browse}
}
\usage{
SURE.merge(merged_dat, sure, save_path = FALSE, verbose = TRUE)
}
\description{
Merge with fine-mapping and \emph{SuRE} results
}
\examples{
sure <- data.table::fread("/sc/arion/projects/pd-omics/data/MPRA/SURE/SuRE_SNP_table_LP190708.txt.gz", nThread=1)
merged_dat <- merge_finemapping_results("Data/GWAS/Nalls23andMe_2019",LD_reference = "UKB", minimum_support = 0)

sure_DT <- SURE.merge(merged_dat=merged_dat, sure=sure, save_path="/sc/arion/projects/pd-omics/brian/Fine_Mapping/Data/GWAS/Nalls23andMe_2019/_genome_wide/SURE/Nalls23andMe_2019.SURE.csv.gz")
}
\seealso{
Other SURE: 
\code{\link{SURE.melt_snp_groups}()},
\code{\link{SURE.plot}()}
}
\concept{SURE}
