% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SURE.R
\name{SURE.plot}
\alias{SURE.plot}
\title{Plot \emph{SuRE} results across SNP groups}
\source{
\href{https://www.nature.com/articles/s41588-019-0455-2}{Publication}
\href{https://github.com/vansteensellab/SuRE-SNV-code}{GitHub}
\href{https://osf.io/w5bzq/wiki/home/?view}{Full SuRE data}
\href{https://sure.nki.nl}{SNP-SuRE data browse}
}
\usage{
SURE.plot(
  sure.melt,
  snp_groups = c("GWAS lead", "UCS", "Consensus (-POLYFUN)", "Consensus"),
  comparisons_filter = function(x) {     if ("Consensus" \%in\% x)          return(x) },
  title = "SuRE MPRA",
  xlabel = "SNP Group",
  facet_formula = ". ~ Cell_type",
  show_padj = T,
  show_signif = T,
  vjust_signif = 0.5,
  show_plot = T,
  save_path = F,
  height = 5,
  width = 5
)
}
\description{
Plot \emph{SuRE} results across SNP groups
}
\examples{
sure.melt <- data.table::fread("/sc/arion/projects/pd-omics/brian/Fine_Mapping/Data/GWAS/Nalls23andMe_2019/_genome_wide/SURE/Nalls23andMe_2019.SURE.snp_groups.mean.csv.gz")
}
\seealso{
Other SURE: 
\code{\link{SURE.melt_snp_groups}()},
\code{\link{SURE.merge}()}
}
\concept{SURE}
