% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SURE.R
\name{SURE.preprocess_data}
\alias{SURE.preprocess_data}
\title{Preprocess \emph{SuRE} MPRA results}
\source{
\href{https://www.nature.com/articles/s41588-019-0455-2}{Publication}
\href{https://github.com/vansteensellab/SuRE-SNV-code}{GitHub}
\href{https://osf.io/w5bzq/wiki/home/?view}{Full SuRE data}
\href{https://sure.nki.nl}{SNP-SuRE data browse}
}
\usage{
SURE.preprocess_data(
  results_path = "./Data/GWAS/Nalls23andMe_2019/LRRK2",
  cor.plot = F
)
}
\description{
Preprocess \emph{SuRE} MPRA results
}
