% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SUSIE.R
\name{SUSIE}
\alias{SUSIE}
\title{Fine-map with SUSIE}
\source{
\url{https://stephenslab.github.io/susieR/}
}
\usage{
SUSIE(
  subset_DT,
  LD_matrix,
  dataset_type = "GWAS",
  n_causal = 5,
  sample_size = NULL,
  var_y = "estimate",
  prior_weights = NULL,
  PP_threshold = 0.95,
  scaled_prior_variance = 0.001,
  estimate_residual_variance = F,
  verbose = T
)
}
\description{
Sum of Single Effects (SuSiE): Iterative Bayesian Step-wise Selection
}
\examples{
data("BST1"); data("LD_matrix");
BST1 <- get_sample_size(subset_DT = BST1)
finemap_DT <- SUSIE(subset_DT=BST1, LD_matrix=LD_matrix)
}
