% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabix.R
\name{TABIX.convert_file}
\alias{TABIX.convert_file}
\title{Convert summary stats file to tabix format}
\usage{
TABIX.convert_file(
  fullSS_path,
  chrom_col = "CHR",
  position_col = "POS",
  remove_header = T
)
}
\arguments{
\item{fullSS_path}{Path to the full summary statistics file (GWAS or QTL) that you want to fine-map.}

\item{chrom_col}{Name of the chromosome column in the full summary stats file.}

\item{position_col}{Name of the genomic position column in the full summary stats file.}

\item{remove_header}{Remove the header to avoid errors when sorting/converting to tabix forrmat.}
}
\description{
Currently removes header and saves it as  \emph{header.txt}
in the same directory as the file it came from.
}
\seealso{
Other query functions: 
\code{\link{TABIX.query}()},
\code{\link{TABIX}()},
\code{\link{extract_SNP_subset}()},
\code{\link{import_topSNPs}()},
\code{\link{query_by_coordinates_merged}()},
\code{\link{query_by_coordinates}()},
\code{\link{query_by_gene}()},
\code{\link{query_by_probe}()},
\code{\link{query_fullSS}()},
\code{\link{query_handler}()}
}
\concept{query functions}
