% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabix.R
\name{TABIX.query}
\alias{TABIX.query}
\title{Query a tabix file}
\usage{
TABIX.query(fullSS.gz, chrom, start_pos, end_pos)
}
\description{
Query by min/max genomic coordinates.
}
\seealso{
Other query functions: 
\code{\link{TABIX.convert_file}()},
\code{\link{TABIX}()},
\code{\link{extract_SNP_subset}()},
\code{\link{import_topSNPs}()},
\code{\link{query_by_coordinates_merged}()},
\code{\link{query_by_coordinates}()},
\code{\link{query_by_gene}()},
\code{\link{query_by_probe}()},
\code{\link{query_fullSS}()},
\code{\link{query_handler}()}
}
\concept{query functions}
