% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabix.R
\name{TABIX}
\alias{TABIX}
\title{Covert and query}
\usage{
TABIX(
  fullSS_path,
  subset_path,
  is_tabix = F,
  chrom_col = "CHR",
  position_col = "POS",
  min_POS = NA,
  max_POS = NA,
  chrom = NULL
)
}
\arguments{
\item{fullSS_path}{Path to the full summary statistics file (GWAS or QTL) that you want to fine-map.}

\item{chrom_col}{Name of the chromosome column in the full summary stats file.}

\item{position_col}{Name of the genomic position column in the full summary stats file.}

\item{min_POS}{Manually set the minimum genomic position for your locus subset.
\code{min_POS} can clip the window size set by \code{bp_distance}.
Only use this parameter when fine-mapping one locus at a time.}

\item{max_POS}{Manually set the maximum genomic position for your locus subset.
\code{max_POS} can clip the window size set by \code{bp_distance}.
Only use this parameter when fine-mapping one locus at a time.}
}
\value{
data.table of locus subset summary statistics
}
\description{
If it is not tabix format already
(determined by checking for a \link{.tbi} file of the same name in the same directory),
the full summary statistics file is converted into tabix format for super fast querying.
A query is then made using the min/max genomic positions to extract a locus-specific summary stats file.
}
\seealso{
Other query functions: 
\code{\link{TABIX.convert_file}()},
\code{\link{TABIX.query}()},
\code{\link{extract_SNP_subset}()},
\code{\link{import_topSNPs}()},
\code{\link{query_by_coordinates_merged}()},
\code{\link{query_by_coordinates}()},
\code{\link{query_by_gene}()},
\code{\link{query_by_probe}()},
\code{\link{query_fullSS}()},
\code{\link{query_handler}()}
}
\concept{query functions}
