% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggbio.R
\name{transcript_model_track}
\alias{transcript_model_track}
\title{Track plot for transcript models}
\usage{
transcript_model_track(
  gr.snp_CHR,
  max_transcripts = 1,
  show.legend = T,
  show_plot = F
)
}
\arguments{
\item{gr.snp_CHR}{\emph{echolocator} fine-mapping results
in \code{\link{GRanges}} format.}

\item{max_transcripts}{Max number of transcripts per gene to show.}

\item{show.legend}{Whether to show legend.}
}
\description{
Uses \code{\link{ggbio}}.
}
\seealso{
Other plot: 
\code{\link{GGBIO.plot}()},
\code{\link{SNP_track}()},
\code{\link{invisible_legend}()}
}
\concept{plot}
\keyword{internal}
