% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PAINTOR.R
\name{transethnic_plot}
\alias{transethnic_plot}
\title{Plot transethnic PAINTOR results}
\usage{
transethnic_plot(
  merged_DT,
  save_path,
  title = locus,
  subtitle = "Trans-ethnic Fine-mapping",
  PAINTOR.label = "PAINTOR\\nTrans-ethnic",
  conditions = c("MESA_AFA", "MESA_CAU", "MESA_HIS")
)
}
\description{
Plot transethnic PAINTOR results
}
\keyword{internal}
