% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{update_cols}
\alias{update_cols}
\title{Update CS cols}
\usage{
update_cols(finemap_dat)
}
\description{
Convert old col format: ".Credible_Set" => ".CS"
}
\examples{
data("BST1");
finemap_DT <- update_cols(finemap_dat=BST1)
}
