% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{update_CS_cols}
\alias{update_CS_cols}
\title{Update CS cols}
\usage{
update_CS_cols(finemap_dat)
}
\description{
Convert old col format: ".Credible_Set" => ".CS"
}
\examples{
\dontrun{
data("BST1");
finemap_DT <- BST1
finemap_DT <- update_CS_cols(finemap_dat=finemap_DT)
}
}
