% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{VALIDATION.bootstrap}
\alias{VALIDATION.bootstrap}
\title{Perform validation bootstrap procedure}
\usage{
VALIDATION.bootstrap(
  metric_df,
  metric,
  validation_method = NULL,
  synthesize_random = F,
  snp_groups = c("Random", "GWAS lead", "UCS", "Consensus (-POLYFUN)", "Consensus"),
  save_path = F,
  nThread = 4
)
}
\description{
Perform validation bootstrap procedure
}
\examples{
\dontrun{
save_path <- root <- "/sc/arion/projects/pd-omics/brian/Fine_Mapping/Data/GWAS/Nalls23andMe_2019/_genome_wide"
ANNOT_MELT <- data.table::fread("/sc/arion/projects/pd-omics/data/IMPACT/IMPACT707/Annotations/IMPACT_overlap.csv.gz", nThread=4)
res.IMPACT <- data.table::fread(file.path(root,"IMPACT/TOP_IMPACT_all.csv.gz"))

boot.h2 <- VALIDATION.bootstrap()
boot.IMPACT
boot.DL
boot.
}
}
\seealso{
Other VALIDATION: 
\code{\link{VALIDATION.compare_distributions}()},
\code{\link{VALIDATION.permute_plot}()},
\code{\link{VALIDATION.super_plot}()}
}
\concept{VALIDATION}
