% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{VALIDATION.bootstrap_multimetric}
\alias{VALIDATION.bootstrap_multimetric}
\title{Conduct bootstrap procedure on multiple columns}
\usage{
VALIDATION.bootstrap_multimetric(
  metric_df,
  metric_names,
  validation_method = NULL,
  test_method = "stats_wilcox.test",
  save_path = F,
  verbose = T
)
}
\description{
Conduct bootstrap procedure on multiple columns
}
\examples{
save_path <- root <-  "/sc/arion/projects/pd-omics/brian/Fine_Mapping/Data/GWAS/Nalls23andMe_2019/_genome_wide"
path <- file.path(root,"Dey_DeepLearning/Nalls23andMe_2019.Dey_DeepLearning.annot.Allelic_Effect.csv.gz")
path <- file.path(root,"Dey_DeepLearning/Nalls23andMe_2019.Dey_DeepLearning.annot.Variant_Level.csv.gz")

# Import and process
metric_df <- data.table::fread(path, nThread=8)
metric_df <- find_consensus_SNPs_no_PolyFun(metric_df)
metric_names <- grep("Basenji.*MAX|DeepSEA.*MAX|Roadmap.*MAX", colnames(metric_df), value = T)
validation_method = "Dey_DeepLearning"

enrich.boot <- VALIDATION.bootstrap_multimetric(metric_df=metric_df, metric_names=metric_names, validation_method=validation_method, save_path=gsub("\\\\.csv\\\\.gz",".bootstrap.stats_wilcox.test.csv.gz",path))
}
\seealso{
Other VALIDATION: 
\code{\link{VALIDATION.bootstrap}()},
\code{\link{VALIDATION.compare_distributions}()},
\code{\link{VALIDATION.permute_plot}()},
\code{\link{VALIDATION.super_plot}()}
}
\concept{VALIDATION}
