% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{VALIDATION.compare_bootstrap_distributions}
\alias{VALIDATION.compare_bootstrap_distributions}
\title{Test whether permutation p-value distributions are different}
\usage{
VALIDATION.compare_bootstrap_distributions(
  boot_res,
  formula_str = "stat ~ SNP_group"
)
}
\description{
Test whether permutation p-value distributions are different
}
\examples{
root <- "/sc/arion/projects/pd-omics/brian/Fine_Mapping/Data/GWAS/Nalls23andMe_2019"
permute.IMPACT <-  data.table::fread(file.path(root,"_genome_wide/IMPACT/Nalls23andMe_2019.IMPACT.permutations.csv.gz"))
res <- VALIDATION.permute_compare_results(permute_res=permute.IMPACT)
}
\seealso{
Other VALIDATION: 
\code{\link{VALIDATION.bootstrap_multimetric}()},
\code{\link{VALIDATION.bootstrap_plot}()},
\code{\link{VALIDATION.bootstrap}()},
\code{\link{VALIDATION.permute}()},
\code{\link{VALIDATION.super_plot}()}
}
\concept{VALIDATION}
