% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{VALIDATION.compare_support}
\alias{VALIDATION.compare_support}
\title{Check whether Support level correlates with some variable}
\usage{
VALIDATION.compare_support()
}
\description{
Check whether Support level correlates with some variable
}
\examples{
# S-LDSC h2
h2_stats <- data.table::fread("/sc/arion/projects/pd-omics/brian/Fine_Mapping/Data/GWAS/Nalls23andMe_2019/_genome_wide/PolyFun/Nalls23andMe_2019.grouped_snpvar_stats.csv.gz")
}
