% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{VALIDATION.permute_plot}
\alias{VALIDATION.permute_plot}
\title{Plot permutation test results}
\usage{
VALIDATION.permute_plot(
  permute_res,
  validation_method = "Dey_DeepLearning",
  show_plot = T
)
}
\description{
Plot permutation test results
}
\examples{
root <- "/sc/arion/projects/pd-omics/brian/Fine_Mapping/Data/GWAS/Nalls23andMe_2019/_genome_wide"

 ## IMPACT
#permute_res <-  data.table::fread(file.path(root,"IMPACT/Nalls23andMe_2019.IMPACT.permutations.csv.gz"))
## Dey_DeepLearning
permute_res <-  data.table::fread(file.path(root,"Dey_DeepLearning/Nalls23andMe_2019.Dey_DeepLearning.permutations.csv.gz"))
gp <- VALIDATION.permute_plot(permute_res=permute_res)
}
\seealso{
Other VALIDATION: 
\code{\link{VALIDATION.compare_distributions}()},
\code{\link{VALIDATION.super_plot}()}
}
\concept{VALIDATION}
