% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{VALIDATION.super_plot}
\alias{VALIDATION.super_plot}
\title{Merge validation assays into a single plot}
\usage{
VALIDATION.super_plot(
 
    root = "/sc/arion/projects/pd-omics/brian/Fine_Mapping/Data/GWAS/Nalls23andMe_2019/_genome_wide",
  height = 10,
  width = 12,
  show_plot = T,
  save_plot = F
)
}
\description{
Merge validation assays into a single plot
}
\examples{
plt.ALL <- VALIDATION.super_plot(root="/sc/arion/projects/pd-omics/brian/Fine_Mapping/Data/GWAS/Nalls23andMe_2019/_genome_wide")
}
\seealso{
Other VALIDATION: 
\code{\link{VALIDATION.bootstrap_multimetric}()},
\code{\link{VALIDATION.bootstrap}()},
\code{\link{VALIDATION.compare_distributions}()},
\code{\link{VALIDATION.permute_plot}()}
}
\concept{VALIDATION}
