% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf_cleaning.R
\name{vcf_cleaning}
\alias{vcf_cleaning}
\title{Remove old vcf files}
\usage{
vcf_cleaning(root, LD_ref = "1KGphase3", loci = NULL, verbose = T)
}
\description{
Remove old vcf files
}
\examples{
vcf_cleaning(root="Data/GWAS/Ripke_2014", LD_ref="1KGphase3", loci=top_SNPs$Locus)
}
