% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloaders.R
\name{wget}
\alias{wget}
\title{wget}
\usage{
wget(
  input.url,
  output.path,
  background = T,
  force_overwrite = F,
  quiet = F,
  show_progress = T,
  continue = T
)
}
\description{
R wrapper for wget
}
\seealso{
Other download functions: 
\code{\link{axel}()}
}
\concept{download functions}
\keyword{internal}
