% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XGR.R
\name{XGR.enrichment}
\alias{XGR.enrichment}
\title{XGR enrichment}
\usage{
XGR.enrichment(
  gr,
  merged_dat,
  foreground_filter = "Consensus_SNP==T",
  background_filter = NULL,
  grouping_vars = c("Study", "Assay", "Cell_type"),
  fg_sample_size = NULL,
  bg_sample_size = NULL,
  background.annotatable.only = F,
  verbose = T
)
}
\description{
XGR enrichment
}
\examples{
root <- "/sc/arion/projects/pd-omics/brian/Fine_Mapping/Data/GWAS/Nalls23andMe_2019/_genome_wide"

merged_dat <- merge_finemapping_results(dataset = dirname(root), LD_reference = "UKB", minimum_support = 0)
merged_dat$Consensus_SNP_noPF <- find_consensus_SNPs(merged_dat, exclude_methods = "POLYFUN_SUSIE", sort_by_support = F)$Consensus_SNP

gr.merged <- merge_celltype_specific_epigenomics()
grouping_vars <- c("Study","Cell_type","Assay")

enrich.lead <- XGR.enrichment(gr=gr, merged_dat=merged_dat, foreground_filter="leadSNP==T",  grouping_vars=grouping_vars)
}
\seealso{
Other XGR: 
\code{\link{DT_to_GRanges}()},
\code{\link{GRanges_to_BED}()},
\code{\link{XGR.download_and_standardize}()},
\code{\link{XGR.enrichment_bootstrap}()},
\code{\link{XGR.enrichment_plot}()},
\code{\link{XGR.filter_assays}()},
\code{\link{XGR.filter_sources}()},
\code{\link{XGR.import_annotations}()},
\code{\link{XGR.iterate_enrichment}()},
\code{\link{XGR.iterate_overlap}()},
\code{\link{XGR.merge_and_process}()},
\code{\link{XGR.plot_enrichment}()},
\code{\link{XGR.plot_peaks}()},
\code{\link{XGR.prepare_foreground_background}()}
}
\concept{XGR}
