% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XGR.R
\name{XGR.enrichment_bootstrap}
\alias{XGR.enrichment_bootstrap}
\title{XGR enrichment (bootstrapped)}
\usage{
XGR.enrichment_bootstrap(
  gr,
  merged_dat,
  snp_groups = c("Random", "GWAS lead", "UCS", "Consensus (-POLYFUN)", "Consensus"),
  background_filter = NULL,
  grouping_vars = c("Study", "Assay", "Cell_type"),
  iterations = 1000,
  fg_sample_size = 20,
  bg_sample_size = NULL,
  bootstrap = T,
  save_path = F,
  nThread = 4,
  verbose = T
)
}
\description{
XGR enrichment (bootstrapped)
}
\examples{
root <- "/sc/arion/projects/pd-omics/brian/Fine_Mapping/Data/GWAS/Nalls23andMe_2019/_genome_wide"

merged_dat <- merge_finemapping_results(dataset = dirname(root), LD_reference = "UKB", minimum_support = 0)
merged_dat <- find_consensus_SNPs_no_PolyFun(merged_dat)
no_no_loci<- c("HLA-DRB5","MAPT","ATG14","SP1","LMNB1","ATP6V0A1", "RETREG3","UBTF","FAM171A2","MAP3K14","CRHR1","MAPT-AS1","KANSL1","NSF","WNT3");
merged_dat <- subset(merged_dat, !Locus \%in\% no_no_loci)

gr.merged <- merge_celltype_specific_epigenomics()
grouping_vars <- c("Study","Cell_type","Assay")

enrich_res <- XGR.enrichment_bootstrap(gr=gr.merged, merged_dat=merged_dat, grouping_vars=grouping_vars,  bootstrap=F)

enrich_boot <- XGR.enrichment_bootstrap(gr=gr.merged, merged_dat=merged_dat, grouping_vars=grouping_vars,  bootstrap=T, fg_sample_size=NULL, bg_sample_size=100, iterations=100)
}
\seealso{
Other XGR: 
\code{\link{DT_to_GRanges}()},
\code{\link{GRanges_to_BED}()},
\code{\link{XGR.download_and_standardize}()},
\code{\link{XGR.enrichment_plot}()},
\code{\link{XGR.enrichment}()},
\code{\link{XGR.filter_assays}()},
\code{\link{XGR.filter_sources}()},
\code{\link{XGR.import_annotations}()},
\code{\link{XGR.iterate_enrichment}()},
\code{\link{XGR.iterate_overlap}()},
\code{\link{XGR.merge_and_process}()},
\code{\link{XGR.plot_enrichment}()},
\code{\link{XGR.plot_peaks}()},
\code{\link{XGR.prepare_foreground_background}()}
}
\concept{XGR}
