% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XGR.R
\name{XGR.filter_assays}
\alias{XGR.filter_assays}
\title{Filter assays}
\usage{
XGR.filter_assays(gr.lib, n_top_assays = 5)
}
\description{
Identify the assays with the most annotations in the locus.
Then only keep these assays
}
\seealso{
Other XGR: 
\code{\link{DT_to_GRanges}()},
\code{\link{GRanges_to_BED}()},
\code{\link{XGR.download_and_standardize}()},
\code{\link{XGR.enrichment_plot}()},
\code{\link{XGR.enrichment}()},
\code{\link{XGR.filter_sources}()},
\code{\link{XGR.import_annotations}()},
\code{\link{XGR.iterate_enrichment}()},
\code{\link{XGR.iterate_overlap}()},
\code{\link{XGR.merge_and_process}()},
\code{\link{XGR.plot_enrichment}()},
\code{\link{XGR.plot_peaks}()},
\code{\link{XGR.prepare_foreground_background}()}
}
\concept{XGR}
\keyword{internal}
