% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XGR.R
\name{XGR.filter_sources}
\alias{XGR.filter_sources}
\title{Filter sources}
\usage{
XGR.filter_sources(gr.lib, n_top_sources = 5)
}
\description{
Identify the sources with the most annotations in the locus.
Then only keep these sources.
}
\seealso{
Other XGR: 
\code{\link{DT_to_GRanges}()},
\code{\link{GRanges_to_BED}()},
\code{\link{XGR.download_and_standardize}()},
\code{\link{XGR.filter_assays}()},
\code{\link{XGR.import_annotations}()},
\code{\link{XGR.iterate_enrichment}()},
\code{\link{XGR.iterate_overlap}()},
\code{\link{XGR.merge_and_process}()},
\code{\link{XGR.plot_enrichment}()},
\code{\link{XGR.plot_peaks}()},
\code{\link{XGR.prepare_foreground_background}()}
}
\concept{XGR}
\keyword{internal}
