% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XGR.R
\name{XGR.import_annotations}
\alias{XGR.import_annotations}
\title{Download XGR annotations}
\usage{
XGR.import_annotations(
  gr.snp,
  anno_data_path = file.path("annotations", paste0("XGR_", lib.name, ".rds")),
  lib.name,
  save_xgr = T,
  annot_overlap_threshold = 5
)
}
\description{
Download XGR annotations
}
\seealso{
Other XGR: 
\code{\link{DT_to_GRanges}()},
\code{\link{GRanges_to_BED}()},
\code{\link{XGR.download_and_standardize}()},
\code{\link{XGR.filter_assays}()},
\code{\link{XGR.filter_sources}()},
\code{\link{XGR.iterate_enrichment}()},
\code{\link{XGR.iterate_overlap}()},
\code{\link{XGR.merge_and_process}()},
\code{\link{XGR.plot_enrichment}()},
\code{\link{XGR.plot_peaks}()},
\code{\link{XGR.prepare_foreground_background}()}
}
\concept{XGR}
\keyword{internal}
