% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XGR.R
\name{XGR.iterate_enrichment}
\alias{XGR.iterate_enrichment}
\title{Conduct enrichment tests for each annotation}
\usage{
XGR.iterate_enrichment(
  subset_DT,
  foreground_filter = "Consensus_SNP",
  background_filter = "leadSNP",
  lib.selections = c("ENCODE_TFBS_ClusteredV3_CellTypes",
    "ENCODE_DNaseI_ClusteredV3_CellTypes", "Broad_Histone", "FANTOM5_Enhancer",
    "Segment_Combined_Gm12878", "TFBS_Conserved", "ReMap_PublicAndEncode_TFBS",
    "Blueprint_VenousBlood_Histone", "Blueprint_DNaseI", "FANTOM5_CAT_Cell",
    "FANTOM5_CAT_MESH", "GWAScatalog_alltraits"),
  save_path = F,
  nCores = 4
)
}
\arguments{
\item{subset_DT}{Data.frame with at least the following columns:
\describe{
\item{SNP}{SNP RSID}
\item{CHR}{chromosome}
\item{POS}{position}
}}

\item{foreground_filter}{Specify foreground by filtering SNPs in \code{subset_DT}.
Write filter as a string (or \code{NULL} to include all SNPs).}

\item{background_filter}{Specify background by filtering SNPs in \code{subset_DT}.
Write filter as a string (or \code{NULL} to include all SNPs).}
}
\description{
XGR uses a binomial enrichment tests for each annotation.
}
\examples{
data("merged_DT")
enrich_res <- XGR.iterate_enrichment(subset_DT=merged_DT, foreground_filter = "Consensus_SNP", background_filter = "leadSNP", lib.selections = c("ENCODE_TFBS_ClusteredV3_CellTypes"), nCores=1)
}
\seealso{
Other XGR: 
\code{\link{DT_to_GRanges}()},
\code{\link{GRanges_to_BED}()},
\code{\link{XGR.download_and_standardize}()},
\code{\link{XGR.import_annotations}()},
\code{\link{XGR.iterate_overlap}()},
\code{\link{XGR.merge_and_process}()},
\code{\link{XGR.plot_enrichment}()},
\code{\link{XGR.plot_peaks}()},
\code{\link{XGR.prepare_foreground_background}()}
}
\concept{XGR}
\keyword{internal}
