% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XGR.R
\name{XGR.merge_and_process}
\alias{XGR.merge_and_process}
\title{Standardize XGR annotations}
\usage{
XGR.merge_and_process(grl.xgr, lib, n_top_sources = 10)
}
\arguments{
\item{grl.xgr}{\code{\link[GenomicRanges]{GenomicRangesList}} of XGR queries.}
}
\description{
Parses the metadata and adds it as columns,
and then merges the results into a single
\code{\link[GenomicRanges]{GenomicRangesList}}
}
\seealso{
Other XGR: 
\code{\link{DT_to_GRanges}()},
\code{\link{GRanges_to_BED}()},
\code{\link{XGR.download_and_standardize}()},
\code{\link{XGR.enrichment_bootstrap}()},
\code{\link{XGR.enrichment_plot}()},
\code{\link{XGR.enrichment}()},
\code{\link{XGR.filter_assays}()},
\code{\link{XGR.filter_sources}()},
\code{\link{XGR.import_annotations}()},
\code{\link{XGR.iterate_enrichment}()},
\code{\link{XGR.iterate_overlap}()},
\code{\link{XGR.plot_enrichment}()},
\code{\link{XGR.plot_peaks}()},
\code{\link{XGR.prepare_foreground_background}()}
}
\concept{XGR}
\keyword{internal}
