% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XGR.R
\name{XGR.plot_enrichment}
\alias{XGR.plot_enrichment}
\title{Plot XGR enrichment}
\usage{
XGR.plot_enrichment(
  enrich_res,
  adjp_thresh = 0.05,
  top_annotations = NULL,
  show_plot = T
)
}
\description{
Plot XGR enrichment
}
\examples{
data("merged_DT")
enrich_res <- XGR.iterate_enrichment(subset_DT=merged_DT, foreground_filter = "Consensus_SNP", background_filter = "leadSNP", lib.selections = c("ENCODE_TFBS_ClusteredV3_CellTypes"), nCores=1)
XGR.plot_enrichment(enrich_res)
}
\seealso{
Other XGR: 
\code{\link{DT_to_GRanges}()},
\code{\link{GRanges_to_BED}()},
\code{\link{XGR.download_and_standardize}()},
\code{\link{XGR.enrichment_bootstrap}()},
\code{\link{XGR.enrichment_plot}()},
\code{\link{XGR.enrichment}()},
\code{\link{XGR.filter_assays}()},
\code{\link{XGR.filter_sources}()},
\code{\link{XGR.import_annotations}()},
\code{\link{XGR.iterate_enrichment}()},
\code{\link{XGR.iterate_overlap}()},
\code{\link{XGR.merge_and_process}()},
\code{\link{XGR.plot_peaks}()},
\code{\link{XGR.prepare_foreground_background}()}
}
\concept{XGR}
\keyword{internal}
