% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XGR.R
\name{XGR.prepare_foreground_background}
\alias{XGR.prepare_foreground_background}
\title{Prepare SNP sets for enrichment}
\usage{
XGR.prepare_foreground_background(
  subset_DT,
  foreground_filter = "Support>0",
  background_filter = NULL
)
}
\arguments{
\item{subset_DT}{Data.frame with at least the following columns:
\describe{
\item{SNP}{SNP RSID}
\item{CHR}{chromosome}
\item{POS}{position}
}}

\item{foreground_filter}{Specify foreground by filtering SNPs in \code{subset_DT}.
Write filter as a string (or \code{NULL} to include all SNPs).}

\item{background_filter}{Specify background by filtering SNPs in \code{subset_DT}.
Write filter as a string (or \code{NULL} to include all SNPs).}
}
\description{
Prepare custom foreground and background SNPs sets for enrichment tests with XGR annotations.
}
\examples{
data("merged_DT")
fg_bg <- XGR.prepare_foreground_background(subset_DT=merged_DT, foreground_filter="Consensus_SNP==T", background_filter="leadSNP==T")
}
\seealso{
Other XGR: 
\code{\link{DT_to_GRanges}()},
\code{\link{GRanges_to_BED}()},
\code{\link{XGR.download_and_standardize}()},
\code{\link{XGR.filter_assays}()},
\code{\link{XGR.filter_sources}()},
\code{\link{XGR.import_annotations}()},
\code{\link{XGR.iterate_enrichment}()},
\code{\link{XGR.iterate_overlap}()},
\code{\link{XGR.merge_and_process}()},
\code{\link{XGR.plot_enrichment}()},
\code{\link{XGR.plot_peaks}()},
\code{\link{save_annotations}()}
}
\concept{XGR}
\keyword{internal}
