% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xLiftOver.R
\name{xLiftOver}
\alias{xLiftOver}
\title{Genome build liftover}
\usage{
xLiftOver(
  data.file,
  format.file = c("data.frame", "bed", "chr:start-end", "GRanges"),
  build.conversion = c(NA, "hg38.to.hg19", "hg19.to.hg38", "hg19.to.hg18",
    "hg18.to.hg38", "hg18.to.hg19"),
  merged = T,
  verbose = T,
  RData.location = "http://galahad.well.ox.ac.uk/bigdata",
  guid = NULL
)
}
\description{
Transfer your genomic coordinates from one genome build to another.
}
\details{

}
