% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{Zscore.get_mean_and_sd}
\alias{Zscore.get_mean_and_sd}
\title{Store info necessary for Z-score}
\usage{
Zscore.get_mean_and_sd(
  fullSS,
  target_col = "statistic",
  effect_col = "beta",
  stderr_col = "se",
  use_saved = T,
  output_path
)
}
\description{
Store the info about the full vector (length, mean, standard deviation),
so that you can later accurately compute Z-scores
when you're only working with a subset of the data.
}
\details{
These functions are necessary for \code{\link{PAINTOR}}.
}
\keyword{internal}
