% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Zscore.R
\name{zscore}
\alias{zscore}
\title{Compute Z-score}
\usage{
zscore(vec)
}
\description{
Computes Z-score when you have the full vector of values (not just a subset).
}
\details{
These functions are necessary for \code{PAINTOR}.
}
\keyword{internal}
