% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arg_list_handler.R
\name{arg_list_handler}
\alias{arg_list_handler}
\title{Argument list handler}
\usage{
arg_list_handler(arg, i, loci)
}
\arguments{
\item{arg}{Name of the argument to evaluate.}

\item{i}{Iterator index.}

\item{loci}{A vector of loci that are being iterated over.}
}
\value{
Selected argument value.
}
\description{
Infer whether the argument should be applied to all loci or matched by index.
}
\concept{general}
\keyword{internal}
