% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloaders.R
\name{axel}
\alias{axel}
\title{axel}
\usage{
axel(
  input.url,
  output.path,
  background = F,
  nThreads = 4,
  force_overwrite = F,
  quiet = F,
  alternate = T
)
}
\description{
R wrapper for axel, which enables multi-threaded download of a single large file.
}
\seealso{
\url{https://github.com/axel-download-accelerator/axel/}

Other download functions: 
\code{\link{wget}()}
}
\concept{download functions}
\keyword{internal}
