% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batapply.R
\name{batapply}
\alias{batapply}
\title{\pkg{echolocatoR}-themed progress bar}
\usage{
batapply(
  X,
  FUN = function(l) Sys.sleep(5/100),
  apply_func = lapply,
  total = length(X),
  name = NULL,
  .envir = parent.frame(),
  cli.progress_show_after = 0,
  clear = FALSE,
  color1 = cli::col_br_cyan,
  color2 = cli::col_br_magenta,
  ...
)
}
\arguments{
\item{X}{a vector (atomic or list) or an \code{\link[base]{expression}}
    object.  Other objects (including classed objects) will be coerced
    by \code{base::\link[base]{as.list}}.}

\item{FUN}{the function to be applied to each element of \code{X}:
    see \sQuote{Details}.  In the case of functions like
    \code{+}, \code{\%*\%}, the function name must be backquoted or quoted.}

\item{apply_func}{Iterator function to use (default: \link[base]{lapply}).}

\item{total}{Passed to \code{\link[cli:cli_progress_bar]{cli_progress_bar()}}.}

\item{name}{Name of the progress bar, a label, passed to
\code{\link[cli:cli_progress_bar]{cli_progress_bar()}}.}

\item{.envir}{Passed to \code{\link[cli:cli_progress_bar]{cli_progress_bar()}}.}

\item{cli.progress_show_after}{How long to wait
before showing the progress bar.}

\item{clear}{Whether to remove the progress bar from the screen after
it has terminated. Defaults to the \code{cli.progress_clear} option, or
\code{TRUE} if unset.}

\item{color1}{First color to use in the palette.}

\item{color2}{Second color to use in the palette.}

\item{...}{Additional arguments passed to \code{apply_func}.}
}
\value{
A (named) list.
}
\description{
Iterator function with \pkg{echolocatoR}-themed progress bar.
}
\examples{
out <- batapply(X = seq_len(30))
}
