% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate.R
\name{biomart_snp_info}
\alias{biomart_snp_info}
\title{Download SNP-wise annotations from Biomart}
\source{
\href{https://bioconductor.org/packages/release/bioc/html/biomaRt.html}{biomaRt}
}
\usage{
biomart_snp_info(snp_list, reference_genome = "grch37", verbose = T)
}
\description{
Download SNP-wise annotations from Biomart
}
\seealso{
Other annotate: 
\code{\link{ANNOTATE.annotate_missense}()},
\code{\link{ANNOTATE.plot_missense}()},
\code{\link{SNPs_by_mutation_type}()},
\code{\link{biomart_geneInfo}()},
\code{\link{biomart_snps_to_geneInfo}()},
\code{\link{epigenetics_enrichment}()},
\code{\link{epigenetics_summary}()},
\code{\link{haploR.HaploReg}()},
\code{\link{haploR.regulomeDB}()}
}
\concept{annotate}
\keyword{internal}
