% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conda.R
\name{CONDA.activate_env}
\alias{CONDA.activate_env}
\title{Activate conda env}
\usage{
CONDA.activate_env(condaenv = "echoR")
}
\description{
Activate conda env
}
\seealso{
Other conda functions: 
\code{\link{CONDA.create_env}()},
\code{\link{CONDA.install}()}
}
\concept{conda functions}
