% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conda.R
\name{CONDA.create_echoR_env}
\alias{CONDA.create_echoR_env}
\title{Create conda env for \emph{echolocatoR}}
\usage{
CONDA.create_echoR_env(
  conda_env = "echoR",
  python_version = NULL,
  channels = c("conda-forge", "bioconda", "r"),
  python_packages = c("pandas>=0.25.0", "pyarrow", "scikit-learn", "bitarray",
    "networkx", "rpy2", "scipy", "pandas-plink"),
  r_packages = c("r-base", "r-devtools"),
  cli_packages = c("tabix", "plink", "macs2"),
  force_install = F,
  auth_token = devtools::github_pat()
)
}
\arguments{
\item{envname}{The conda environment where you want to install \emph{plink}.
By default uses \emph{echoR}, the conda environment distributed with \emph{echolocatoR}.}
}
\description{
Create a new env (or update and existing one)
with the necessary Python, R, and command line packages
to run \emph{echolocatoR}.
}
\details{
\describe{
\item{plink}{https://anaconda.org/bioconda/plink}
\item{tabix}{https://anaconda.org/bioconda/tabix}
}
}
\seealso{
Other conda: 
\code{\link{CONDA.activate_env}()},
\code{\link{CONDA.env_from_yaml}()},
\code{\link{CONDA.find_env_Rlib}()},
\code{\link{CONDA.find_python_path}()},
\code{\link{CONDA.install}()}
}
\concept{conda}
