% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conda.R
\name{CONDA.install}
\alias{CONDA.install}
\title{Install conda if it's missing}
\usage{
CONDA.install(conda_path = "auto", verbose = F)
}
\description{
Install conda if it's missing
}
\seealso{
Other conda: 
\code{\link{CONDA.activate_env}()},
\code{\link{CONDA.create_echoR_env}()},
\code{\link{CONDA.env_from_yaml}()},
\code{\link{CONDA.find_env_Rlib}()},
\code{\link{CONDA.find_python_path}()}
}
\concept{conda}
