% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{createDT_html}
\alias{createDT_html}
\title{Interactive DT (html)}
\usage{
createDT_html(DF, caption = "", scrollY = 400)
}
\description{
Generate an interactive data table with download buttons.
Use this function when manually constructing rmarkdown chunks using cat() in a for loop.
}
\seealso{
Other general: 
\code{\link{.rbind.file.list}()},
\code{\link{createDT}()},
\code{\link{dt.replace}()},
\code{\link{example_fullSS}()},
\code{\link{fillNA_CS_PP}()},
\code{\link{get_os}()},
\code{\link{get_sample_size}()},
\code{\link{startup_image}()},
\code{\link{tryFunc}()}
}
\concept{general}
\keyword{internal}
