% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dey_DeepLearning.R
\name{DEEPLEARNING.melt}
\alias{DEEPLEARNING.melt}
\title{Melt deep learning annotations into long-format}
\usage{
DEEPLEARNING.melt(
  ANNOT,
  model = c("Basenji", "BiClassCNN", "DeepSEA", "ChromHMM", "Roadmap", "Others"),
  aggregate_func = "mean",
  replace_NA = NA,
  replace_negInf = NA,
  save_path = F
)
}
\description{
Melt deep learning annotations into long-format
}
\examples{
root <- "/sc/arion/projects/pd-omics/brian/Fine_Mapping/Data/GWAS/Nalls23andMe_2019/_genome_wide"
## merged_dat <- merge_finemapping_results(dataset = "Data/GWAS/Nalls23andMe_2019", minimum_support = 0, LD_reference = "UKB")
## ANNOT <- DEEPLEARNING.query(merged_dat=merged_dat, level="Allelic_Effect", type="annot")

#### Allelic_Effect ####
path <- file.path(root,"Dey_DeepLearning/Nalls23andMe_2019.Dey_DeepLearning.annot.Allelic_Effect.csv.gz")

#### Variant_Level ####
path <- file.path(root,"Dey_DeepLearning/Nalls23andMe_2019.Dey_DeepLearning.annot.Variant_Level.csv.gz")

ANNOT <- data.table::fread(path, nThread=8)
ANNOT <- find_consensus_SNPs_no_PolyFun(ANNOT)
annot.melt <- DEEPLEARNING.melt(ANNOT=ANNOT, aggregate_func="mean", save_path=gsub("\\\\.csv\\\\.gz",".snp_groups_mean.csv.gz",path))
}
