% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dey_DeepLearning.R
\name{DEEPLEARNING.plot}
\alias{DEEPLEARNING.plot}
\title{Plot deep learning predictions}
\usage{
DEEPLEARNING.plot(
  annot.melt,
  snp_groups = c("GWAS lead", "UCS", "Consensus (-POLYFUN)", "Consensus"),
  comparisons_filter = function(x) {     if ("Consensus" \%in\% x)          return(x) },
  model_metric = c("MAX"),
  facet_formula = ". ~ Model",
  remove_outliers = T,
  show_padj = T,
  show_signif = T,
  vjust_signif = 0,
  show_plot = T,
  save_path = F,
  height = 6,
  width = 8
)
}
\description{
Plot deep learning predictions
}
\examples{
\dontrun{
root = "/sc/arion/projects/pd-omics/brian/Fine_Mapping/Data/GWAS/Nalls23andMe_2019/_genome_wide"

#### Allelic_Effect ####
path <- file.path(root,"Dey_DeepLearning/Nalls23andMe_2019.Dey_DeepLearning.annot.Allelic_Effect.snp_groups_mean.csv.gz")

#### Variant_Level ####
path <- file.path(root,"Dey_DeepLearning/Nalls23andMe_2019.Dey_DeepLearning.annot.Variant_Level.snp_groups_mean.csv.gz")

annot.melt <-  data.table::fread(path, nThread=8)

gp <- DEEPLEARNING.plot(annot.melt=annot.melt, facet_formula="Tissue ~ Model", comparisons_filter=NULL, save_path=gsub("\\\\.csv\\\\.gz",".png",path))
}
}
