% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.rbind.file.list}
\alias{.rbind.file.list}
\title{Bind stored files}
\usage{
.rbind.file.list(file.list, verbose = T, nCores = 4)
}
\description{
Rapidly read a list of files from storage and concatenate them by rows.
}
\seealso{
Other general: 
\code{\link{createDT_html}()},
\code{\link{createDT}()},
\code{\link{dt.replace}()},
\code{\link{get_os}()},
\code{\link{startup_image}()},
\code{\link{tryFunc}()}
}
\concept{general}
\keyword{internal}
