% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dt.replace}
\alias{dt.replace}
\title{Replace items within DT object}
\usage{
dt.replace(DT, target, replacement)
}
\value{
data.frame
}
\description{
Annoyingly, there is no native function to do simple find-and-replace in the `DT` library.
}
\seealso{
Other general: 
\code{\link{.rbind.file.list}()},
\code{\link{createDT_html}()},
\code{\link{createDT}()},
\code{\link{get_os}()},
\code{\link{get_sample_size}()},
\code{\link{startup_image}()},
\code{\link{tryFunc}()}
}
\concept{general}
\keyword{internal}
