% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{example_fullSS}
\alias{example_fullSS}
\title{Store \code{\link{fullSS_dat}}}
\usage{
example_fullSS(
  fullSS_path = "./Nalls23andMe_2019.fullSS_subset.tsv",
  nThread = 1
)
}
\arguments{
\item{fullSS_path}{Path to the full summary statistics file (GWAS or QTL) that you want to fine-map.
It is usually best to provide the absolute path rather than the relative path.}
}
\description{
Store \code{\link{fullSS_dat}}
}
\examples{
\dontrun{
data("fullSS_dat")
fullSS_path <- example_fullSS(fullSS_path="./Nalls23andMe_2019.fullSS_subset.tsv")
}
}
\seealso{
Other general: 
\code{\link{.rbind.file.list}()},
\code{\link{createDT_html}()},
\code{\link{createDT}()},
\code{\link{dt.replace}()},
\code{\link{fillNA_CS_PP}()},
\code{\link{get_os}()},
\code{\link{get_sample_size}()},
\code{\link{startup_image}()},
\code{\link{tryFunc}()}
}
\concept{general}
\keyword{internal}
