% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extra_installs.R
\name{extra_installs}
\alias{extra_installs}
\title{Install tricky R packages}
\usage{
extra_installs(
  cran_packages = F,
  xgr = T,
  bioc_packages = T,
  github_packages = T
)
}
\description{
Some of \pkg{echolocatoR}'s optional R package dependencies are especially tricky to install.
Rather than including them in the DESCRIPTION file,
which would require them to install \pkg{echolocatoR} at all,
this functions installs them afterwards.
Only packages not already installed will be installed.
}
\details{
Some of the main packages installed via this function include:
\describe{
\item{gaston}{CRAN}
\item{plotly}{CRAN}

\item{XGR}{xgr}
\item{foreign}{xgr}
\item{refGenome}{xgr}

\item{Rgraphviz}{Bioconductor}
\item{biomaRt}{Bioconductor}

\item{knitrBootstrap}{GitHub}
\item{susieR}{GitHub}
}
}
\examples{
library(echolocatoR)
extra_installs()
}
