% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_finemap.R
\name{find_consensus_SNPs}
\alias{find_consensus_SNPs}
\title{Adds fine-mapping summary columns}
\usage{
find_consensus_SNPs(
  finemap_dat,
  credset_thresh = 0.95,
  consensus_thresh = 2,
  sort_by_support = T,
  exclude_methods = NULL,
  verbose = T
)
}
\arguments{
\item{consensus_thresh}{Threshold for determining \strong{Consensus_SNP} status.}

\item{verbose}{Whether \pkg{echolocatoR} should be verbose or silent.}
}
\description{
Adds several columns that summarise the results across all fine-mapping tools that were run:
\describe{
\item{Support}{The number of tools in which the SNP was proposed in a credible set.}
\item{mean.PP}{The mean per-SNP PP across all fine-mapping tools used.}
\item{Consensus_SNP}{Whether or not the SNP was in the credible set of ≥ \code{consensus_thresh} SNPs (\emph{default=2}).}
}
}
\seealso{
Other finemapping functions: 
\code{\link{multi_finemap}()}
}
\concept{finemapping functions}
\keyword{internal}
