% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FINEMAP.R
\name{FINEMAP.construct_data}
\alias{FINEMAP.construct_data}
\title{Prepare input files for \code{FINEMAP}}
\source{
\url{http://www.christianbenner.com}
}
\usage{
FINEMAP.construct_data(
  locus_dir,
  subset_DT,
  LD_matrix,
  nThread = 4,
  verbose = T
)
}
\description{
Creates and saves 1) the summary stats file, and 2) the LD matrix.
}
\examples{
data("locus_dir"); data("BST1"); data("BST1_LD_matrix");
finemap_DT <- BST1
dir.create(file.path(locus_dir,"FINEMAP"), showWarnings = FALSE, recursive = TRUE)
out <- subset_common_snps(LD_matrix=LD_matrix, finemap_DT=finemap_DT)
LD_matrix <- out$LD
finemap_DT <- out$DT
dat_paths <- FINEMAP.construct_data(locus_dir=locus_dir, subset_DT=finemap_DT, LD_matrix=LD_matrix)
}
\seealso{
Other FINEMAP: 
\code{\link{FINEMAP.construct_master}()},
\code{\link{FINEMAP.find_executable}()},
\code{\link{FINEMAP.process_results}()},
\code{\link{FINEMAP}()}
}
\concept{FINEMAP}
\keyword{internal}
