% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FINEMAP.R
\name{FINEMAP.construct_master}
\alias{FINEMAP.construct_master}
\title{Construct the \code{FINAMAP} master file}
\source{
\url{http://www.christianbenner.com}
}
\usage{
FINEMAP.construct_master(
  locus_dir,
  n_samples,
  dataset_number = 1,
  file.k = NA,
  verbose = TRUE
)
}
\description{
Creates and saves the master file
which tells \code{FINEMAP} where to find each input file.
}
\examples{
locus_dir <- echodata::locus_dir;
master_path <- FINEMAP.construct_master(locus_dir=locus_dir, n_samples=25000)
}
\seealso{
Other FINEMAP: 
\code{\link{FINEMAP.construct_data}()},
\code{\link{FINEMAP.find_executable}()},
\code{\link{FINEMAP.process_results}()},
\code{\link{FINEMAP}()}
}
\concept{FINEMAP}
\keyword{internal}
