% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FINEMAP.R
\name{FINEMAP.process_results}
\alias{FINEMAP.process_results}
\title{Post-processing of \code{FINEMAP} results}
\source{
\url{http://www.christianbenner.com}
}
\usage{
FINEMAP.process_results(
  locus_dir,
  subset_DT,
  credset_thresh = 0.95,
  pvalue_thresh = 0.05,
  finemap_version = "1.4",
  results_file = ".cred",
  nThread = 1,
  sort_by_CS = TRUE,
  verbose = TRUE
)
}
\description{
Post-processing of \code{FINEMAP} results
}
\examples{
\dontrun{
locus_dir <- echodata::locus_dir; BST1 <- echodata::BST1;
finemap_DT <- BST1
subset_DT <-FINEMAP.process_results(locus_dir=locus_dir, subset_DT=finemap_DT)
}
}
\seealso{
Other FINEMAP: 
\code{\link{FINEMAP.construct_data}()},
\code{\link{FINEMAP.construct_master}()},
\code{\link{FINEMAP.find_executable}()},
\code{\link{FINEMAP}()}
}
\concept{FINEMAP}
\keyword{internal}
