% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FINEMAP.R
\name{FINEMAP}
\alias{FINEMAP}
\title{Fine-map locus with \code{FINEMAP}}
\source{
\url{http://www.christianbenner.com}
}
\usage{
FINEMAP(
  subset_DT,
  locus_dir,
  LD_matrix,
  FINEMAP_path = NULL,
  n_samples = NULL,
  n_causal = 5,
  model = "cond",
  remove_tmps = T,
  credset_thresh = 0.95,
  finemap_version = "1.3.1",
  server = F
)
}
\description{
The stepwise conditional search starts with a causal configuration containing the
SNP with the lowest P-value alone and then iteratively adds to the causal configuration
the SNP given the highest posterior model probability until no further SNP yields
a higher posterior model probability.
}
\examples{
data("locus_dir"); data("BST1"); data("BST1_LD_matrix");
finemap_DT <- BST1
dir.create(file.path(locus_dir,"FINEMAP"), showWarnings = FALSE, recursive = TRUE)
out <- subset_common_snps(LD_matrix=LD_matrix, finemap_DT=finemap_DT)
LD_matrix <- out$LD
finemap_DT <- out$DT
finemap_DT <- FINEMAP(subset_DT=finemap_DT, locus_dir=locus_dir, LD_matrix=LD_matrix)
}
\seealso{
Other FINEMAP: 
\code{\link{FINEMAP.construct_data}()},
\code{\link{FINEMAP.construct_master}()},
\code{\link{FINEMAP.find_executable}()},
\code{\link{FINEMAP.process_results}()}
}
\concept{FINEMAP}
\keyword{internal}
