% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finemap_loci.R
\name{finemap_loci}
\alias{finemap_loci}
\title{Fine-map multiple loci}
\usage{
finemap_loci(
  loci,
  fullSS_path,
  fullSS_genome_build = NULL,
  results_dir = file.path(tempdir(), "results"),
  dataset_name = "dataset_name",
  dataset_type = "GWAS",
  top_SNPs = "auto",
  force_new_subset = FALSE,
  force_new_LD = FALSE,
  force_new_finemap = TRUE,
  finemap_methods = c("ABF", "FINEMAP", "SUSIE", "POLYFUN_SUSIE"),
  finemap_args = NULL,
  n_causal = 5,
  PP_threshold = 0.95,
  consensus_threshold = 2,
  fillNA = 0,
  conditioned_snps = "auto",
  munged = FALSE,
  chrom_col = "CHR",
  chrom_type = NULL,
  position_col = "POS",
  snp_col = "SNP",
  pval_col = "P",
  effect_col = "Effect",
  stderr_col = "StdErr",
  tstat_col = "t-stat",
  locus_col = "Locus",
  freq_col = "Freq",
  MAF_col = "MAF",
  A1_col = "A1",
  A2_col = "A2",
  gene_col = "Gene",
  N_cases_col = "N_cases",
  N_controls_col = "N_controls",
  N_cases = NULL,
  N_controls = NULL,
  proportion_cases = "calculate",
  sample_size = NULL,
  LD_reference = "1KGphase1",
  LD_genome_build = "hg19",
  leadSNP_LD_block = FALSE,
  superpopulation = "EUR",
  download_method = "axel",
  bp_distance = 5e+05,
  min_POS = NA,
  max_POS = NA,
  min_MAF = NA,
  trim_gene_limits = FALSE,
  max_snps = NULL,
  min_r2 = 0,
  remove_variants = FALSE,
  remove_correlates = FALSE,
  file_sep = "\\t",
  query_by = "tabix",
  PAINTOR_QTL_datasets = NULL,
  case_control = TRUE,
  qtl_prefixes = NULL,
  plot_types = c("simple"),
  zoom = "1x",
  nott_epigenome = FALSE,
  nott_show_placseq = FALSE,
  nott_binwidth = 200,
  nott_bigwig_dir = NULL,
  xgr_libnames = NULL,
  roadmap = FALSE,
  roadmap_query = NULL,
  remove_tmps = TRUE,
  conda_env = "echoR",
  nThread = 1,
  verbose = TRUE,
  plot.Nott_epigenome = deprecated(),
  plot.Nott_show_placseq = deprecated(),
  plot.Nott_binwidth = deprecated(),
  plot.Nott_bigwig_dir = deprecated(),
  plot.Roadmap = deprecated(),
  plot.Roadmap_query = deprecated(),
  plot.XGR_libnames = deprecated(),
  server = deprecated(),
  plot.types = deprecated(),
  plot.zoom = deprecated(),
  QTL_prefixes = deprecated(),
  vcf_folder = deprecated(),
  probe_path = deprecated()
)
}
\arguments{
\item{loci}{The list of loci you want to fine-map.
If \code{subset_path="auto"} (\emph{default}),
 a locus subset file name is automatically constructed as:
\emph{Data/<dataset_type>/<dataset_name>/<locus>/
Multi-finemap/<locus>_<dataset_name>_Multi-finemap.tsv.gz}}

\item{fullSS_path}{Path to the full summary statistics file (GWAS or QTL)
that you want to fine-map.
It is usually best to provide the absolute path rather than the relative path.}

\item{results_dir}{Where to store all results.
\strong{IMPORTANT!:} It is usually best to provide the absolute path
rather than the relative path.
This is especially important for \emph{FINEMAP}.}

\item{dataset_name}{The name you want to assign to the dataset being fine-mapped,
This will be used to name the subdirectory where your results will be stored
(e.g. \emph{Data/GWAS/<dataset_name>}).
Don't use special characters (e.g.".", "/").}

\item{dataset_type}{The kind dataset you're fine-mapping
(e.g. GWAS, eQTL, tQTL).
This will also be used when creating the subdirectory where your results
will be stored
(e.g. \emph{Data/<dataset_type>/Kunkle_2019}).}

\item{top_SNPs}{A data.frame with the genomic coordinates of the lead SNP
for each locus.
The lead SNP will be used as the center of the window when extracting
subset from the full GWAS/QTL summary statistics file.
Only one SNP per \strong{Locus} should be included.
At minimum, \code{top_SNPs} should include the following columns:
\describe{
\item{\emph{Locus}}{A unique name for each locus. Often,
 loci are named after a relevant gene (e.g. LRRK2) or based on
  the name/coordinates of the lead SNP (e.g. locus_chr12_40734202) }
\item{\emph{CHR}}{The chromosome that the SNP is on.
 Can be "chr12" or "12" format.}
\item{\emph{POS}}{The genomic position of the SNP (in basepairs)}
}}

\item{force_new_subset}{By default, if a subset of the full
 summary stats file for a given locus is already present,
then \pkg{echolocatoR} will just use the pre-existing file.
Set \code{force_new_subset=T} to override this and extract a new subset.
Subsets are saved in the following path structure:
\emph{Data/\<dataset_type\>/\<dataset_name\>/\<locus\>/Multi-finemap/
\<locus\>_\<dataset_name\>_Multi-finemap.tsv.gz}}

\item{force_new_LD}{If LD file exists, create a new one.}

\item{force_new_finemap}{By default, if an fine-mapping results file for
 a given locus is already present,
then \pkg{echolocatoR} will just use the preexisting file.
Set \code{force_new_finemap=T} to override this and re-run fine-mapping.}

\item{finemap_methods}{Which fine-mapping methods you want to use.}

\item{n_causal}{The maximum number of potential causal SNPs per locus.
This parameter is used somewhat differntly by different fine-mapping tools.
See tool-specific functions for details.}

\item{PP_threshold}{The minimum fine-mapped posterior probability
 for a SNP to be considered part of a Credible Set.
For example, \code{PP_threshold=.95} means that all Credible Set SNPs
will be 95\% Credible Set SNPs.}

\item{consensus_threshold}{The minimum number of fine-mapping tools
that include a SNP
 in their 95\% Credible Sets to consider that it a "Consensus SNP"
 (\emph{default=2}).}

\item{fillNA}{Value to fill LD matrix NAs with.}

\item{conditioned_snps}{Which SNPs to conditions on when fine-mapping
with \emph{COJO}.}

\item{munged}{Whether \code{fullSS_path} have already been
standardised/filtered  full summary stats
with \link[MungeSumstats]{format_sumstats}.
If \code{munged=FALSE} you'll need to provide the necessary
 column name arguments.}

\item{chrom_col}{Name of the chromosome column in the full summary stats
file.
Can be "chr1" or "1" format.
(\emph{default: ="CHR"})}

\item{position_col}{Name of the genomic position column in the full summary
stats file.
Must be in units of basepairs.
(\emph{default: ="POS"})}

\item{snp_col}{Name of the SNP RSID column in the full summary stats file.
(\emph{default: ="SNP"})}

\item{pval_col}{Name of the p-value column in the full summary stats file.
Raw p-values are preferred, but if not available corrected p-values
(e.g. FDR) can be used instead.
(\emph{default: ="P"})}

\item{effect_col}{Name of the effect size column in the full summary stats file.
Effect size is preferred, but if not available other metrics like Beta for
Odds Ratio can be used instead.
(\emph{default: ="Effect"})}

\item{stderr_col}{Name of the standard error  column in the full summary stats file.
You can also set \code{stderr_col="calculate"} to infer standard error
using: \code{effect / tstat}.
(\emph{default: ="StdErr"})}

\item{tstat_col}{Name of the t-statistic column in the full
summary stats file.
This column is not necessary unless \code{stderr_col="calculate"}
or the standard error column is missing.
(\emph{default: ="t-stat"})}

\item{locus_col}{Name of the locus column in the full summary stats file.
(\emph{default: ="Locus"})}

\item{freq_col}{Name of the allele frequency column in the full
 summary stats file.
Effect allele frequency is preferred, but the non-effect allele can
be provided instead (though this may be less accurate).
This column is not necessary unless \code{MAF_col="calculate"} or
 the MAF column is missing.
(\emph{default: ="Freq"})}

\item{MAF_col}{Name of the minor allele frequency column in the
full summary stats file.
Can be inferred from \strong{freq_col} if missing from the dataset.
(\emph{default: ="MAF"})}

\item{A1_col}{Name of the effect/risk allele column in the full
summary stats.
 \strong{\emph{IMPORTANT}}: Make sure this actually the case for your
  full summary stats file.
Unfortunately, different studies report different kinds of allele
information in a non-standardized way.
Meaning that A1/A2 can refer to any number of things:
 \describe{
 \item{effect/other alleles}{in the case of diseases}
 \item{ref/alt alleles}{where ref is the reference genome being used}
 \item{major/minor alleles}{This dichotomy holds true for bi-allelic
  SNPs but not necessary multi-allelic SNPs}
 }
 This makes comparing summary stats across GWAS/QTL datasets very
 confusing for several reasons:
 \describe{
 \item{Multi-allelic SNPs}{SNPs can have more than just 2 possible
 alleles (multi-allelic SNPs). Even if you compare the same SNP
 between two studies, you may accidentally be comparing
 totally different alleles.}
 \item{Valence}{The valence (+/-) of per-SNP GWAS effect sizes/beta
  can be relative to different allele types between studies.
 For example, let's say in one GWAS study your effect size for
  SNP A is 1.5 relative to the major allele in one study,
  and the minor allele happens to be the one found in the reference genome.
  You then try to compare that effect size to that of the same
  SNP in another GWAS.
  But, the valence of the effect sizes in the 2nd GWAS study are all
  relative to the reference genome (instead of the minor allele),
  giving the same SNP a value of -1.2. If you took the effect sizes
   at face value you'd say the signals are in opposite directions.
  But once you take into account how the valences were determined in
   each study you realize that they're actually both positive relative
    to the major allele.}
 }
This process of reversing per-SNP valences based on aligning the alleles
 is known as allele flipping.
This is important when comparing individual SNPs, but can also have an
impact on colocalization results.}

\item{gene_col}{For QTL studies, the name of the \[e\]gene column in the
full summary stats file (\emph{default: "gene"}).
This column will be used for filtering summary stats if supplying a named
list of gene:Locus pairs to \code{loci}.}

\item{N_cases_col}{Name of the column in the full summary stats that has
 the number of case subjects in the study.
This can either be per SNP sample sizes, or one number repeated
across all rows.
Proxy cases (e.g. relatives of people with the disease being investigated)
should be included in this estimate if any were used in the study.
This column is not necesssary if \code{N_cases} parameter is provided.
(\emph{default: ="N_cases"})}

\item{N_controls_col}{Name of the column in the full summary stats that
 has the number of control subjects in the study.
 This can either be per SNP sample sizes, or one number repeated across
  all rows.
 This column is not necesssary if \code{N_controls} parameter is provided.
(\emph{default: ="N_controls"})}

\item{N_cases}{The number of case subjects in the study.
Instead of providing a redundant \strong{N_cases_col} column,
you can simply enter one value here.}

\item{N_controls}{The number of control subjects in the study.
Instead of providing a redundant \strong{N_controls_col} column,
you can simply enter one value here.}

\item{proportion_cases}{The proportion of total subjects in the
study that were cases.
 if \code{proportion_cases="calculate"} then this is inferred:
 \code{N_controls / N_controls}.}

\item{sample_size}{The overall sample size of the study.
If none is given, and \strong{N_cases} and \strong{N_controls}
columns are present,
then sample_size is inferred to be:  \code{max(N_cases) + max(N_controls)}.}

\item{LD_reference}{LD reference to use:
\itemize{
\item{"1KGphase1" : }{1000 Genomes Project Phase 1}
\item{"1KGphase3" : }{1000 Genomes Project Phase 3}
\item{"UKB" : }{Pre-computed LD from a British
European-decent subset of UK Biobank.}
}}

\item{leadSNP_LD_block}{Only return SNPs within the same LD block
as the lead SNP (the SNP with the smallest p-value).}

\item{superpopulation}{Superpopulation to subset LD panel by
(used only if \code{LD_reference} is "1KGphase1" or "1KGphase3".)}

\item{download_method}{\itemize{
\item{\code{"axel"} : }{Multi-threaded}
\item{\code{"wget"} : }{Single-threaded}
\item{\code{"download.file"} : }{Single-threaded}
\item{\code{"internal"} : }{Single-threaded 
(passed to \link[utils]{download.file})}
\item{\code{"wininet"} : }{Single-threaded 
(passed to \link[utils]{download.file})}
\item{\code{"libcurl"} : }{Single-threaded 
(passed to \link[utils]{download.file})}
\item{\code{"curl"} : }{Single-threaded 
(passed to \link[utils]{download.file})}
}
 or
\code{"download.file"} (single-threaded) .}

\item{bp_distance}{Distance around the lead SNP to include.}

\item{min_POS}{Minimum genomic position to include.}

\item{max_POS}{Maximum genomic position to include.}

\item{min_MAF}{Minimum Minor Allele Frequency (MAF) of SNPs to include.}

\item{trim_gene_limits}{If a gene name is supplied to this argument 
(e.g. \code{trim_gene_limits="BST"}), only SNPs within the gene body
 will be included.}

\item{max_snps}{Maximum number of SNPs to include.}

\item{min_r2}{Correlation threshold for \code{remove_correlates}.}

\item{remove_variants}{A list of SNP RSIDs to remove.}

\item{remove_correlates}{A list of SNPs.
If provided, all SNPs that correlates with these SNPs (at r2>=\code{min_r2})
will be removed from both \code{DT} and \code{LD} list items..}

\item{file_sep}{The separator in the full summary stats file.
This parameter is only necessary if \code{query_by!="tabix"}.}

\item{query_by}{Choose which method you want to use to extract
 locus subsets from the full summary stats file.
Methods include:
\describe{
\item{"tabix"}{Convert the full summary stats file in an indexed tabix file.
 Makes querying lightning fast after the initial conversion is done.
  (\emph{default})}
\item{"coordinates"}{Extract locus subsets using min/max genomic c
oordinates with \emph{awk}.}
}}

\item{PAINTOR_QTL_datasets}{A list of QTL datasets to be used when
conducting joint functional fine-mapping with \emph{PAINTOR}.}

\item{plot_types}{Which kinds of plots to include.
Options:
\itemize{
\item{"simple"}{Just plot the following tracks: GWAS,
fine-mapping, gene models}
\item{"fancy"}{Additionally plot XGR annotation tracks
(XGR, Roadmap, Nott2019).}
' \item{"LD"}{LD heatmap showing the 10 SNPs surrounding the lead SNP.}
}}

\item{zoom}{Zoom into the center of the locus when plotting
(without editing the fine-mapping results file).
You can provide either:
\itemize{
\item{The size of your plot window in terms of basepairs
(e.g. \code{zoom=50000} for a 50kb window)}.
\item{How much you want to zoom in (e.g. \code{zoom="1x"}
for the full locus, \code{zoom="2x"}
for 2x zoom into the center of the locus, etc.)}.
}
You can pass a list of window sizes (e.g. \code{c(50000,100000,500000)})
to automatically generate
multiple views of each locus.
This can even be a mix of different style inputs: e.g.
 \code{c("1x","4.5x",25000)}.}

\item{nott_binwidth}{When including Nott et al. (2019)
epigenomic data in the track plots,
adjust the bin width of the histograms.}

\item{nott_bigwig_dir}{Instead of pulling Nott et al. (2019)
 epigenomic data
from the \emph{UCSC Genome Browser}, use a set of local bigwig files.}

\item{roadmap}{Find and plot annotations from Roadmap.}

\item{roadmap_query}{Only plot annotations from Roadmap whose
metadata contains a string or any items from  a list of strings
(e.g. \code{"brain"} or \code{c("brain","liver","monocytes")}).}

\item{remove_tmps}{Whether to remove any temporary files
(e.g. FINEMAP output files) after the pipeline is done running.}

\item{conda_env}{Conda environment to use.}

\item{nThread}{Number of threads to parallelize over.}

\item{verbose}{Print messages.}
}
\value{
A merged data.frame with all fine-mapping results from all loci.
}
\description{
\pkg{echolocatoR} will automatically fine-map each locus.
Uses the \code{top_SNPs} data.frame to define locus coordinates.
}
\examples{
top_SNPs <- echodata::topSNPs_Nalls2019
fullSS_path <- echodata::example_fullSS(dataset = "Nalls2019")

Nalls23andMe_2019.results <- echolocatoR::finemap_loci(
  fullSS_path = fullSS_path,
  top_SNPs = top_SNPs,
  loci = c("BST1","MEX3C"),
  dataset_name = "Nalls23andMe_2019",
  fullSS_genome_build = "hg19",
  munged = TRUE)
}
\seealso{
Other MAIN: 
\code{\link{finemap_locus}()}
}
\concept{MAIN}
