% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gene_trimmer}
\alias{gene_trimmer}
\title{Remove all SNPs outside of of a given gene.}
\usage{
gene_trimmer(subset_DT, gene, min_POS = NULL, max_POS = NULL)
}
\description{
Get the min/max coordinates of a given gene (including known regulatory regions, introns, and exons).
Remove any SNPs from the data.frame that fall outside these coordinates.
}
\seealso{
Other SNP filters: 
\code{\link{filter_snps}()},
\code{\link{limit_SNPs}()},
\code{\link{snps_to_condition}()},
\code{\link{subset_common_snps}()}
}
\concept{SNP filters}
\keyword{internal}
