% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_os}
\alias{get_os}
\title{Identify current operating system (OS).}
\usage{
get_os()
}
\description{
Identify current operating system (OS).
}
\examples{
get_os()
}
\seealso{
Other general functions: 
\code{\link{.rbind.file.list}()},
\code{\link{createDT_html}()},
\code{\link{createDT}()},
\code{\link{dt.replace}()},
\code{\link{printer}()},
\code{\link{startup_image}()},
\code{\link{tryFunc}()}
}
\concept{general functions}
\keyword{internal}
