% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_os.R
\name{get_os}
\alias{get_os}
\title{Identify current operating system (OS).}
\usage{
get_os()
}
\description{
Identify current operating system (OS).
}
\examples{
get_os()
}
\seealso{
Other general: 
\code{\link{createDT_html}()},
\code{\link{createDT}()},
\code{\link{dt.replace}()},
\code{\link{example_fullSS}()},
\code{\link{fillNA_CS_PP}()},
\code{\link{get_sample_size}()},
\code{\link{tryFunc}()}
}
\concept{general}
\keyword{internal}
