% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_sample_size}
\alias{get_sample_size}
\title{Infer sample size from summary stats}
\usage{
get_sample_size(subset_DT, sample_size = NULL, effective_ss = T, verbose = T)
}
\description{
Infer sample size from summary stats
}
\examples{
data("BST1")
BST1 <- finemap_DT
subset_DT <- get_sample_size(subset_DT = finemap_DT)
}
\seealso{
Other general: 
\code{\link{.rbind.file.list}()},
\code{\link{createDT_html}()},
\code{\link{createDT}()},
\code{\link{dt.replace}()},
\code{\link{example_fullSS}()},
\code{\link{fillNA_CS_PP}()},
\code{\link{get_os}()},
\code{\link{startup_image}()},
\code{\link{tryFunc}()}
}
\concept{general}
\keyword{internal}
