% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize.R
\name{get_UKB_MAF}
\alias{get_UKB_MAF}
\title{Get MAF from UK Biobank.}
\usage{
get_UKB_MAF(
  subset_DT,
  output_path = "./Data/Reference/UKB_MAF",
  force_new_maf = FALSE,
  download_method = "axel",
  nThread = 1,
  verbose = TRUE,
  conda_env = "echoR"
)
}
\description{
If \strong{MAF} column is missing,
download MAF from UK Biobank and use that instead.
}
\examples{
\dontrun{
BST1 <- echodata::BST1;
subset_DT <- data.frame(BST1)[,colnames(BST1)!="MAF"]
BST1 <- get_UKB_MAF(subset_DT=subset_DT )
}
}
\seealso{
Other standardizing functions: 
\code{\link{standardize_subset}()}
}
\concept{standardizing functions}
