% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggbio.R
\name{get_window_limits}
\alias{get_window_limits}
\title{Get window size limits for plot}
\usage{
get_window_limits(
  finemap_dat,
  index_as_center = T,
  plot.zoom = NULL,
  verbose = T
)
}
\description{
Get window size limits for plot
}
\examples{
data("BST1");
xlims <- get_window_limits(finemap_dat=BST1, plot.zoom=50000)
xlims <- get_window_limits(finemap_dat=BST1, plot.zoom="all")
xlims <- get_window_limits(finemap_dat=BST1, plot.zoom="5x")
}
\seealso{
Other plot: 
\code{\link{GGBIO.QTL_track}()},
\code{\link{GGBIO.SNP_track}()},
\code{\link{GGBIO.add_lines}()},
\code{\link{GGBIO.dot_summary}()},
\code{\link{GGBIO.invisible_legend}()},
\code{\link{GGBIO.plot}()},
\code{\link{GGBIO.track_heights_dict}()},
\code{\link{GGBIO.transcript_model_track}()},
\code{\link{get_window_suffix}()}
}
\concept{plot}
\keyword{internal}
