% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggbio.R
\name{GGBIO.dot_summary}
\alias{GGBIO.dot_summary}
\title{Multi-fine-map summary dot plot}
\usage{
GGBIO.dot_summary(finemap_dat, PP_threshold = 0.95, show_plot = T)
}
\description{
Multi-fine-map summary dot plot
}
\examples{
data("BST1")
gp <- GGBIO.dot_summary(finemap_dat=BST1)
}
