% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggbio.R
\name{GGBIO.transcript_model_track}
\alias{GGBIO.transcript_model_track}
\title{Track plot for transcript models}
\usage{
GGBIO.transcript_model_track(
  gr.snp_CHR,
  max_transcripts = 1,
  show.legend = T,
  show_plot = F
)
}
\arguments{
\item{gr.snp_CHR}{\emph{echolocator} fine-mapping results
in \code{\link{GRanges}} format.}

\item{max_transcripts}{Max number of transcripts per gene to show.}

\item{show.legend}{Whether to show legend.}
}
\description{
Uses \code{\link{ggbio}}.
}
\examples{
data("BST1");
gr.snp_CHR <- biovizBase::transformDfToGr(BST1, seqnames = "CHR", start = "POS", end = "POS")
track.genes <- GGBIO.transcript_model_track(gr.snp_CHR, max_transcripts=1)
}
\seealso{
Other plot: 
\code{\link{GGBIO.QTL_track}()},
\code{\link{GGBIO.SNP_track}()},
\code{\link{GGBIO.add_lines}()},
\code{\link{GGBIO.dot_summary}()},
\code{\link{GGBIO.invisible_legend}()},
\code{\link{GGBIO.plot}()},
\code{\link{GGBIO.track_heights_dict}()},
\code{\link{get_window_limits}()},
\code{\link{get_window_suffix}()}
}
\concept{plot}
\keyword{internal}
