% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GoShifter.R
\name{GOSHIFTER.create_LD}
\alias{GOSHIFTER.create_LD}
\title{Prepare LD input for \emph{GoShifter}}
\usage{
GOSHIFTER.create_LD(
  locus_dir,
  finemap_dat = NULL,
  LD_path = NULL,
  conda_env = "goshifter",
  nThread = 4,
  verbose = T
)
}
\description{
Prepare LD input for \emph{GoShifter}
}
\examples{
data("BST1"); data("locus_dir")
LD_folder <- GOSHIFTER.create_LD(locus_dir=locus_dir, finemap_dat=subset(BST1, P<5e-8))
}
\seealso{
Other GOSHIFTER: 
\code{\link{GOSHIFTER.create_snpmap}()}
}
\concept{GOSHIFTER}
