% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GoShifter.R
\name{GOSHIFTER.create_snpmap}
\alias{GOSHIFTER.create_snpmap}
\title{Prepare SNP input for \emph{GoShifter}}
\usage{
GOSHIFTER.create_snpmap(dat, GS_results, nThread = 1, verbose = TRUE)
}
\description{
Prepare SNP input for \emph{GoShifter}
}
\examples{
BST1 <- echodata::BST1; locus_dir <- echodata::locus_dir
snpmap <- GOSHIFTER.create_snpmap(dat=BST1, GS_results=file.path(locus_dir,"GoShifter"))
}
\seealso{
Other GOSHIFTER: 
\code{\link{GOSHIFTER.create_LD}()}
}
\concept{GOSHIFTER}
