% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GoShifter.R
\name{GOSHIFTER}
\alias{GOSHIFTER}
\title{Run \code{GoShifter} enrichment pipeline}
\source{
\href{https://github.com/immunogenomics/goshifter}{GitHub}
\href{https://pubmed.ncbi.nlm.nih.gov/26140449/}{Publication}
}
\usage{
GOSHIFTER(
  locus_dir,
  dat,
  SNP.Group = "",
  goshifter_path = NULL,
  permutations = 1000,
  ROADMAP_search = "",
  ROADMAP_type = NA,
  chromatin_states = c("TssA"),
  R2_filter = 0.8,
  overlap_threshold = 1,
  force_new_goshifter = FALSE,
  remove_tmps = TRUE,
  verbose = TRUE,
  save_results = TRUE
)
}
\description{
\code{GoShifter}: Locus-specific enrichment tool.
}
\examples{
locus_dir <- echodata::locus_dir
dat <- echodata::BST1
gs_out <- echolocatoR:::GOSHIFTER(locus_dir = locus_dir, dat = dat)
}
