% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GoShifter.R
\name{GOSHIFTER.run}
\alias{GOSHIFTER.run}
\title{Run \code{GoShifter} enrichment test}
\source{
\href{https://github.com/immunogenomics/goshifter}{GitHub}
\href{https://pubmed.ncbi.nlm.nih.gov/26140449/}{Publication}
}
\usage{
GOSHIFTER.run(
  locus_dir,
  grl.ap,
  permutations = 1000,
  goshifter_path = NULL,
  chromatin_state = "TssA",
  R2_filter = 0.8,
  remove_tmps = T,
  verbose = T,
  overlap_threshold = 1
)
}
\description{
\code{GoShifter}: Locus-specific enrichment tool.
}
