% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{GRanges_overlap}
\alias{GRanges_overlap}
\title{Find overlap between genomic coordinates/ranges}
\usage{
GRanges_overlap(
  dat1,
  dat2,
  chrom_col.1 = "chrom",
  start_col.1 = "start",
  end_col.1 = "end",
  chrom_col.2 = "chrom",
  start_col.2 = "start",
  end_col.2 = "end",
  return_merged = T,
  verbose = T
)
}
\description{
Find overlap between genomic coordinates/ranges
}
\concept{GRanges}
\keyword{internal}
