% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{GRanges_overlap}
\alias{GRanges_overlap}
\title{Find overlap between genomic coordinates/ranges}
\usage{
GRanges_overlap(
  finemap_dat,
  regions,
  return_merged = T,
  chrom_col.1 = "CHR",
  start_col.1 = "POS",
  end_col.1 = "POS",
  chrom_col.2 = "CHR",
  start_col.2 = "start",
  end_col.2 = "end"
)
}
\description{
Find overlap between genomic coordinates/ranges
}
\concept{GRanges}
\keyword{internal}
