% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IMPACT.R
\name{IMPACT.get_annotation_key}
\alias{IMPACT.get_annotation_key}
\title{Get IMPACT annotation key}
\usage{
IMPACT.get_annotation_key(
 
    URL = "https://github.com/immunogenomics/IMPACT/raw/master/IMPACT707/IMPACT_annotation_key.txt",
  save_path = "./echolocatoR/annotations/IMPACT/IMPACT_annotation_key.txt.gz",
  force_new_download = F
)
}
\description{
Inlcludes the study source, tissue,cell type/line,
and cell subtype of each of the 500 annotations in IMPACT
}
\keyword{internal}
