% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IMPACT.R
\name{IMPACT.get_annotations}
\alias{IMPACT.get_annotations}
\title{Download IMPACT annotations}
\usage{
IMPACT.get_annotations(
  baseURL = "https://github.com/immunogenomics/IMPACT/raw/master/IMPACT707/Annotations",
  chrom = NULL,
  subset_DT = NULL,
  nThread = 4,
  all_snps_in_range = F
)
}
\description{
Includes the raw annotation itself, as well as per-SNP IMPACT scores for each annotation.
}
\keyword{internal}
