% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{IMPACT_annotation_key}
\alias{IMPACT_annotation_key}
\title{IMPACT annotation key}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 1511 rows and 7 columns.
}
\source{
\url{https://github.com/immunogenomics/IMPACT}
}
\usage{
IMPACT_annotation_key
}
\description{

}
\concept{IMPACT}
\keyword{datasets}
