% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{IMPACT_annotation_key}
\alias{IMPACT_annotation_key}
\title{IMPACT annotation key}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 1511 rows and 7 columns.
}
\source{
\url{https://github.com/immunogenomics/IMPACT}

\code{
path <- file.path("~/Desktop/Fine_Mapping/echolocatoR",
                  "annotations/IMPACT/IMPACT_annotation_key.txt.gz")
IMPACT_annotation_key <- data.table::fread(path)
usethis::use_data(IMPACT_annotation_key, overwrite = TRUE)
}
}
\usage{
IMPACT_annotation_key
}
\description{
Metadata for each of the IMPACT
(Inference and Modeling of Phenotype-related ACtive Transcription)
annotation files.
Originally from
\href{https://www.cell.com/ajhg/supplemental/S0002-9297%2819%2930108-9}{
Amariuta et al. (2019)}.
}
\seealso{
Other IMPACT: 
\code{\link{IMPACT.get_annotation_key}()},
\code{\link{IMPACT.get_annotations}()},
\code{\link{IMPACT.get_top_annotations}()},
\code{\link{IMPACT.iterate_get_annotations}()},
\code{\link{IMPACT.postprocess_annotations}()},
\code{\link{IMPACT.snp_group_boxplot}()},
\code{\link{IMPACT_heatmap}()},
\code{\link{prepare_mat_meta}()}
}
\concept{IMPACT}
\keyword{datasets}
