% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LD.R
\name{LD.1KG}
\alias{LD.1KG}
\title{Compute LD from 1000 Genomes}
\usage{
LD.1KG(
  locus_dir,
  subset_DT,
  locus = NULL,
  LD_reference = "1KGphase1",
  superpopulation = "EUR",
  vcf_folder = NULL,
  download_reference = T,
  min_r2 = F,
  LD_block = F,
  LD_block_size = 0.7,
  min_Dprime = F,
  remove_correlates = F,
  remove_tmps = T,
  fillNA = 0,
  nThread = 4,
  download_method = "wget",
  verbose = T
)
}
\arguments{
\item{download_reference}{When acquiring LD matrixes,
the default is to delete the full vcf or npz files after \emph{echolocator} has extracted the necssary subset.
However, if you wish to keep these full files (which can be quite large) set \code{download_reference=T}.}

\item{min_r2}{Remove any SNPs are below the LD r2 threshold with the lead SNP within their respective locus.}

\item{LD_block}{Calculate LD blocks with \emph{plink} and only include the block to which the lead SNP belongs.}

\item{LD_block_size}{Adjust the granularity of block sizes when \code{LD_block=T}.}

\item{min_Dprime}{Remove any SNPs are below the LD D' threshold with the lead SNP within their respective locus.
This is paramter currently only works when \code{LD_reference!="UKB"}.}

\item{remove_correlates}{If \code{remove_correlates} is set to a value between 0-1,
removes any SNPs that are in LD with any of the \code{remove_variants} above the threshold provided by \code{remove_correlates}.}

\item{remove_tmps}{Whether to remove any temporary files (e.g. FINEMAP output files) after the pipeline is done running.}

\item{fillNA}{When pairwise LD (r) between two SNPs is \code{NA}, replace with 0.}

\item{verbose}{Whether \emph{echolocatoR} should be verbose or silent.}
}
\description{
Downloads a subset vcf of the 1KG database that matches your locus coordinates.
Then uses \emph{plink} to calculate LD on the fly.
}
\details{
This approach is taken, because other API query tools have limitations with the window size being queried.
This approach does not have this limitations, allowing you to fine-map loci more completely.
}
\examples{
\dontrun{
data("BST1"); data("locus_dir");
LD_matrix <- LD.1KG(locus_dir=file.path("~/Desktop",locus_dir), subset_DT=BST1, locus="BST1", LD_reference="1KGphase1")
}
}
\seealso{
Other LD: 
\code{\link{LD.1KG_download_vcf}()},
\code{\link{LD.LD_blocks}()},
\code{\link{LD.UKBiobank}()},
\code{\link{LD.calculate_LD}()},
\code{\link{LD.dprime_table}()},
\code{\link{LD.filter_vcf_gaston}()},
\code{\link{LD.filter_vcf}()},
\code{\link{LD.get_vcf_folder}()},
\code{\link{LD.leadSNP_block}()},
\code{\link{LD.load_or_create}()},
\code{\link{LD.plink_LD}()},
\code{\link{LD.plink_file}()},
\code{\link{LD.plot}()},
\code{\link{LD.read_bin}()},
\code{\link{LD.read_ld_table}()},
\code{\link{LD.run_plink_LD}()},
\code{\link{LD.translate_population}()},
\code{\link{LD.vcf_to_bed}()}
}
\concept{LD}
\keyword{internal}
