% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LD.R
\name{LD.1KG}
\alias{LD.1KG}
\title{Compute LD from 1000 Genomes}
\usage{
LD.1KG(
  results_path,
  subset_DT,
  locus,
  LD_reference = "1KG_Phase1",
  superpopulation = "EUR",
  vcf_folder = "./Data/Reference/1000_Genomes",
  download_reference = T,
  min_r2 = F,
  LD_block = F,
  LD_block_size = 0.7,
  min_Dprime = F,
  remove_correlates = F,
  remove_tmps = T,
  fillNA = 0
)
}
\arguments{
\item{download_reference}{When acquiring LD matrixes,
the default is to delete the full vcf or npz files after \emph{echolocator} has extracted the necssary subset.
However, if you wish to keep these full files (which can be quite large) set \code{download_reference=T}.}

\item{min_r2}{Remove any SNPs are below the LD r2 threshold with the lead SNP within their respective locus.}

\item{LD_block}{Calculate LD blocks with \emph{plink} and only include the block to which the lead SNP belongs.}

\item{LD_block_size}{Adjust the granularity of block sizes when \code{LD_block=T}.}

\item{min_Dprime}{Remove any SNPs are below the LD D' threshold with the lead SNP within their respective locus.
This is paramter currently only works when \code{LD_reference!="UKB"}.}

\item{remove_correlates}{If \code{remove_correlates} is set to a value between 0-1,
removes any SNPs that are in LD with any of the \code{remove_variants} above the threshold provided by \code{remove_correlates}.}

\item{remove_tmps}{Whether to remove any temporary files (e.g. FINEMAP output files) after the pipeline is done running.}

\item{fillNA}{When pairwise LD (r) between two SNPs is \code{NA}, replace with 0.}
}
\description{
Downloads a subset vcf of the 1KG database that matches your locus coordinates.
Then uses \emph{plink} to calculate LD on the fly.
}
\details{
This approach is taken, because other API query tools have limitations with the window size being queried.
This approach does not have this limitations, allowing you to fine-map loci more completely.
}
