% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LD.R
\name{LD.1KG_download_vcf}
\alias{LD.1KG_download_vcf}
\title{Download vcf subset from 1000 Genomes}
\usage{
LD.1KG_download_vcf(
  subset_DT,
  LD_reference = "1KG_Phase1",
  vcf_folder = "./Data/Reference/1000_Genomes",
  locus_dir,
  locus,
  download_reference = T,
  whole_vcf = F
)
}
\arguments{
\item{download_reference}{When acquiring LD matrixes,
the default is to delete the full vcf or npz files after \emph{echolocator} has extracted the necssary subset.
However, if you wish to keep these full files (which can be quite large) set \code{download_reference=T}.}
}
\description{
Download vcf subset from 1000 Genomes
}
