% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LD.R
\name{LD.calculate_LD}
\alias{LD.calculate_LD}
\title{Calculate LD}
\usage{
LD.calculate_LD(locus_dir, ld_window = 1000, ld_format = "r", verbose = T)
}
\arguments{
\item{locus_dir}{Locus-specific results directory.}

\item{ld_window}{Set --r/--r2 max variant ct pairwise distance (usu. 10).}

\item{ld_format}{Whether to produce an LD matrix with
r (\code{ld_format="r"}) or D' (\code{ld_format="D"}) as the pairwise SNP correlation metric.}
}
\description{
Calculate a pairwise LD matrix from a vcf file using \emph{plink}.
}
\seealso{
Other LD: 
\code{\link{LD.1KG_download_vcf}()},
\code{\link{LD.1KG}()},
\code{\link{LD.LD_blocks}()},
\code{\link{LD.UKBiobank}()},
\code{\link{LD.construct_subset_vcf_name}()},
\code{\link{LD.custom_panel}()},
\code{\link{LD.dprime_table}()},
\code{\link{LD.filter_vcf_gaston}()},
\code{\link{LD.filter_vcf}()},
\code{\link{LD.get_locus_vcf_folder}()},
\code{\link{LD.index_vcf}()},
\code{\link{LD.leadSNP_block}()},
\code{\link{LD.load_or_create}()},
\code{\link{LD.plink_LD}()},
\code{\link{LD.plink_file}()},
\code{\link{LD.plot_LD}()},
\code{\link{LD.query_vcf}()},
\code{\link{LD.rds_to_npz}()},
\code{\link{LD.read_bin}()},
\code{\link{LD.read_ld_table}()},
\code{\link{LD.run_plink_LD}()},
\code{\link{LD.save_LD_matrix}()},
\code{\link{LD.translate_population}()},
\code{\link{LD.vcf_to_bed}()},
\code{\link{saveSparse}()}
}
\concept{LD}
\keyword{internal}
