% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LD.R
\name{LD.custom_panel}
\alias{LD.custom_panel}
\title{Compute LD from user-supplied vcf file}
\usage{
LD.custom_panel(
  LD_reference,
  LD_genome_build = "hg19",
  subset_DT,
  locus_dir,
  min_r2 = F,
  min_Dprime = F,
  remove_correlates = F,
  fillNA = 0,
  LD_block = F,
  LD_block_size = 0.7,
  remove_tmps = T,
  nThread = 4,
  conda_env = "echoR",
  verbose = T
)
}
\description{
Compute LD from user-supplied vcf file
}
\examples{
data("BST1"); data("locus_dir")
LD_reference="~/Desktop/results/Reference/custom_panel_chr4.vcf"
LD_matrix <- LD.custom_panel(LD_reference=LD_reference, subset_DT=BST1, locus_dir=locus_dir)
}
\seealso{
Other LD: 
\code{\link{LD.1KG_download_vcf}()},
\code{\link{LD.1KG}()},
\code{\link{LD.LD_blocks}()},
\code{\link{LD.UKBiobank}()},
\code{\link{LD.calculate_LD}()},
\code{\link{LD.construct_subset_vcf_name}()},
\code{\link{LD.dprime_table}()},
\code{\link{LD.filter_vcf_gaston}()},
\code{\link{LD.filter_vcf}()},
\code{\link{LD.get_locus_vcf_folder}()},
\code{\link{LD.index_vcf}()},
\code{\link{LD.leadSNP_block}()},
\code{\link{LD.load_or_create}()},
\code{\link{LD.plink_LD}()},
\code{\link{LD.plink_file}()},
\code{\link{LD.plot}()},
\code{\link{LD.query_vcf}()},
\code{\link{LD.rds_to_npz}()},
\code{\link{LD.read_bin}()},
\code{\link{LD.read_ld_table}()},
\code{\link{LD.run_plink_LD}()},
\code{\link{LD.save_LD_matrix}()},
\code{\link{LD.translate_population}()},
\code{\link{LD.vcf_to_bed}()}
}
\concept{LD}
\keyword{internal}
