% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LD.R
\name{LD.dprime_table}
\alias{LD.dprime_table}
\title{Calculate LD (D')}
\usage{
LD.dprime_table(SNP_list, LD_folder)
}
\description{
This appriach computes an LD matrix of D' (instead of r or r2) from a vcf.
See \code{\link{LD.run_plink_LD}} for a faster (but less flexible) alternative to computing LD.
}
\seealso{
Other LD: 
\code{\link{LD.1KG_download_vcf}()},
\code{\link{LD.1KG}()},
\code{\link{LD.LD_blocks}()},
\code{\link{LD.UKBiobank}()},
\code{\link{LD.calculate_LD}()},
\code{\link{LD.filter_vcf_gaston}()},
\code{\link{LD.filter_vcf}()},
\code{\link{LD.get_vcf_folder}()},
\code{\link{LD.leadSNP_block}()},
\code{\link{LD.load_or_create}()},
\code{\link{LD.plink_LD}()},
\code{\link{LD.plink_file}()},
\code{\link{LD.plot}()},
\code{\link{LD.read_bin}()},
\code{\link{LD.read_ld_table}()},
\code{\link{LD.run_plink_LD}()},
\code{\link{LD.translate_population}()},
\code{\link{LD.vcf_to_bed}()}
}
\concept{LD}
\keyword{internal}
