% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LD.R
\name{LD.filter_vcf_population}
\alias{LD.filter_vcf_population}
\title{Subset a vcf by superpopulation}
\usage{
LD.filter_vcf_population(
  subset_vcf,
  subset_DT,
  results_path,
  superpopulation,
  popDat
)
}
\arguments{
\item{subset_vcf}{Path to the locus subset vcf.}

\item{popDat}{The metadata file listing the superpopulation to which each sample belongs.}
}
\description{
Subset a vcf by superpopulation
}
