% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LD.R
\name{LD.filter_vcf_rsamtools}
\alias{LD.filter_vcf_rsamtools}
\title{Filter a vcf by min/max coordinates}
\usage{
LD.filter_vcf_rsamtools()
}
\description{
Uses the \code{Rsamtools} package to filter a vcf fiile.
}
\keyword{internal}
