% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LD.R
\name{LD.plink_LD}
\alias{LD.plink_LD}
\title{Calculate LD}
\usage{
LD.plink_LD(
  leadSNP,
  plink_folder,
  min_r2 = F,
  min_Dprime = F,
  remove_correlates = F,
  fillNA = 0
)
}
\description{
Use \emph{plink} to calculate LD from a vcf.
}
