% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LD.R
\name{LD.plink_LD}
\alias{LD.plink_LD}
\title{Calculate LD}
\usage{
LD.plink_LD(
  leadSNP = NULL,
  subset_DT,
  bim_path = NULL,
  remove_excess_snps = T,
  merge_by_RSID = F,
  LD_folder,
  min_r2 = F,
  min_Dprime = F,
  remove_correlates = F,
  fillNA = 0,
  plink_prefix = "plink",
  verbose = T,
  conda_env = NULL
)
}
\description{
Use \emph{plink} to calculate LD from a vcf.
}
\examples{
locus_dir <- "/sc/arion/projects/pd-omics/brian/Fine_Mapping/Data/GWAS/Kunkle_2019/ACE"
LD_folder <- file.path(locus_dir,"LD")
ld.matrix <- LD.plink_LD(subset_DT=BST1, LD_folder=LD_folder)
}
\seealso{
Other LD: 
\code{\link{LD.1KG_download_vcf}()},
\code{\link{LD.1KG}()},
\code{\link{LD.LD_blocks}()},
\code{\link{LD.UKBiobank}()},
\code{\link{LD.calculate_LD}()},
\code{\link{LD.construct_subset_vcf_name}()},
\code{\link{LD.custom_panel}()},
\code{\link{LD.dprime_table}()},
\code{\link{LD.filter_LD}()},
\code{\link{LD.filter_vcf_gaston}()},
\code{\link{LD.filter_vcf}()},
\code{\link{LD.get_locus_vcf_folder}()},
\code{\link{LD.index_vcf}()},
\code{\link{LD.leadSNP_block}()},
\code{\link{LD.load_or_create}()},
\code{\link{LD.plink_file}()},
\code{\link{LD.plot_LD}()},
\code{\link{LD.query_vcf}()},
\code{\link{LD.rds_to_npz}()},
\code{\link{LD.read_bin}()},
\code{\link{LD.read_ld_table}()},
\code{\link{LD.run_plink_LD}()},
\code{\link{LD.save_LD_matrix}()},
\code{\link{LD.snpstats_get_LD}()},
\code{\link{LD.snpstats_get_MAF}()},
\code{\link{LD.translate_population}()},
\code{\link{LD.vcf_to_bed}()},
\code{\link{LDlinkR.LDproxy_batch}()},
\code{\link{popDat_1KGphase1}},
\code{\link{popDat_1KGphase3}},
\code{\link{saveSparse}()}
}
\concept{LD}
\keyword{internal}
