% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LD.R
\name{LD.plot}
\alias{LD.plot}
\title{Plot a subset of the LD matrix}
\usage{
LD.plot(LD_matrix, subset_DT, span = 10)
}
\arguments{
\item{span}{This is very computationally intensive,
so you need to limit the number of SNPs with span.
If \code{span=10}, only 10 SNPs upstream and 10 SNPs downstream of the lead SNP will be plotted.}
}
\description{
Uses \code{gaston} to plot a SNP-annotated LD matrix.
}
\seealso{
Other LD: 
\code{\link{LD.1KG_download_vcf}()},
\code{\link{LD.1KG}()},
\code{\link{LD.LD_blocks}()},
\code{\link{LD.UKBiobank}()},
\code{\link{LD.calculate_LD}()},
\code{\link{LD.dprime_table}()},
\code{\link{LD.filter_vcf_population}()},
\code{\link{LD.filter_vcf_rsamtools}()},
\code{\link{LD.filter_vcf}()},
\code{\link{LD.leadSNP_block}()},
\code{\link{LD.load_or_create}()},
\code{\link{LD.plink_LD}()},
\code{\link{LD.read_bin}()},
\code{\link{LD.read_ld_table}()},
\code{\link{LD.run_plink_LD}()},
\code{\link{LD.translate_population}()},
\code{\link{LD.vcf_to_bed}()},
\code{\link{get_vcf_folder}()}
}
\concept{LD}
\keyword{internal}
