% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LD.R
\name{LD.read_bin}
\alias{LD.read_bin}
\title{Create LD matrix from plink output}
\usage{
LD.read_bin(ld.path)
}
\arguments{
\item{ld.path}{Directory that contains the bin/bim files.}
}
\description{
Depending on which parameters you give \emph{plink} when calculating LD, you get different file outputs.
When it produces bin and bim files, use this function to create a proper LD matrix.
For example, this happens when you try to calculate D' with the \code{--r dprime-signed} flag (instead of just r).
}
\seealso{
Other LD: 
\code{\link{LD.1KG_download_vcf}()},
\code{\link{LD.1KG}()},
\code{\link{LD.LD_blocks}()},
\code{\link{LD.UKBiobank}()},
\code{\link{LD.calculate_LD}()},
\code{\link{LD.dprime_table}()},
\code{\link{LD.filter_vcf_gaston}()},
\code{\link{LD.filter_vcf}()},
\code{\link{LD.get_vcf_folder}()},
\code{\link{LD.leadSNP_block}()},
\code{\link{LD.load_or_create}()},
\code{\link{LD.plink_LD}()},
\code{\link{LD.plink_file}()},
\code{\link{LD.plot}()},
\code{\link{LD.read_ld_table}()},
\code{\link{LD.run_plink_LD}()},
\code{\link{LD.translate_population}()},
\code{\link{LD.vcf_to_bed}()}
}
\concept{LD}
\keyword{internal}
