% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LD.R
\name{LD.read_ld_table}
\alias{LD.read_ld_table}
\title{Create LD matrix from plink output.}
\usage{
LD.read_ld_table(ld.path, snp.subset = F, fillNA = 0, verbose = T)
}
\description{
Depending on which parameters you give \emph{plink} when calculating LD, you get different file outputs.
When it produces an LD table, use this function to create a proper LD matrix.
}
\seealso{
Other LD: 
\code{\link{LD.1KG_download_vcf}()},
\code{\link{LD.1KG}()},
\code{\link{LD.LD_blocks}()},
\code{\link{LD.UKBiobank}()},
\code{\link{LD.calculate_LD}()},
\code{\link{LD.construct_subset_vcf_name}()},
\code{\link{LD.custom_panel}()},
\code{\link{LD.dprime_table}()},
\code{\link{LD.filter_vcf_gaston}()},
\code{\link{LD.filter_vcf}()},
\code{\link{LD.get_locus_vcf_folder}()},
\code{\link{LD.index_vcf}()},
\code{\link{LD.leadSNP_block}()},
\code{\link{LD.load_or_create}()},
\code{\link{LD.plink_LD}()},
\code{\link{LD.plink_file}()},
\code{\link{LD.plot}()},
\code{\link{LD.query_vcf}()},
\code{\link{LD.rds_to_npz}()},
\code{\link{LD.read_bin}()},
\code{\link{LD.run_plink_LD}()},
\code{\link{LD.save_LD_matrix}()},
\code{\link{LD.translate_population}()},
\code{\link{LD.vcf_to_bed}()}
}
\concept{LD}
\keyword{internal}
