% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LD.R
\name{LD.translate_population}
\alias{LD.translate_population}
\title{Translate superopulation acronyms}
\usage{
LD.translate_population(superpopulation)
}
\arguments{
\item{superpopulation}{Three-letter superpopulation name.}
}
\description{
Ensures a common ontology for synonynmous superpopulation names.
}
\seealso{
Other LD: 
\code{\link{LD.1KG_download_vcf}()},
\code{\link{LD.1KG}()},
\code{\link{LD.LD_blocks}()},
\code{\link{LD.UKBiobank}()},
\code{\link{LD.calculate_LD}()},
\code{\link{LD.dprime_table}()},
\code{\link{LD.filter_vcf_gaston}()},
\code{\link{LD.filter_vcf}()},
\code{\link{LD.get_vcf_folder}()},
\code{\link{LD.leadSNP_block}()},
\code{\link{LD.load_or_create}()},
\code{\link{LD.plink_LD}()},
\code{\link{LD.plink_file}()},
\code{\link{LD.plot}()},
\code{\link{LD.rds_to_npz}()},
\code{\link{LD.read_bin}()},
\code{\link{LD.read_ld_table}()},
\code{\link{LD.run_plink_LD}()},
\code{\link{LD.vcf_to_bed}()}
}
\concept{LD}
\keyword{internal}
