% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UKBiobank_LD.R
\name{LD.UKBiobank}
\alias{LD.UKBiobank}
\title{Download LD matrices from UK Biobank}
\usage{
LD.UKBiobank(
  subset_DT = NULL,
  locus_dir,
  sumstats_path = NULL,
  chrom = NULL,
  min_pos = NULL,
  force_new_LD = F,
  chimera = F,
  server = T,
  download_full_ld = F,
  download_method = "direct",
  nThreads = 4,
  return_matrix = F,
  remove_tmps = T
)
}
\description{
Download pre-computed LD matrices from UK Biobank in 3Mb windows,
then subset to the region that overlaps with \code{subset_DT}.
}
\concept{LD}
\keyword{internal}
