% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LD.R
\name{LD.vcf_to_bed}
\alias{LD.vcf_to_bed}
\title{Convert vcf file to BED file}
\usage{
LD.vcf_to_bed(vcf.gz.subset, locus_dir, verbose = T)
}
\arguments{
\item{vcf.gz.subset}{Path to the gzipped locus subset vcf.}

\item{locus_dir}{Locus-specific results directory.}
}
\description{
Uses plink to convert vcf to BED.
}
\seealso{
Other LD: 
\code{\link{LD.1KG_download_vcf}()},
\code{\link{LD.1KG}()},
\code{\link{LD.LD_blocks}()},
\code{\link{LD.UKBiobank}()},
\code{\link{LD.calculate_LD}()},
\code{\link{LD.construct_subset_vcf_name}()},
\code{\link{LD.custom_panel}()},
\code{\link{LD.dprime_table}()},
\code{\link{LD.filter_vcf_gaston}()},
\code{\link{LD.filter_vcf}()},
\code{\link{LD.get_locus_vcf_folder}()},
\code{\link{LD.index_vcf}()},
\code{\link{LD.leadSNP_block}()},
\code{\link{LD.load_or_create}()},
\code{\link{LD.plink_LD}()},
\code{\link{LD.plink_file}()},
\code{\link{LD.plot}()},
\code{\link{LD.query_vcf}()},
\code{\link{LD.rds_to_npz}()},
\code{\link{LD.read_bin}()},
\code{\link{LD.read_ld_table}()},
\code{\link{LD.run_plink_LD}()},
\code{\link{LD.translate_population}()}
}
\concept{LD}
\keyword{internal}
