% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LD.R
\name{LDlinkR.LDproxy_batch}
\alias{LDlinkR.LDproxy_batch}
\title{Extract LD proxies from 1KGphase3}
\source{
\href{https://www.rdocumentation.org/packages/LDlinkR/versions/1.0.2}{website}
}
\usage{
LDlinkR.LDproxy_batch(
  snp,
  pop = "CEU",
  r2d = "r2",
  threshold = 0.8,
  verbose = T
)
}
\description{
Wrapper for \code{LDlinkR::LDproxy_batch}.
Eeasy to use but doesn't scale up well to many SNPs (takes way too long).
}
\examples{
data("merged_DT")
lead.snps <- setNames(subset(merged_DT, leadSNP)$Locus, subset(merged_DT, leadSNP)$SNP)
proxies <- LDlinkR.LDproxy_batch(snp=lead.snps)
}
\seealso{
Other LD: 
\code{\link{LD.1KG_download_vcf}()},
\code{\link{LD.1KG}()},
\code{\link{LD.LD_blocks}()},
\code{\link{LD.UKBiobank}()},
\code{\link{LD.calculate_LD}()},
\code{\link{LD.construct_subset_vcf_name}()},
\code{\link{LD.custom_panel}()},
\code{\link{LD.dprime_table}()},
\code{\link{LD.filter_LD}()},
\code{\link{LD.filter_vcf_gaston}()},
\code{\link{LD.filter_vcf}()},
\code{\link{LD.get_locus_vcf_folder}()},
\code{\link{LD.index_vcf}()},
\code{\link{LD.leadSNP_block}()},
\code{\link{LD.load_or_create}()},
\code{\link{LD.plink_LD}()},
\code{\link{LD.plink_file}()},
\code{\link{LD.plot_LD}()},
\code{\link{LD.query_vcf}()},
\code{\link{LD.rds_to_npz}()},
\code{\link{LD.read_bin}()},
\code{\link{LD.read_ld_table}()},
\code{\link{LD.run_plink_LD}()},
\code{\link{LD.save_LD_matrix}()},
\code{\link{LD.snpstats_get_LD}()},
\code{\link{LD.snpstats_get_MAF}()},
\code{\link{LD.translate_population}()},
\code{\link{LD.vcf_to_bed}()},
\code{\link{popDat_1KGphase1}},
\code{\link{popDat_1KGphase3}},
\code{\link{saveSparse}()}
}
\concept{LD}
