% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{limit_SNPs}
\alias{limit_SNPs}
\title{Limit the number of SNPs per locus.}
\usage{
limit_SNPs(max_snps = 500, subset_DT)
}
\arguments{
\item{max_snps}{The maximum number of SNPs to keep in the resulting data.frame.}

\item{subset_DT}{A data.frame that contains at least the following columns:
\describe{
  \item{SNP}{RSID for each SNP.}
  \item{POS}{Each SNP's genomic position (in basepairs).}
}}
}
\description{
Start with the lead SNP and keep expanding the window until you reach the desired number of snps.
\code{subset_DT} should only contain one locus from one chromosome.
}
\seealso{
Other SNP filters: 
\code{\link{filter_snps}()},
\code{\link{gene_trimmer}()},
\code{\link{snps_to_condition}()},
\code{\link{subset_common_snps}()}
}
\concept{SNP filters}
\keyword{internal}
