% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{LRRK2}
\alias{LRRK2}
\title{\pkg{echolocatoR} output example (LRRK2 locus)}
\format{
data.table
\describe{
  \item{SNP}{SNP RSID}
  \item{CHR}{Chromosome}
  \item{POS}{Genomic position (in basepairs)}
  \item{...}{Optional: extra columns}
}
}
\source{
\url{https://www.biorxiv.org/content/10.1101/388165v3}
}
\usage{
LRRK2
}
\description{
An example results file after running
\code{\link[=finemap_loci]{finemap_loci()}} on the \emph{LRRK2} locus.
}
\details{
Data originally comes from the Parkinson's disease GWAS
by \href{https://www.biorxiv.org/content/10.1101/388165v3}{Nalls et al. (bioRxiv)}.
}
\examples{
\dontrun{
root_dir <- "~/Desktop/Fine_Mapping/Data/GWAS/Nalls23andMe_2019/LRRK2/Multi-finemap"
LRRK2 <- data.table::fread(file.path(root_dir,"Multi-finemap_results.txt"))
LRRK2 <- update_CS_cols(finemap_dat=LRRK2)
LRRK2 <- find_consensus_SNPs(finemap_dat=LRRK2)
usethis::use_data(LRRK2, overwrite = T)
}
}
\keyword{datasets}
