% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_locus_dir}
\alias{make_locus_dir}
\title{Make locus-specific results folder}
\usage{
make_locus_dir(
  results_dir = "./results",
  dataset_type = "dataset_type",
  dataset_name = "dataset_name",
  locus
)
}
\description{
Make locus-specific results folder
}
\examples{
locus_dir <- make_locus_dir(results_dir="./results", dataset_type="GWAS", dataset_name="Nalls23andMe_2019", locus="BST1")
}
\seealso{
Other directory functions: 
\code{\link{get_locus_dir}()},
\code{\link{get_multifinemap_path}()},
\code{\link{get_subset_path}()}
}
\concept{directory functions}
\keyword{internal}
