% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{merge_coloc_results}
\alias{merge_coloc_results}
\title{Convert Jack's coloc results to \emph{echolocatoR} format}
\usage{
merge_coloc_results(
  all_obj,
  results_level = c("summary"),
  nThread = 4,
  verbose = T,
  save_path = F
)
}
\arguments{
\item{all_obj}{Nested list object created by Jack.}

\item{results_level}{Return coloc results at the
"summary" (one row per Locus:eGene pair) or "snp" level (one row per SNP).}

\item{save_path}{Directory where you want the merged results to be saved as a compressed .tsv file.}
}
\description{
Convert Jack's coloc results to \emph{echolocatoR} format
}
\examples{
\dontrun{
load("~/Desktop/Microglia_all_regions_Kunkle_2019_COLOC.RData")
merged_results <- merge_coloc_results(all_obj=all_obj, results_level="snp", save_path="~/Desktop")
}
}
\seealso{
Other general: 
\code{\link{.rbind.file.list}()},
\code{\link{createDT_html}()},
\code{\link{createDT}()},
\code{\link{dt.replace}()},
\code{\link{example_fullSS}()},
\code{\link{fillNA_CS_PP}()},
\code{\link{get_os}()},
\code{\link{get_sample_size}()},
\code{\link{startup_image}()},
\code{\link{tryFunc}()}
}
\concept{general}
\keyword{internal}
