% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{merge_coloc_results_each}
\alias{merge_coloc_results_each}
\title{Iterate merging of coloc results across files}
\usage{
merge_coloc_results_each(
  coloc_rda_files,
  save_path = F,
  results_level = "summary",
  ppH4_thresh = 0.5,
  filter = F,
  force_new = T,
  nThread = 4,
  verbose = T
)
}
\arguments{
\item{coloc_rda_files}{A list of full paths to coloc .RData/.rda results files generated by Jack's coloc scripts.}

\item{save_path}{File path where the merged results will be saved.}

\item{results_level}{Return coloc results at the
"summary" (one row per Locus:eGene pair) or "snp" level (one row per SNP).}

\item{ppH4_thresh}{Threshold to filter summary-level results.}

\item{filter}{Flexible row filtering.}

\item{force_new}{If the \code{save_path} file already exists, overwrite it.}
}
\description{
Iterate merging of coloc results across files
}
\examples{
# List RDS files you want to extract info from
coloc_rds <- list.files("/sc/hydra/projects/ad-omics/microglia_omics/COLOC", pattern = "*_COLOC.RData", recursive = T, full.names = T)
coloc_rds <- coloc_rds[!grepl("*_sQTL_*|Microglia_all_regions_Young",coloc_rds)]
coloc_rds <- coloc_rds[grep("*Microglia_all_regions_*",coloc_rds)]

# Summary-level
coloc_summary <- merge_coloc_results_each(coloc_rds_files=coloc_rds, save_path="/sc/arion/projects/pd-omics/brian/all_COLOC_results.Microglia_all_regions.summary-level.tsv.gz", results_level="summary")
# SNP-level
coloc_snp <- merge_coloc_results_each(coloc_rds_files=coloc_rds, save_path="/sc/arion/projects/pd-omics/brian/all_COLOC_results.Microglia_all_regions.snp-level.tsv.gz", results_level="snp", filter="leadSNP==T")
}
\seealso{
Other coloc: 
\code{\link{merge_coloc_results}()}
}
\concept{coloc}
\keyword{internal}
