% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MEX3C}
\alias{MEX3C}
\title{\emph{echolocatoR} output example (MEX3C locus)}
\format{
data.table
\describe{
  \item{SNP}{SNP RSID}
  \item{CHR}{Chromosome}
  \item{POS}{Genomic positiion (in basepairs)}
  ...
}
}
\source{
\url{https://www.biorxiv.org/content/10.1101/388165v3}
}
\usage{
MEX3C
}
\description{
An example results file after running
\code{\link{finemap_loci}} on the \emph{MEX3C} locus.
}
\details{
Data originally comes from the Parkinson's disease GWAS
by \href{https://www.biorxiv.org/content/10.1101/388165v3}{Nalls et al. (bioRxiv)}.
}
\examples{
\dontrun{
MEX3C <- data.table::fread("~/Desktop/Fine_Mapping/Data/GWAS/Nalls23andMe_2019/MEX3C/Multi-finemap/Multi-finemap_results.txt")
MEX3C <- update_CS_cols(finemap_dat=MEX3C)
MEX3C <- find_consensus_SNPs(finemap_dat=MEX3C)
usethis::use_data(MEX3C, overwrite = T)
}
}
\keyword{datasets}
