% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motifbreakR.R
\name{MOTIFBREAKR.plot}
\alias{MOTIFBREAKR.plot}
\title{Plot \code{\link{motifbreakR}} results}
\source{
\strong{Publication:}
\url{https://pubmed.ncbi.nlm.nih.gov/26272984/}

\strong{GitHub:}
\url{https://github.com/Simon-Coetzee/MotifBreakR}
}
\usage{
MOTIFBREAKR.plot(mb.results, rsid = NULL, effect = c("strong", "weak"))
}
\description{
Plot \code{\link{motifbreakR}} results
}
\examples{
\dontrun{
# mb.results <- readRDS("/sc/arion/projects/pd-omics/brian/motifbreakR/motifbreakR_results.rds")
mb.results <- readRDS("~/Desktop/Fine_Mapping/Data/GWAS/Nalls23andMe_2019/_genome_wide/motifbreakR/motifbreakR_results.rds")
MOTIFBREAKR.plot(mb.results=mb.results, rsid="rs114528427")

# Example from motifbreakR
library(motifbreakR)
data("example.results")
motifbreakR::plotMB(mb.results=example.results, "rs2661839", effect = "strong")
}
}
