% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motifbreakR.R
\name{MOTIFBREAKR}
\alias{MOTIFBREAKR}
\title{Run \code{\link{motifbreakR}}}
\source{
\strong{Publication:}
\url{https://pubmed.ncbi.nlm.nih.gov/26272984/}
\strong{GitHub:}
\url{https://github.com/Simon-Coetzee/MotifBreakR}
\strong{Vignette:}
\url{http://simon-coetzee.github.io/motifBreakR}
}
\usage{
MOTIFBREAKR(
  rsid_list,
  save_rds = TRUE,
  dataset_dir = "./results",
  pwmList = NULL,
  organism = NULL,
  threshold = 0.85,
  show.neutral = FALSE,
  method = "default",
  verbose = TRUE,
  calculate_all_pval = TRUE,
  force_new = FALSE
)
}
\value{
motifbreakr results
}
\description{
\code{\link{motifbreakR}} is a package to predict how much a SNP will disrupt
a transcription factor binding motif (if it falls witihn one).
}
\examples{
\dontrun{
# data("merged_DT")
# rsid_list <- unique(subset(merged_DT, Locus \%in\% c("LRRK2","MED12L","DYRK1A","FCGR2A") & (Consensus_SNP | leadSNP))$SNP)
# rsid_list <- unique(subset(merged_DT, Consensus_SNP | leadSNP)$SNP)
rsid_list <- c("rs11175620","rs7294619","rs74324737")
mb.results <- MOTIFBREAKR(rsid_list=rsid_list, calculate_all_pval=TRUE, force_new = TRUE)
}
}
\seealso{
Other motifbreakR: 
\code{\link{MOTIFBREAKR.filter_by_metadata}()}
}
\concept{motifbreakR}
