% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_finemap.R
\name{multi_finemap}
\alias{multi_finemap}
\title{Fine-map using multiple fine-mapping tools}
\usage{
multi_finemap(
  locus_dir,
  fullSS_path,
  finemap_method_list,
  subset_DT,
  dataset_type,
  LD_matrix = NULL,
  n_causal = 5,
  sample_size = NULL,
  snp_col = "SNP",
  freq_col = "Freq",
  effect_col = "Effect",
  stderr_col = "StdErr",
  pval_col = "P",
  N_cases_col = "N_cases",
  N_controls_col = "N_controls",
  A1_col = "A1",
  A2_col = "A2",
  PAINTOR_QTL_datasets = NULL,
  PP_threshold = 0.95,
  case_control = T,
  verbose = T,
  conda_env = "echoR"
)
}
\arguments{
\item{fullSS_path}{Path to the full summary statistics file (GWAS or QTL) that you want to fine-map.}

\item{dataset_type}{The kind dataset you're fine-mapping (e.g. GWAS, eQTL, tQTL).
This will also be used when creating the subdirectory where your results will be stored
(e.g. \emph{Data/<dataset_type>/Kunkle_2019}).}

\item{n_causal}{The maximum number of potential causal SNPs per locus.
This parameter is used somewhat differntly by different fine-mapping tools.
See tool-specific functions for details.}

\item{sample_size}{The overall sample size of the study.
If none is given, and \strong{N_cases} and \strong{N_controls} columns are present,
then sample_size is inferred to be:  \code{max(N_cases) + max(N_controls)}.}

\item{snp_col}{Name of the SNP RSID column in the full summary stats file.
(\emph{default: ="SNP"})}

\item{freq_col}{Name of the allele frequency column in the full summary stats file.
Effect allele frequency is preferred, but the non-effect allele can be provided instead (though this may be less accurate).
This column is not necessary unless \code{MAF_col="calculate"} or the MAF column is missing.
(\emph{default: ="Freq"})}

\item{effect_col}{Name of the effect size column in the full summary stats file.
Effect size is preferred, but if not available other metrics like Beta for Odds Ratio can be used instead.
(\emph{default: ="Effect"})}

\item{stderr_col}{Name of the standard error  column in the full summary stats file.
You can also set \code{stderr_col="calculate"} to infer standard error using: \code{effect / tstat}.
(\emph{default: ="StdErr"})}

\item{pval_col}{Name of the p-value column in the full summary stats file.
Raw p-values are preferred, but if not available corrected p-values (e.g. FDR) can be used instead.
(\emph{default: ="P"})}

\item{N_cases_col}{Name of the column in the full summary stats that has the number of case subjects in the study.
This can either be per SNP sample sizes, or one number repeated across all rows.
Proxy cases (e.g. relatives of people with the disease being investigated) should be included in this estimate if any were used in the study.
This column is not necesssary if \code{N_cases} parameter is provided.
(\emph{default: ="N_cases"})}

\item{N_controls_col}{Name of the column in the full summary stats that has the number of control subjects in the study.
 This can either be per SNP sample sizes, or one number repeated across all rows.
 This column is not necesssary if \code{N_controls} parameter is provided.
(\emph{default: ="N_controls"})}

\item{A1_col}{Name of the effect/risk allele column in the full summary stats.
 \strong{\emph{IMPORTANT}}: Make sure this actually the case for your full summary stats file.
Unfortunately, different studies report different kinds of allele information in a non-standardized way.
Meaning that A1/A2 can refer to any number of things:
 \describe{
 \item{effect/other alleles}{in the case of diseases}
 \item{ref/alt alleles}{where ref is the reference genome being used}
 \item{major/minor alleles}{This dichotomy holds true for bi-allelic SNPs but not necessary multi-allelic SNPs}
 }
 This makes comparing summary stats across GWAS/QTL datasets very confusing for several reasons:
 \describe{
 \item{Multi-allelic SNPs}{SNPs can have more than just 2 possible alleles (multi-allelic SNPs). Even if you compare the same SNP between two studies, you may accidentally be comparing totally different alleles.}
 \item{Valence}{The valence (+/-) of per-SNP GWAS effect sizes/beta can be relative to different allele types between studies.
 For example, let's say in one GWAS study your effect size for SNP A is 1.5 relative to the major allele in one study,
  and the minor allele happens to be the one found in the reference genome.
  You then try to compare that effect size to that of the same SNP in another GWAS.
  But, the valence of the effect sizes in the 2nd GWAS study are all relative to the reference genome (instead of the minor allele),
  giving the same SNP a value of -1.2. If you took the effect sizes at face value you'd say the signals are in opposite directions.
  But once you take into account how the valences were determined in each study you realize that they're actually both positive relative to the major allele.}
 }
This process of reversing per-SNP valences based on aligning the alleles is known as allele flipping.
This is important when comparing individual SNPs, but can also have an impact on colocalization results.}

\item{PAINTOR_QTL_datasets}{A list of QTL datasets to be used when conducting joint functional fine-mapping with \emph{PAINTOR}.}

\item{PP_threshold}{The minimum fine-mapped posterior probability for a SNP to be considered part of a Credible Set.
For example, \code{PP_threshold=.95} means that all Credible Set SNPs will be 95\% Credible Set SNPs.}

\item{verbose}{Whether \pkg{echolocatoR} should be verbose or silent.}

\item{conda_env}{The name of a conda environment to use.}
}
\description{
Fine-mapping will be repeated on the same locus using each of the tools in \code{finemap_method_list}.
Then, all results will be merged into the locus-specific multi-finemap file,
along with the original per-SNP GWAS/QTL summary statistics.
Each tools will have the following columns:
\describe{
 \item{<tool>.PP}{The posterior probability (PP) of a SNP being causal for the trait. Though this is a generalization and the exact meaning of PP will differ by tools (e.g. Posterior Inclusion Probability for SUSIE).}
 \item{<tool>.CS}{Which credible set the SNP is part of (within a locus). If \code{=0}, then the SNP was not part of any credible set. Some tools only produce one credible set per locus.}
}
}
\examples{
\dontrun{
data("BST1"); data("BST1_LD_matrix");
subset_DT <- BST1
finemap_method_list <- c("ABF","SUSIE")
}
}
\seealso{
Other finemapping functions: 
\code{\link{find_consensus_SNPs}()}
}
\concept{finemapping functions}
\keyword{internal}
