% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NOTT_2019.interactome}
\alias{NOTT_2019.interactome}
\title{Brain cell type-specific enhancers, promoters, and interactomes}
\format{
An object of class \code{list} of length 12.
}
\source{
\url{https://science.sciencemag.org/content/366/6469/1134}
}
\usage{
NOTT_2019.interactome
}
\description{
Originally from \href{https://science.sciencemag.org/content/366/6469/1134}{Nott et al. (2019)}.
Specifically: \emph{aay0793-Nott-Table-S5.xlsx}.
}
\seealso{
Other NOTT_2019: 
\code{\link{NOTT_2019.bigwig_metadata}},
\code{\link{NOTT_2019.superenhancer_interactome}}
}
\concept{NOTT_2019}
\keyword{datasets}
