% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Nott_2019.R
\name{NOTT_2019.plac_seq_plot}
\alias{NOTT_2019.plac_seq_plot}
\title{Plot brain cell-specific interactome data}
\source{
\href{https://science.sciencemag.org/content/366/6469/1134}{Nott et al. (2019)}
\url{https://genome.ucsc.edu/cgi-bin/hgTracks?db=hg19&lastVirtModeType=default&lastVirtModeExtraState=&virtModeType=default&virtMode=0&nonVirtPosition=&position=chr2:127770344-127983251&hgsid=778249165_ySowqECRKNxURRn6bafH0yewAiuf}
}
\usage{
NOTT_2019.plac_seq_plot(
  finemap_dat = NULL,
  locus_dir = NULL,
  title = NULL,
  print_plot = T,
  save_plot = T,
  return_interaction_track = F,
  xlims = NULL,
  zoom_window = NULL,
  index_SNP = NULL,
  return_consensus_overlap = T,
  show_arches = T,
  show_regulatory_rects = T,
  save_annot = F
)
}
\description{
Plot brain cell-specific interactome data
}
\seealso{
Other NOTT_2019: 
\code{\link{NOTT_2019.bigwig_metadata}},
\code{\link{NOTT_2019.epigenomic_histograms}()},
\code{\link{NOTT_2019.get_epigenomic_peaks}()},
\code{\link{NOTT_2019.get_interactions}()},
\code{\link{NOTT_2019.get_interactome}()},
\code{\link{NOTT_2019.get_promoter_celltypes}()},
\code{\link{NOTT_2019.get_promoter_interactome_data}()},
\code{\link{NOTT_2019.get_regulatory_regions}()},
\code{\link{NOTT_2019.interactome}},
\code{\link{NOTT_2019.superenhancer_interactome}},
\code{\link{NOTT_2019.superenhancers}()}
}
\concept{NOTT_2019}
\keyword{internal}
