% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order_loci.R
\name{order_loci}
\alias{order_loci}
\title{Order loci}
\usage{
order_loci(dat, by_UCS_size = FALSE, descending = TRUE, verbose = FALSE)
}
\description{
Order loci by UCS size, or alphabetically.
}
\examples{
merged_DT <- echodata::get_Nalls2019_merged()
merged_DT2 <- echolocatoR:::order_loci(dat=merged_DT)
}
\keyword{internal}
