% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{order_loci}
\alias{order_loci}
\title{Order loci by UCS size, or alphabetically}
\usage{
order_loci(dat, merged_DT, by_UCS_size = F, descending = T, verbose = F)
}
\description{
Order loci by UCS size, or alphabetically
}
\examples{
data("merged_DT");
... by UCS size ...
merged_DT <- order_loci(dat=merged_DT, merged_DT=merged_DT, descending=F)
... alphabetically ...
}
