% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PAINTOR.R
\name{PAINTOR.install}
\alias{PAINTOR.install}
\title{Install PAINTOR via command line}
\usage{
PAINTOR.install(paintor_path = NULL, force_reinstall = FALSE)
}
\description{
Currently there is no R or conda distribution of PAINTOR.
}
\keyword{internal}
