% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PAINTOR.R
\name{PAINTOR.prepare_LD.transethnic}
\alias{PAINTOR.prepare_LD.transethnic}
\title{Prepare transethnic LD files for PAINTOR}
\usage{
PAINTOR.prepare_LD.transethnic(
  subset_DT,
  PT_results_path,
  locus,
  GWAS_populations = NULL,
  QTL_datasets,
  QTL_populations = c("AFA", "CAU", "HIS"),
  LD_reference = "1KG_Phase1",
  force_new_LD = F,
  shared_snps_only = T,
  fillNA = 0
)
}
\description{
Prepare transethnic LD files for PAINTOR
}
\keyword{internal}
