% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/POLYFUN.R
\name{POLYFUN.get_precomputed_priors}
\alias{POLYFUN.get_precomputed_priors}
\title{Extract pre-computed prior}
\source{
https://www.biorxiv.org/content/10.1101/807792v3
}
\usage{
POLYFUN.get_precomputed_priors(
  polyfun = NULL,
  locus_dir,
  finemap_dat = NULL,
  force_new_priors = TRUE,
  remove_tmps = FALSE,
  nThread = 1,
  conda_env = "echoR"
)
}
\description{
Extract SNP-wise prior probabilities pre-computed from many UK Biobank traits.
}
\examples{
\dontrun{

BST1 <- echodata::BST1; locus_dir <- echodata::locus_dir;
priors <- POLYFUN.get_precomputed_priors(locus_dir=locus_dir, finemap_dat=BST1)
}
}
\seealso{
Other polyfun: 
\code{\link{POLYFUN.compute_priors}()},
\code{\link{POLYFUN.download_ref_files}()},
\code{\link{POLYFUN.find_polyfun_folder}()},
\code{\link{POLYFUN.finemapper}()},
\code{\link{POLYFUN.functional_enrichment}()},
\code{\link{POLYFUN.gather_annotations}()},
\code{\link{POLYFUN.gather_ldscores}()},
\code{\link{POLYFUN.h2_enrichment_SNPgroups_plot}()},
\code{\link{POLYFUN.h2_enrichment_SNPgroups}()},
\code{\link{POLYFUN.h2_enrichment}()},
\code{\link{POLYFUN.help}()},
\code{\link{POLYFUN.initialize}()},
\code{\link{POLYFUN.ldsc_annot_enrichment}()},
\code{\link{POLYFUN.munge_summ_stats}()},
\code{\link{POLYFUN.plot}()},
\code{\link{POLYFUN.prepare_snp_input}()},
\code{\link{POLYFUN.read_parquet}()},
\code{\link{POLYFUN.run_ldsc}()},
\code{\link{POLYFUN_SUSIE}()}
}
\concept{polyfun}
\keyword{internal}
