% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/POLYFUN.R
\name{POLYFUN.h2_enrichment_SNPgroups_plot}
\alias{POLYFUN.h2_enrichment_SNPgroups_plot}
\title{Plot heritability (h2) enrichment}
\usage{
POLYFUN.h2_enrichment_SNPgroups_plot(
  RES,
  snp_groups = c("GWAS lead", "UCS", "Consensus (-PolyFun)", "Consensus"),
  comparisons_filter = function(x) {     if ("Consensus" \%in\% x)          return(x) },
  method = "wilcox.test",
  remove_outliers = TRUE,
  title = "S-LDSC heritability enrichment",
  xlabel = "SNP group",
  ylabel = bquote(~"h"^2 ~ "enrichment"),
  show_xtext = TRUE,
  show_padj = TRUE,
  show_signif = TRUE,
  vjust_signif = 0.5,
  show_plot = TRUE,
  save_path = FALSE
)
}
\description{
Plot heritability (h2) enrichment
}
\examples{
\dontrun{
root <- "/sc/arion/projects/pd-omics/brian/Fine_Mapping"
merged_dat <- merge_finemapping_results(dataset = file.path("Data/GWAS/Nalls23andMe_2019"), LD_reference = "UKB", minimum_support = 0)
RES <- POLYFUN.h2_enrichment_SNPgroups(merged_dat=merged_dat, out.path="/sc/arion/projects/pd-omics/brian/Fine_Mapping/Data/GWAS/Nalls23andMe_2019/_genome_wide/PolyFun/output")

plot.h2 <- POLYFUN.h2_enrichment_SNPgroups_plot(RES = RES, show_plot = TRUE)
}
}
\seealso{
Other polyfun: 
\code{\link{POLYFUN.compute_priors}()},
\code{\link{POLYFUN.download_ref_files}()},
\code{\link{POLYFUN.find_polyfun_folder}()},
\code{\link{POLYFUN.finemapper}()},
\code{\link{POLYFUN.functional_enrichment}()},
\code{\link{POLYFUN.gather_annotations}()},
\code{\link{POLYFUN.gather_ldscores}()},
\code{\link{POLYFUN.get_precomputed_priors}()},
\code{\link{POLYFUN.h2_enrichment_SNPgroups}()},
\code{\link{POLYFUN.h2_enrichment}()},
\code{\link{POLYFUN.help}()},
\code{\link{POLYFUN.initialize}()},
\code{\link{POLYFUN.ldsc_annot_enrichment}()},
\code{\link{POLYFUN.munge_summ_stats}()},
\code{\link{POLYFUN.plot}()},
\code{\link{POLYFUN.prepare_snp_input}()},
\code{\link{POLYFUN.read_parquet}()},
\code{\link{POLYFUN.run_ldsc}()},
\code{\link{POLYFUN_SUSIE}()}
}
\concept{polyfun}
