% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IMPACT.R
\name{prepare_mat_meta}
\alias{prepare_mat_meta}
\title{Prepare \emph{IMPACT} data for for \pkg{ComplexHeatmap}}
\usage{
prepare_mat_meta(
  TOP_IMPACT,
  TOP_IMPACT_all,
  snp.group = "Consensus",
  value.var = "mean_IMPACT",
  fill_na = 0
)
}
\description{
Prepare \emph{IMPACT} data for for \pkg{ComplexHeatmap}
}
\seealso{
Other IMPACT: 
\code{\link{IMPACT.get_annotation_key}()},
\code{\link{IMPACT.get_annotations}()},
\code{\link{IMPACT.get_top_annotations}()},
\code{\link{IMPACT.iterate_get_annotations}()},
\code{\link{IMPACT.postprocess_annotations}()},
\code{\link{IMPACT.snp_group_boxplot}()},
\code{\link{IMPACT_heatmap}()}
}
\concept{IMPACT}
\keyword{internal}
