% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{query_by_coordinates_merged}
\alias{query_by_coordinates_merged}
\title{Use \emph{awk} to query locus subsets.}
\usage{
query_by_coordinates_merged(
  top_SNPs,
  fullSS_path,
  subset_path,
  locus,
  chrom_col,
  file_sep = " ",
  location_sep = ":",
  min_POS,
  max_POS,
  bp_distance
)
}
\arguments{
\item{top_SNPs}{A data.frame with the genomic coordinates of the lead SNP for each locus.
The lead SNP will be used as the center of the window when extracting subset from the full GWAS/QTL summary statistics file.
Only one SNP per \strong{Locus} should be included.
At minimum, \code{top_SNPs} should include the following columns:
\describe{
\item{\emph{Locus}}{A unique name for each locus. Often, loci are named after a relevant gene (e.g. LRRK2) or based on the name/coordinates of the lead SNP (e.g. locus_chr12_40734202) }
\item{\emph{CHR}}{The chromosome that the SNP is on. Can be "chr12" or "12" format.}
\item{\emph{POS}}{The genomic position of the SNP (in basepairs)}
}}

\item{fullSS_path}{Path to the full summary statistics file (GWAS or QTL) that you want to fine-map.}

\item{chrom_col}{Name of the chromosome column in the full summary stats file.
Can be "chr1" or "1" format.
(\emph{default: ="CHR"})}

\item{file_sep}{The separator in the full summary stats file.
This parameter is only necessary if \code{query_by!="tabix"}.}

\item{location_sep}{The separator character when \strong{CHR} and \strong{POS} are merged into one column (e.g. ":" when formatted like chr12:12209944).}

\item{min_POS}{Manually set the minimum genomic position for your locus subset.
\code{min_POS} can clip the window size set by \code{bp_distance}.
Only use this parameter when fine-mapping one locus at a time.}

\item{max_POS}{Manually set the maximum genomic position for your locus subset.
\code{max_POS} can clip the window size set by \code{bp_distance}.
Only use this parameter when fine-mapping one locus at a time.}

\item{bp_distance}{The width of the window size you want each locus to be.
For example, if \code{bp_distance=500000} then the locus will span 500kb from the lead SNP in either direction,
resulting in a locus that is ~1Mb long (depending on the dataset).}
}
\description{
Search full summary stats file by genomics coordinates.
To be used when \strong{CHR} and \strong{POS} are merged into one column (e.g. chr12:12209944).
}
\seealso{
Other query functions: 
\code{\link{TABIX.convert_file}()},
\code{\link{TABIX.query}()},
\code{\link{TABIX}()},
\code{\link{extract_SNP_subset}()},
\code{\link{import_topSNPs}()},
\code{\link{query_by_coordinates}()},
\code{\link{query_by_gene}()},
\code{\link{query_by_probe}()},
\code{\link{query_fullSS}()},
\code{\link{query_handler}()}
}
\concept{query functions}
\keyword{internal}
