% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{query_by_probe}
\alias{query_by_probe}
\title{Use \emph{awk} to query locus subsets.}
\usage{
query_by_probe(
  fullSS_path,
  subset_path,
  gene,
  gene_col,
  chrom_col,
  file_sep,
  probe_path,
  coordinates_merged = T,
  location_sep = ":"
)
}
\arguments{
\item{fullSS_path}{Path to the full summary statistics file (GWAS or QTL) that you want to fine-map.}

\item{gene_col}{For QTL studies, the name of the \link{e}gene column in the full summary stats file (\emph{default: "gene"}).
This column will be used for filtering summary stats if supplying a named list of gene:Locus pairs to \code{loci}.}

\item{chrom_col}{Name of the chromosome column in the full summary stats file.
Can be "chr1" or "1" format.
(\emph{default: ="CHR"})}

\item{file_sep}{The separator in the full summary stats file.
This parameter is only necessary if \code{query_by!="tabix"}.}

\item{probe_path}{The location of the file containing translations between probe IDs and gene symbols.
Only used for certain eQTL datasets.}

\item{coordinates_merged}{Whether \strong{CHR} and \strong{POS} are merged into one column (e.g. chr12:12209944).}

\item{location_sep}{The separator character when \strong{CHR} and \strong{POS} are merged into one column (e.g. ":" when formatted like chr12:12209944).}
}
\description{
To be used when probe name (but not gene name) is a column in the full summary stats file.
Requires a gene-probe key file (\strong{probe_path}).
More commonly useful for QTL full summary stats files.
}
\seealso{
Other query functions: 
\code{\link{TABIX.convert_file}()},
\code{\link{TABIX.query}()},
\code{\link{TABIX}()},
\code{\link{extract_SNP_subset}()},
\code{\link{import_topSNPs}()},
\code{\link{query_by_coordinates_merged}()},
\code{\link{query_by_coordinates}()},
\code{\link{query_by_gene}()},
\code{\link{query_fullSS}()},
\code{\link{query_handler}()}
}
\concept{query functions}
\keyword{internal}
