% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{query_fullSS}
\alias{query_fullSS}
\title{Munge the full sumary stats file.}
\usage{
query_fullSS(fullSS_path, subset_path)
}
\arguments{
\item{fullSS_path}{Path to the full summary statistics file (GWAS or QTL) that you want to fine-map.}
}
\description{
Read in a standardize the entire full summary stats file at once.
}
