% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_handler.R
\name{query_handler}
\alias{query_handler}
\title{Query handler}
\usage{
query_handler(
  fullSS_path,
  colmap,
  locus_dir = NULL,
  locus = NULL,
  topSNPs,
  subset_path,
  bp_distance = 5e+05,
  query_by = c("tabix", "fullSS"),
  force_new_subset = FALSE,
  conda_env = "echoR_mini",
  nThread = 1,
  verbose = TRUE
)
}
\arguments{
\item{fullSS_path}{Path to the full summary statistics file (GWAS or QTL)
that you want to fine-map.
It is usually best to provide the absolute path rather
than the relative path.}

\item{colmap}{Column name mappings in in \code{fullSS_path}. Must be a named
list. Can use \link[echodata]{construct_colmap} to assist with this. This
function can be used in two different ways:
\itemize{
\item{\code{munged=FALSE} : }{When \code{munged=FALSE},
 you will need to provide the necessary column names to the
 \code{colmap} argument (\emph{default}).}
 \item{\code{munged=TRUE} : }{ Alternatively, instead of filling out
 each argument in
\link[echodata]{construct_colmap}, you can simply set \code{munged=TRUE}
 if  \code{fullSS_path} has already been munged with
 \link[MungeSumstats]{format_sumstats}.
 }
}}

\item{locus}{Locus name to fine-map (e.g. \code{"BIN1"}).
Can be named to indicate a specific gene within a QTL locus
(e.g. \code{c(ENSG00000136731="BIN1")}).}

\item{topSNPs}{A data.frame with the genomic coordinates of the lead SNP
for each locus.
The lead SNP will be used as the center of the window when extracting
subset from the full GWAS/QTL summary statistics file.
Only one SNP per \strong{Locus} should be included.
At minimum, \code{topSNPs} should include the following columns:
\describe{
\item{\emph{Locus}}{A unique name for each locus. Often,
 loci are named after a relevant gene (e.g. LRRK2) or based on
  the name/coordinates of the lead SNP (e.g. locus_chr12_40734202) }
\item{\emph{CHR}}{The chromosome that the SNP is on.
 Can be "chr12" or "12" format.}
\item{\emph{POS}}{The genomic position of the SNP (in basepairs)}
}}

\item{subset_path}{Path of the resulting locus subset file.}

\item{bp_distance}{Distance around the lead SNP to include.}

\item{query_by}{Choose which method you want to use to extract
 locus subsets from the full summary stats file.
Methods include:
\describe{
\item{"tabix"}{Convert the full summary stats file in an indexed tabix file.
 Makes querying lightning fast after the initial conversion is done.
  (\emph{default})}
\item{"coordinates"}{Extract locus subsets using min/max genomic
coordinates with \emph{awk}.}
}}

\item{force_new_subset}{By default, if a subset of the full
 summary stats file for a given locus is already present,
then \pkg{echolocatoR} will just use the pre-existing file.
Set \code{force_new_subset=T} to override this and extract a new subset.
Subsets are saved in the following path structure:
\emph{Data/\<dataset_type\>/\<dataset_name\>/\<locus\>/Multi-finemap/
\<locus\>_\<dataset_name\>_Multi-finemap.tsv.gz}}

\item{conda_env}{Conda environment to use.}

\item{nThread}{Number of threads to parallelise saving across.}

\item{verbose}{Print messages.}
}
\description{
Handles which query method to use
}
\seealso{
Other query functions: 
\code{\link{extract_snp_subset}()}
}
\concept{query functions}
\keyword{internal}
