% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggbio.R
\name{SNP_track}
\alias{SNP_track}
\title{Track plot for SNPs}
\usage{
SNP_track(
  gr.snp,
  method = "original",
  labels_subset = c("Lead SNP", "Credible Set", "Consensus SNP"),
  r2 = NULL,
  show.legend = T,
  PP_threshold = 0.95
)
}
\arguments{
\item{gr.snp}{\emph{echolocator} fine-mapping results
in \code{\link{GRanges}} format.}

\item{method}{Which method to plot.}

\item{labels_subset}{Which SNP groups to label.}

\item{r2}{Whether color the points by LD r2.}

\item{show.legend}{Whether to show legend.}

\item{PP_threshold}{The minimum fine-mapped posterior probability for a SNP to be considered part of a Credible Set.
For example, \code{PP_threshold=.95} means that all Credible Set SNPs will be 95\% Credible Set SNPs.}
}
\description{
Uses \code{\link{ggbio}}.
}
\seealso{
Other plot: 
\code{\link{GGBIO.plot}()},
\code{\link{invisible_legend}()},
\code{\link{transcript_model_track}()}
}
\concept{plot}
\keyword{internal}
