% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate.R
\name{SNPs_by_mutation_type}
\alias{SNPs_by_mutation_type}
\title{Return only the missense SNPs}
\source{
\href{https://bioconductor.org/packages/release/bioc/html/biomaRt.html}{biomaRt}
}
\usage{
SNPs_by_mutation_type(merged_results, mutation_type = "missense_variant")
}
\description{
Return only the missense SNPs
}
\seealso{
Other annotate: 
\code{\link{ANNOTATE.annotate_missense}()},
\code{\link{ANNOTATE.plot_missense}()},
\code{\link{biomart_geneInfo}()},
\code{\link{biomart_snp_info}()},
\code{\link{biomart_snps_to_geneInfo}()},
\code{\link{epigenetics_enrichment}()},
\code{\link{epigenetics_summary}()},
\code{\link{haploR.HaploReg}()},
\code{\link{haploR.regulomeDB}()},
\code{\link{merge_finemapping_results_each}()}
}
\concept{annotate}
\keyword{internal}
