% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snps_to_condition.R
\name{snps_to_condition}
\alias{snps_to_condition}
\title{Identify SNPs to condition on}
\usage{
snps_to_condition(conditioned_snps, topSNPs, loci)
}
\description{
When running conditional analyses (e.g. \emph{GCTA-COJO}),
this functions automatically identifies SNP to condition on.
}
\concept{SNP filters}
