% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{subset_common_snps}
\alias{subset_common_snps}
\title{Subset LD matrix and dataframe to only their shared SNPs}
\usage{
subset_common_snps(LD_matrix, finemap_dat, fillNA = 0, verbose = F)
}
\value{
data.frame
}
\description{
Find the SNPs that are shared between an LD matrix and another data.frame with a \code{SNP} column.
Then remove any non-shared SNPs from both objects.
}
\examples{
data("BST1"); data('LD_matrix');
finemap_dat=BST1
}
\seealso{
Other SNP filters: 
\code{\link{filter_snps}()},
\code{\link{gene_trimmer}()},
\code{\link{limit_SNPs}()},
\code{\link{snps_to_condition}()}
}
\concept{SNP filters}
\keyword{internal}
