% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise.R
\name{SUMMARISE.peak_overlap_plot}
\alias{SUMMARISE.peak_overlap_plot}
\title{Plot overlap between some SNP group and various epigenomic data}
\source{
\href{https://science.sciencemag.org/content/366/6469/1134}{Nott et al. (2019)}
\href{https://www.biorxiv.org/content/10.1101/2020.01.06.896159v1}{Corces et al. (2020/bioRxiv)}
}
\usage{
SUMMARISE.peak_overlap_plot(
  merged_DT,
  snp_filter = "Consensus_SNP==T",
  include.NOTT_2019_peaks = T,
  include.NOTT_2019_enhancers_promoters = T,
  include.NOTT_2019_PLACseq = T,
  include.CORCES_2020_scATACpeaks = T,
  include.CORCES_2020_Cicero_coaccess = T,
  include.CORCES_2020_bulkATACpeaks = T,
  include.CORCES_2020_HiChIP_FitHiChIP_coaccess = T,
  include.CORCES_2020_gene_annotations = T,
  plot_celltype_specificity = T,
  plot_celltype_specificity_genes = F,
  facets_formula = ". ~ Cell_type",
  show_plot = T,
  label_yaxis = T,
  x_strip_angle = 90,
  x_tick_angle = 40,
  drop_empty_cols = F,
  fill_title = paste(snp_filter, "\\nin epigenomic peaks"),
  save_path = F,
  height = 11,
  width = 12,
  subplot_widths = c(1, 0.5),
  verbose = T
)
}
\arguments{
\item{include.NOTT_2019_peaks}{Plot SNP subset overlap with
peaks from cell-type-specific bulk ATAC, H3K27ac, and H3K4me3 assays.}

\item{include.NOTT_2019_enhancers_promoters}{Plot SNP subset overlap with
cell enhancers and promoters.}

\item{include.CORCES_2020_scATACpeaks}{Plot SNP subset overlap with
cell-type-specific scATAC-seq peaks.}

\item{include.CORCES_2020_Cicero_coaccess}{Plot SNP subset overlap with
Cicero coaccessibility peaks (derived from scATACseq).}
}
\description{
Plot overlap between some SNP group and various epigenomic data
}
\examples{
data("merged_DT");

... Consensus SNPs ...
gg_peaks <- SUMMARISE.peak_overlap_plot(merged_DT=merged_DT, snp_filter="Consensus_SNP==T", fill_title="Consensus SNPs in epigenomic peaks")
... UCS SNPs ...
gg_peaks <- SUMMARISE.peak_overlap_plot(merged_DT=merged_DT, snp_filter="Support>0", fill_title="UCS SNPs in epigenomic peaks")
}
\seealso{
Other summarise: 
\code{\link{SUMMARISE.CS_bin_plot}()},
\code{\link{SUMMARISE.CS_counts_plot}()},
\code{\link{SUMMARISE.get_CS_bins}()},
\code{\link{SUMMARISE.get_CS_counts}()},
\code{\link{SUMMARISE.get_SNPgroup_counts}()},
\code{\link{results_report}()}
}
\concept{summarise}
\keyword{internal}
