% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SURE.R
\name{SURE.melt_snp_groups}
\alias{SURE.melt_snp_groups}
\title{Aggregate \emph{SuRE} p-values for each SNP group}
\source{
\href{https://www.nature.com/articles/s41588-019-0455-2}{Publication}
\href{https://github.com/vansteensellab/SuRE-SNV-code}{GitHub}
\href{https://osf.io/w5bzq/wiki/home/?view}{Full SuRE data}
\href{https://sure.nki.nl}{SNP-SuRE data browse}
}
\usage{
SURE.melt_snp_groups(
  sure_DT,
  grouping_vars = c("Locus"),
  metric_str = "mean",
  replace_NA = NA
)
}
\description{
Aggregate \emph{SuRE} p-values for each SNP group
}
\examples{
sure_DT <- data.table::fread("/sc/arion/projects/pd-omics/brian/Fine_Mapping/Data/GWAS/Nalls23andMe_2019/_genome_wide/SURE/Nalls23andMe_2019.SURE.csv.gz", nThread=4)
}
\seealso{
Other SURE: 
\code{\link{SURE.merge}()},
\code{\link{SURE.plot}()}
}
\concept{SURE}
