% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SURE.R
\name{SURE.merge}
\alias{SURE.merge}
\title{Merge with fine-mapping and \emph{SuRE} results}
\source{
\href{https://www.nature.com/articles/s41588-019-0455-2}{Publication}
\href{https://github.com/vansteensellab/SuRE-SNV-code}{GitHub}
\href{https://osf.io/w5bzq/wiki/home/?view}{Full SuRE data}
\href{https://sure.nki.nl}{SNP-SuRE data browse}
}
\usage{
SURE.merge(merged_DT, sure)
}
\description{
Merge with fine-mapping and \emph{SuRE} results
}
\examples{
sure <- data.table::fread("/sc/arion/projects/pd-omics/data/MPRA/SURE/SuRE_SNP_table_LP190708.txt.gz", nThread=4)
}
\seealso{
Other SURE: 
\code{\link{SURE.melt_snp_groups}()},
\code{\link{SURE.plot}()}
}
\concept{SURE}
