% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabix.R
\name{TABIX.convert_file}
\alias{TABIX.convert_file}
\title{Convert summary stats file to tabix format}
\usage{
TABIX.convert_file(
  fullSS_path,
  study_dir = NULL,
  chrom_col = "CHR",
  position_col = "POS",
  conda_env = "echoR",
  verbose = T
)
}
\arguments{
\item{fullSS_path}{Path to the full summary statistics file (GWAS or QTL) that you want to fine-map.
It is usually best to provide the absolute path rather than the relative path.}

\item{chrom_col}{Name of the chromosome column in the full summary stats file.
Can be "chr1" or "1" format.
(\emph{default: ="CHR"})}

\item{position_col}{Name of the genomic position column in the full summary stats file.
Must be in units of basepairs.
(\emph{default: ="POS"})}

\item{conda_env}{The name of a conda environment to use.}

\item{verbose}{Whether \pkg{echolocatoR} should be verbose or silent.}
}
\description{
Convert summary stats file to tabix format
}
\examples{
\dontrun{
data("genome_wide_dir");
fullSS_path <-  "~/Desktop/nallsEtAl2019_allSamples_allVariants.mod.txt"
fullSS.gz <- TABIX.convert_file(fullSS_path=fullSS_path, results_path=results_path_genome_wide, chrom_col="CHR", position_col="POS")
}
}
\seealso{
Other query functions: 
\code{\link{TABIX.query}()},
\code{\link{TABIX}()},
\code{\link{extract_SNP_subset}()},
\code{\link{import_topSNPs}()},
\code{\link{query_by_coordinates_merged}()},
\code{\link{query_by_coordinates}()},
\code{\link{query_by_gene}()},
\code{\link{query_by_probe}()},
\code{\link{query_fullSS}()},
\code{\link{query_handler}()}
}
\concept{query functions}
