% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabix.R
\name{TABIX.query}
\alias{TABIX.query}
\title{Query a tabix file}
\usage{
TABIX.query(fullSS_tabix, chrom, start_pos, end_pos, verbose = TRUE)
}
\description{
Query by genomic coordinates.
}
\examples{
\dontrun{
fullSS_path <- example_fullSS(fullSS_path=file.path(tempdir(),"Nalls23andMe_2019.fullSS_subset.tsv") )
fullSS_tabix <- TABIX.convert_file(fullSS_path=fullSS_path, chrom_col="CHR", position_col="BP")
TABIX.query(fullSS_tabix=fullSS_tabix, chrom=4, start_pos=13737637, end_pos=13837637)
}
}
\seealso{
Other query functions: 
\code{\link{TABIX.convert_file}()},
\code{\link{TABIX}()},
\code{\link{extract_SNP_subset}()},
\code{\link{import_topSNPs}()},
\code{\link{query_by_coordinates_merged}()},
\code{\link{query_by_coordinates}()},
\code{\link{query_by_gene}()},
\code{\link{query_by_probe}()},
\code{\link{query_fullSS}()},
\code{\link{query_handler}()}
}
\concept{query functions}
