% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{tryFunc}
\alias{tryFunc}
\title{tryCatch extension}
\usage{
tryFunc(input, func)
}
\arguments{
\item{input}{Function input.}

\item{func}{Function.}
}
\description{
Extension of tryCatch function.
}
\examples{
finemap_DT$effective_sample_size <- effective_sample_size(finemap_DT)
}
\concept{developer functions}
