% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tryFunc}
\alias{tryFunc}
\title{tryCatch extension}
\usage{
tryFunc(input, func, ...)
}
\arguments{
\item{input}{Function input.}

\item{func}{Function.}
}
\description{
Extension of tryCatch function.
}
\seealso{
Other general: 
\code{\link{.rbind.file.list}()},
\code{\link{createDT_html}()},
\code{\link{createDT}()},
\code{\link{dt.replace}()},
\code{\link{get_os}()},
\code{\link{startup_image}()}
}
\concept{general}
\keyword{internal}
