% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{VALIDATION.bootstrap}
\alias{VALIDATION.bootstrap}
\title{Perform validation bootstrap procedure}
\source{
\href{https://www.datanovia.com/en/lessons/wilcoxon-test-in-r/}{Wilcox test tutorial}
}
\usage{
VALIDATION.bootstrap(
  metric_df,
  metric,
  predictor = "SNP_group",
  validation_method = NULL,
  synthesize_random = FALSE,
  snp_groups = c("Random", "GWAS lead", "UCS (-PolyFun)", "UCS",
    "Consensus (-PolyFun)", "Consensus"),
  test_method = "coin_wilcox_test",
  locus_means = TRUE,
  iterations = 1000,
  save_path = FALSE,
  nThread = 1,
  verbose = TRUE
)
}
\description{
Perform validation bootstrap procedure
}
\examples{
\dontrun{
save_path <- root <- "/sc/arion/projects/pd-omics/brian/Fine_Mapping/Data/GWAS/Nalls23andMe_2019/_genome_wide"

#### h2 ####
## mean
path <- file.path(root,"PolyFun/Nalls23andMe_2019.h2_enrich.snp_groups.csv.gz")
metric <- "h2.enrichment"
## raw
path <- file.path(root,"PolyFun/h2_merged.csv.gz")
metric <- "SNPVAR"

#### IMPACT ####
## mean
path <- file.path(root,"IMPACT/TOP_IMPACT_all.csv.gz");
metric <- "mean_IMPACT"
## raw
path <- file.path(root,"IMPACT/IMPACT_overlap.csv.gz")
metric <- "IMPACT_score"


## Import and Process ##
metric_df <- data.table::fread(path, nThread=8)
if(metric=="IMPACT_score") metric_df <- subset(metric_df, select=c(SNP,leadSNP,ABF.Credible_Set,ABF.PP,SUSIE.Credible_Set,SUSIE.PP,POLYFUN_SUSIE.Credible_Set,POLYFUN_SUSIE.PP,FINEMAP.Credible_Set,FINEMAP.PP,Consensus_SNP,Support,Locus,IMPACT_score))
if(metric=="mean_IMPACT") metric_df <- echodata::find_consensus_snps_no_polyfun(metric_df)

#### run bootstrap ####
boot_res <- VALIDATION.bootstrap(metric_df=metric_df, metric=metric, nThread=8, save_path=gsub("\\\\.csv\\\\.gz",".bootstrap.coin_wilcox_test.csv.gz",path) )
}
}
\seealso{
Other VALIDATION: 
\code{\link{VALIDATION.bootstrap_multimetric}()},
\code{\link{VALIDATION.bootstrap_plot}()},
\code{\link{VALIDATION.compare_bootstrap_distributions}()},
\code{\link{VALIDATION.permute}()},
\code{\link{VALIDATION.super_plot}()}
}
\concept{VALIDATION}
