% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{VALIDATION.permute_plot}
\alias{VALIDATION.permute_plot}
\title{Plot permutation test results}
\usage{
VALIDATION.permute_plot(
  permute_res,
  validation_method = NULL,
  facet_formula = ". ~ .",
  y_var = "stat",
  show_plot = T,
  save_plot = F,
  height = 5,
  width = 7,
  verbose = T
)
}
\description{
Plot permutation test results
}
\examples{
root <- "/sc/arion/projects/pd-omics/brian/Fine_Mapping/Data/GWAS/Nalls23andMe_2019/_genome_wide"

## h2
path <-  file.path(root,"PolyFun/Nalls23andMe_2019.h2.bootstrap.coin_wilcox_test.csv.gz")
validation_method <- "h2_score"

## IMPACT
path <- file.path(root,"IMPACT//Nalls23andMe_2019.IMPACT_score.bootstrap.coin_wilcox_test.csv.gz")
validation_method <- "IMPACT"

 ## Dey_DeepLearning
path <- file.path(root,"Dey_DeepLearning/Nalls23andMe_2019.Dey_DeepLearning.bootstrap.stats_wilcox.test.csv.gz")
validation_method <- "Dey_DeepLearning"

## SURE MPRA
path <- file.path(root,"SURE/Nalls23andMe_2019.SURE.permutations.csv.gz")
validation_method <- "SuRE MPRA"

# ------  Import
permute_res <-  data.table::fread(path)
## permute_res <- dplyr::rename(permute_res, stat=Stat, p=p.value)

# ---Plot ---
library(patchwork)
permute_sub <- subset(permute_res, endsWith(Metric,"MAX"))
gp1 <- VALIDATION.permute_plot(permute_res=permute_sub, validation_method=validation_method, y_var="stat", save_plot=gsub("\\\\.csv\\\\.gz",".stat_values.png",path), width=9, facet_formula = "Assay ~ Model + Tissue")
gp2 <- VALIDATION.permute_plot(permute_res=subset(permute_sub, stat>0), validation_method=validation_method, y_var="p", save_plot=gsub("\\\\.csv\\\\.gz",".p_values.png",path), width=9,facet_formula = "Assay ~ Model + Tissue")
gp12 <- (gp1 / gp2) + patchwork::plot_annotation(tag_levels = "a")
ggsave(gsub("\\\\.csv\\\\.gz",".png",path),gp12, dpi=400, height=10, width=9)

}
\seealso{
Other VALIDATION: 
\code{\link{VALIDATION.bootstrap_multimetric}()},
\code{\link{VALIDATION.bootstrap}()},
\code{\link{VALIDATION.compare_distributions}()},
\code{\link{VALIDATION.super_plot}()}
}
\concept{VALIDATION}
