% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XGR.R
\name{XGR.enrichment}
\alias{XGR.enrichment}
\title{XGR enrichment}
\usage{
XGR.enrichment(
  gr,
  merged_dat,
  foreground_filter = "Consensus_SNP==T",
  background_filter = "!is.na(SNP)",
  grouping_vars = c("Cell_type")
)
}
\description{
XGR enrichment
}
\examples{
root <- "/sc/arion/projects/pd-omics/brian/Fine_Mapping/Data/GWAS/Nalls23andMe_2019/_genome_wide"

merged_dat <- merge_finemapping_results(dataset = dirname(root), LD_reference = "UKB", minimum_support = 0)
gr <- peaks <- NOTT_2019.get_epigenomic_peaks(convert_to_GRanges = T, nThread = 1) \%>\% subset(select = -c(start,end))
gr <- regions <- NOTT_2019.get_regulatory_regions(as.granges = T)
grouping_vars <- c("Cell_type","Assay")
enrich.lead <- XGR.enrichment(gr=gr, merged_dat=merged_dat, foreground_filter="leadSNP==T",  grouping_vars=grouping_vars)
enrich.UCS <- XGR.enrichment(gr=gr, merged_dat=merged_dat, foreground_filter="Support>0", grouping_vars=grouping_vars)
enrich.consensus <- XGR.enrichment(gr=gr, merged_dat=merged_dat, foreground_filter="Consensus_SNP==T", grouping_vars=grouping_vars)
enrich_res <- rbind(enrich.lead, enrich.UCS, enrich.consensus)

data.table::fwrite(enrich_res, file.path(root,"XGR/Nott2019.enrich_celltype_peaks.csv.gz"))
}
\seealso{
Other XGR: 
\code{\link{DT_to_GRanges}()},
\code{\link{GRanges_to_BED}()},
\code{\link{XGR.download_and_standardize}()},
\code{\link{XGR.enrichment_plot}()},
\code{\link{XGR.filter_assays}()},
\code{\link{XGR.filter_sources}()},
\code{\link{XGR.import_annotations}()},
\code{\link{XGR.iterate_enrichment}()},
\code{\link{XGR.iterate_overlap}()},
\code{\link{XGR.merge_and_process}()},
\code{\link{XGR.plot_enrichment}()},
\code{\link{XGR.plot_peaks}()},
\code{\link{XGR.prepare_foreground_background}()}
}
\concept{XGR}
