% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XGR.R
\name{XGR.enrichment_bootstrap}
\alias{XGR.enrichment_bootstrap}
\title{XGR enrichment (bootstrapped)}
\usage{
XGR.enrichment_bootstrap(
  gr,
  merged_dat,
  snp_groups = c("Random", "GWAS lead", "UCS", "Consensus (-POLYFUN)", "Consensus"),
  background_filter = NULL,
  grouping_vars = c("Cell_type"),
  iterations = 1000,
  fg_sample_size = 20,
  bg_sample_size = 10000,
  save_path = F,
  nThread = 4,
  verbose = T
)
}
\description{
XGR enrichment (bootstrapped)
}
\examples{
root <- "/sc/arion/projects/pd-omics/brian/Fine_Mapping/Data/GWAS/Nalls23andMe_2019/_genome_wide"

merged_dat <- merge_finemapping_results(dataset = dirname(root), LD_reference = "UKB", minimum_support = 0)
merged_dat$Consensus_SNP_noPF <- find_consensus_SNPs(merged_dat, exclude_methods = "POLYFUN_SUSIE", sort_by_support = F)$Consensus_SNP

## NOTT 2019
gr <- peaks <- NOTT_2019.get_epigenomic_peaks(convert_to_GRanges = T, nThread = 1) \%>\% subset(select = -c(start,end))
gr <- regions <- NOTT_2019.get_regulatory_regions(as.granges = T)
## CORCES 2020
corces.scATAC <- data.table::melt.data.table(echolocatoR::CORCES_2020.scATACseq_celltype_peaks, measure.vars = c("ExcitatoryNeurons","InhibitoryNeurons","NigralNeurons","Microglia","Oligodendrocytes","Astrocytes","OPCs"), variable.name = "Cell_type") \%>\% subset(value==1)
corces.scATAC$Assay <- "scATAC"
gr <- LIFTOVER(dat = corces.scATAC, build.conversion = "hg38.to.hg19", chrom_col = "hg38_Chromosome", start_col = "hg38_Start", end_col = "hg38_Stop")
grouping_vars <- c("Cell_type","Assay")


enrich.lead <- XGR.enrichment_bootstrap(gr=gr, merged_dat=merged_dat, foreground_filter="leadSNP==T",  grouping_vars=grouping_vars)
}
\seealso{
Other XGR: 
\code{\link{DT_to_GRanges}()},
\code{\link{GRanges_to_BED}()},
\code{\link{XGR.download_and_standardize}()},
\code{\link{XGR.enrichment_plot}()},
\code{\link{XGR.enrichment}()},
\code{\link{XGR.filter_assays}()},
\code{\link{XGR.filter_sources}()},
\code{\link{XGR.import_annotations}()},
\code{\link{XGR.iterate_enrichment}()},
\code{\link{XGR.iterate_overlap}()},
\code{\link{XGR.merge_and_process}()},
\code{\link{XGR.plot_enrichment}()},
\code{\link{XGR.plot_peaks}()},
\code{\link{XGR.prepare_foreground_background}()}
}
\concept{XGR}
